% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ansi_collapse_quot}
\alias{ansi_collapse_quot}
\title{Utility function that adds quotation marks on elements printed by
ansi_collapse from cli.}
\usage{
ansi_collapse_quot(x, sep = "and")
}
\arguments{
\item{x}{Character vector.}

\item{sep}{One of "or" and "and".}
}
\value{
Scalar character vector collapsed by "and" or "or".
}
\description{
Utility function that adds quotation marks on elements printed by
ansi_collapse from cli.
}
