% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_plate_randomizer.R
\name{assign_subject2plate}
\alias{assign_subject2plate}
\title{assign subject to a plate for longitudinal randomization}
\usage{
assign_subject2plate(plate_map, manifest, subject_id)
}
\arguments{
\item{plate_map}{character vector of locations available for samples
including plate, row and columns}

\item{manifest}{sample manifest mapping sample IDs and subject IDs}

\item{subject_id}{id of subject}
}
\value{
plate_map adding sample IDs to plates keeping samples from the
same subject on the same plate
}
\description{
assign subject to a plate for longitudinal randomization
}
