% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeable.R
\name{bridgeable}
\alias{bridgeable}
\title{Identify if assays shared between Olink Explore 3072 and Olink Explore HT are
can be bridged}
\usage{
bridgeable(data_Explore384, data_HT)
}
\arguments{
\item{data_Explore384}{Olink Explore 3072 dataset to be used for
normalization (required).}

\item{data_HT}{Olink Explore HT dataset to be used for normalization
(required).}
}
\value{
A "tibble" in long format with the following columns:
\itemize{
\item{OlinkID:} Underscore-separated Olink identifiers of matching assays
between Olink Explore HT and Olink Explore 3072.
\item{IsBridgeable:} A boolean flag indicating whether the matching assays
are considered as bridgeable or not.
}
}
\description{
The function uses a dataset from Olink Explore 3072 and a datasets from Olink
Explore HT, and examines if the matched assays between the two products can
be normalized to each other. The input datasets should be exported from Olink
software and should not be altered prior to importing them to this function.
}
\examples{
\donttest{
results <- OlinkAnalyze:::bridgeable(
  data_Explore384 = OlinkAnalyze:::data_3k,
  data_HT = OlinkAnalyze:::data_ht
)
}

}
\author{
Amrita Kar
Marianne Sandin
Danai G. Topouza
Klev Diamanti
}
\keyword{Bridging}
\keyword{NPX}
