% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeable_plot.R
\name{bridgeable_plts}
\alias{bridgeable_plts}
\title{Plots for each bridgeable assays between two platforms / products
Author : Amrita Kar}
\usage{
bridgeable_plts(data = npx_br_data, median_counts_threshold = 150)
}
\arguments{
\item{median_counts_threshold}{Numeric value to use for median counts threshold for the two platforms. Default is 150}

\item{npx_br_data}{a data frame of results from bridge normalization generated by olink_normalization()}
}
\value{
An object of class "ggplot", 4 combined plots for each protein.
}
\description{
Generates a combined plot per assay containing a violin and boxplot plot for IQR ranges ;
correlation plot of NPX values ;
a median count bar plot and
KS plots from 2 platforms
}
\examples{
\donttest{

library(dplyr)

npx_df <- npx_data1 \%>\% filter(!grepl('control',SampleID, ignore.case = TRUE))
ttest_results <- olink_ttest(df=npx_df,
                             variable = 'Treatment',
                             alternative = 'two.sided')
olink_volcano_plot(ttest_results)}

}
