% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all_na_assays.R
\name{check_all_na_assays}
\alias{check_all_na_assays}
\title{Help function to identify assays with Only NAs}
\usage{
check_all_na_assays(df)
}
\arguments{
\item{df}{An arrow object containing columns OlinkID and
either NPX or Quantified_value}
}
\value{
A character vector of assays with all NA values.
}
\description{
Help function to identify assays with Only NAs
}
\author{
Simon Forsberg
 Masoumeh Sheikhi
}
