% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_zip.R
\name{check_checksum}
\alias{check_checksum}
\title{Help function comparing the checksum reported by Olink software to the
checksum of the Olink data file from the input zip-compressed file.}
\usage{
check_checksum(checksum_file, npx_file)
}
\arguments{
\item{checksum_file}{Extracted checksum file from the zip-compressed file
that contains the checksum file from Olink software.}

\item{npx_file}{Extracted file from the zip-compressed file that contains the
Olink data file from Olink software.}
}
\value{
NULL or an error if the files could not be opened or if checksum did
not match.
}
\description{
Runs only if `MD5_checksum.txt` or `checksum_sha256.txt` are present in the
input zip-compressed file. This function does not check if the
\var{checksum_file} is in acceptable format.
}
\author{
Klev Diamanti
}
