% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_col_key}
\alias{check_col_key}
\title{Check if col_key is valid.}
\usage{
check_col_key(col_key)
}
\arguments{
\item{col_key}{Column key for which to retrieve alternative column names. One
of "sample_id", "sample_type", "assay_type", "olink_id", "uniprot", "assay", "panel", "block", "plate_id", "panel_version", "lod", "quant", "ext_npx", "count", "qc_warning", "assay_warn", and "normalization".}
}
\value{
Nothing or an error message if col_key is not valid.
}
\description{
Check if col_key is valid.
}
\keyword{internal}
