% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_columns}
\alias{check_columns}
\title{Check that an input tibble or arrow object contains the required columns.}
\usage{
check_columns(df, col_list)
}
\arguments{
\item{df}{A tibble or arrow object.}

\item{col_list}{A list, where each element is a character vector (usually of
length one). The function will check whether the column names of \var{df}
contain elements from \var{col_list}. When an element of \var{col_list}
contains more than one elements, the function will check whether the column
names of \var{df} include \strong{at least one} of the elements of that
vector.}
}
\value{
Nothing or an error message if any column is missing.
}
\description{
Checks if the input tibble or arrow object \var{df} contains all of the
columns in \var{col_list}.
}
\details{
\var{col_list} might contain also vectors of characters as elements. In this
case one of the characters in the vector is expected to be present. See
examples.
}
\examples{
\dontrun{
tmp_data <- dplyr::tibble(
  "A" = c(1,2,3),
  "B" = c(T,T,F),
  "C" = c("A", "B", "C"),
  "D" = c(F,F,T)
)

# OK
check_columns(df = tmp_data,
              col_list = list("A", "B"))

# ERROR: E is missing
check_columns(df = tmp_data,
              col_list = list("A", "E"))

# ERROR: E and F are missing
check_columns(df = tmp_data,
              col_list = list("A", "E", "F"))

# OK
check_columns(df = tmp_data,
              col_list = list("A", c("B", "C")))

# OK
check_columns(df = tmp_data,
              col_list = list("A", c("B", "E")))

# ERROR: c(F, E) are missing
check_columns(df = tmp_data,
              col_list = list("A", c("F", "E")))

# ERROR: c(F, E) and c(M, N) are missing
check_columns(df = tmp_data,
              col_list = list("A", c("F", "E"), c("M", "N")))
}

}
\author{
Klev Diamanti
Albin Lundin
Lei Conze
Pascal Pucholt
Gil Henriques
}
