% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_NPX_data.R
\name{check_data_completeness}
\alias{check_data_completeness}
\title{Check data completeness}
\usage{
check_data_completeness(df)
}
\arguments{
\item{df}{a NPX dataframe, e.g. from read_NPX()}
}
\value{
None. Used for side effects (warnings)
}
\description{
Throw informative warnings if a dataset appears to have problems
}
\examples{

npx_data1 \%>\%
    mutate(NPX = if_else(SampleID == "A1" & Panel == "Olink Cardiometabolic",
                         NA_real_,
                         NPX)) \%>\%
    check_data_completeness()
}
