% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_delim.R
\name{check_field_separator}
\alias{check_field_separator}
\title{Help function to check and get the separator of a delimited file from Olink
software.}
\usage{
check_field_separator(file, sep = NULL)
}
\arguments{
\item{file}{Path to Olink software output delimited file in wide or long
format. Expecting file extensions `csv` or `txt`.}

\item{sep}{Character separator of delimited input file. One of `NULL` for
auto-detection (default), `,` for comma or `;` for semicolon. Used only for
delimited output files from Olink software.}
}
\value{
The file separator comma `,` or semicolon `;`.
}
\description{
Help function to check and get the separator of a delimited file from Olink
software.
}
\seealso{
[read_npx_delim()]
  [read_npx_delim_long()]
  [read_npx_delim_wide()]
}
\author{
Klev Diamanti
}
