% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_exists.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Help function checking if file exists.}
\usage{
check_file_exists(file)
}
\arguments{
\item{file}{Path to the file.}
}
\description{
Help function checking if file exists.
}
