% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_exists.R
\name{check_file_extension}
\alias{check_file_extension}
\title{Help function checking if file extension is acceptable.}
\usage{
check_file_extension(file)
}
\arguments{
\item{file}{Path to the file.}
}
\value{
The type of the file extension based on the global variable
\code{accepted_npx_file_ext}
}
\description{
Help function checking if file extension is acceptable.
}
\author{
Klev Diamanti
}
