% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_is_data_frame.R
\name{check_is_data_frame}
\alias{check_is_data_frame}
\title{Help function to check if the variable is a data frame.}
\usage{
check_is_data_frame(df, error = FALSE)
}
\arguments{
\item{df}{Data frame to check.}

\item{error}{Boolean to return error or a boolean (default).}
}
\value{
Boolean if the object is a data frame or not, and an error if
`error = TRUE`.
}
\description{
Help function to check if the variable is a data frame.
}
