% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_integer.R
\name{check_is_integer}
\alias{check_is_integer}
\title{Help function checking if a variable is an integer vector.}
\usage{
check_is_integer(int, error = FALSE)
}
\arguments{
\item{int}{Input to check.}

\item{error}{Boolean to return error or a boolean (default).}
}
\value{
Boolean if the variable is an integer vector or not, and an error if
`error = TRUE`.or.
}
\description{
Help function checking if a variable is an integer vector.
}
\seealso{
[check_is_boolean()]
  [check_is_character()]
  [check_is_numeric()]
}
\author{
Klev Diamanti
}
