% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_list.R
\name{check_is_list}
\alias{check_is_list}
\title{Help function checking if a variable is a list.}
\usage{
check_is_list(lst, error = FALSE)
}
\arguments{
\item{lst}{Variable to check.}

\item{error}{Scalar boolean to return an error instead of a \code{FALSE}
(\code{default = FALSE}).}
}
\value{
\code{TRUE} if the variable is a list, and \code{FALSE} if not; error if the
variable is not a list and \code{error = TRUE}.
}
\description{
Help function checking if a variable is a list.
}
\author{
Klev Diamanti
}
