% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_is_numeric}
\alias{check_is_numeric}
\title{Help function checking if a variable is a numeric vector.}
\usage{
check_is_numeric(num, error = FALSE)
}
\arguments{
\item{num}{Number to check.}

\item{error}{Return error or a boolean (default = FALSE).}
}
\value{
Boolean if the variable is a numeric vector or not, and an error if
"error = TRUE".
}
\description{
Help function checking if a variable is a numeric vector.
}
\seealso{
\code{\link{check_is_boolean}}
\code{\link{check_is_character}}
\code{\link{check_is_integer}}
}
\author{
Klev Diamanti
}
