% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boolean.R
\name{check_is_scalar_boolean}
\alias{check_is_scalar_boolean}
\title{Help function checking if a variable is a boolean vector of length 1.}
\usage{
check_is_scalar_boolean(bool, error = FALSE)
}
\arguments{
\item{bool}{Boolean variable to check.}

\item{error}{Return error or a boolean (default).}
}
\value{
Boolean if the variable is a boolean vector of length 1 or not, and
an error if "error = TRUE".
}
\description{
Help function checking if a variable is a boolean vector of length 1.
}
\seealso{
\code{\link{check_is_scalar_character}}
\code{\link{check_is_scalar_integer}}
\code{\link{check_is_scalar_numeric}}
}
\author{
Klev Diamanti
}
