% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_integer.R
\name{check_is_scalar_integer}
\alias{check_is_scalar_integer}
\title{Help function checking if a variable is an integer vector of length 1.}
\usage{
check_is_scalar_integer(int, error = FALSE)
}
\arguments{
\item{int}{Integer vector to check.}

\item{error}{Return error or a boolean (default = FALSE).}
}
\value{
Boolean if the variable is an integer vector of length 1 or not, and
an error if "error = TRUE".
}
\description{
Help function checking if a variable is an integer vector of length 1.
}
\seealso{
\code{\link{check_is_scalar_boolean}}
\code{\link{check_is_scalar_character}}
\code{\link{check_is_scalar_numeric}}
}
\author{
Klev Diamanti
}
