% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_library_installed.R
\name{check_library_installed}
\alias{check_library_installed}
\title{Help function to check if suggested libraries are installed when required.}
\usage{
check_library_installed(x, error = FALSE)
}
\arguments{
\item{x}{A character vector of R libraries.}

\item{error}{Boolean to return error or a boolean (default).}
}
\value{
Boolean if the library is installed or not, and an error if
\code{error = TRUE}.
}
\description{
Help function to check if suggested libraries are installed when required.
}
\author{
Klev Diamanti
}
