% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_lod.R
\name{check_lod_darid}
\alias{check_lod_darid}
\title{Check fixed LOD file version vs DarID
Check for old Fixed LOD when using D.07, 08, 10, and 14}
\usage{
check_lod_darid(lod_file, data)
}
\arguments{
\item{lod_file}{}

\item{data}{}
}
\description{
Check fixed LOD file version vs DarID
Check for old Fixed LOD when using D.07, 08, 10, and 14
}
\keyword{internal}
