% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_log_colname}
\alias{check_log_colname}
\title{Check if check_log has identified required column in the dataset.}
\usage{
check_log_colname(check_log, col_key)
}
\arguments{
\item{check_log}{A named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{\link[=check_npx]{check_npx()}} will be run internally using \code{df}.}

\item{col_key}{Column key for which to retrieve alternative column names. One
of "sample_id", "sample_type", "assay_type", "olink_id", "uniprot", "assay", "panel", "block", "plate_id", "panel_version", "lod", "quant", "ext_npx", "count", "qc_warning", "assay_warn", "normalization", and "qc_version".}
}
\value{
Nothing or an error message if the required column is missing.
}
\description{
Check if check_log has identified required column in the dataset.
}
\keyword{internal}
