% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx}
\alias{check_npx}
\title{Check NPX data format}
\usage{
check_npx(df = df, preferred_names = NULL)
}
\arguments{
\item{df}{A tibble or and arrow object from \code{\link{read_npx}}.}

\item{preferred_names}{A named character vector where names are internal
column names and values are column names to be selected from the input data
frame. Read the \emph{description} for further information.}
}
\value{
A list containing the following elements:
\itemize{
\item \strong{col_names} List of column names from the input data frame
marking the columns to be used in downstream analyses.
\item \strong{oid_invalid} Character vector of invalid \var{OlinkID}.
\item \strong{assay_na} Character vector of assays with all samples having
\emph{NA} values.
\item \strong{sample_id_dups} Character vector of duplicate \var{SampleID}.
}
}
\description{
This function performs various checks on NPX data, including checking
column names, validating Olink identifiers, identifying assays with \emph{NA}
values for all samples and detecting duplicate sample identifiers.
}
\details{
OlinkAnalyze uses pre-defined names of columns of data frames to perform
downstream analyses. At the same time, different Olink platforms export data
with different column names (e.g. different protein quantification metric).
This function aims to instruct each function of OlinkAnalyze on the column it
should be using for the downstream analysis. This should be seamless for data
exported from Olink Software and imported to R using the read_npx function.

However, in certain cases the columns of interest might be named differently.
This function allows assigning custom-named columns of a data frame to
internally expected variables that will in turn instruct Olink Analyze
functions to use them for downstream analysis. For example, if one wished to
use the column \var{PCNormalizedNPX} for their analysis instead of the
column \var{NPX}, then they can assign this new name to the internal
variable \var{quant} to inform the package that in the downstream analysis
\var{PCNormalizedNPX} should be used. See example 3.

Similarly, in case of multiple matches (e.g. the data frame contains both
columns \var{LOD} and \var{PlateLOD}) the ties will need to be resolved by
the user using the argument \var{preferred_names} from this function.  See
example 4.

The argument \var{preferred_names} is a named character vector with internal
column names as names and column names of the current data set as values.
Names of the input vector can be one or more of the following:
sample_id, sample_type, assay_type, olink_id, plate_id, qc_warning, lod, and quant
}
\examples{
\dontrun{
# Example 0: Use npx_data1 to check that check_npx works
check_npx_result <- OlinkAnalyze::npx_data1 |>
  OlinkAnalyze::check_npx() |>
  suppressWarnings()
}

}
\author{
Masoumeh Sheikhi
}
