% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx_col_class}
\alias{check_npx_col_class}
\title{Help function checking types of columns in data.}
\usage{
check_npx_col_class(df, col_names)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{col_names}{A list of matched column names. This is the output of the
\var{check_npx_col_names} function.}
}
\value{
A data frame with the columns \var{col_name}, \var{col_key},
\var{col_class} and \var{expected_col_class} marking columns with the
incorrect type.
}
\description{
This function checks if certain columns from \var{df} have the correct type
to enable downstream analysis. Columns to be checked are marked as such in
the columns \var{col_class} and \var{col_class_check} of
\var{column_name_dict}.
}
