% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx_col_names.R
\name{check_npx_col_names}
\alias{check_npx_col_names}
\title{Check and update column names}
\usage{
check_npx_col_names(df, preferred_names = NULL)
}
\arguments{
\item{df}{An arrow object to check and update column names.}

\item{preferred_names}{A named list where names are desired column names
and values are current column names.}
}
\value{
A list of matched column names based on the preferred names.
}
\description{
This function checks the column names of Olink data and updates them if specified.
}
\details{
This function checks if the preferred column names are present in the data. If the
preferred names are provided, it updates the column names according to the provided preferences.
Otherwise, it uses the existing column names in the dataset.
}
\author{
Klev Diamanti
 Masoumeh Sheikhi
}
