% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx_nonunique_uniprot}
\alias{check_npx_nonunique_uniprot}
\title{Help function checking for assays mapping to multiple UniProt identifiers.}
\usage{
check_npx_nonunique_uniprot(df, col_names)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}. The output of \code{\link[=read_npx]{read_npx()}} is an \code{olink_class}
object that carries the check log as an attribute. If the check log
indicates that the data still needs cleaning (e.g. invalid identifiers,
duplicate samples), the downstream function will automatically clean the
data before proceeding.}

\item{col_names}{A list of matched column names. This is the output of the
\var{check_npx_col_names} function.}
}
\value{
A character vector of assay identifiers \var{OlinkID} that map to
more than one \var{UniProt} identifiers.
}
\description{
Occasionally, updates in panel versions include updates in \var{UniProt}
identifiers (e.g. change in formatting). This function identifies cases where
an assay identifier \var{OlinkID} maps to multiple \var{UniProt} identifiers.
}
\author{
Kathleen Nevola
Kang Dong
Klev Diamanti
}
\keyword{internal}
