% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx_olinkid}
\alias{check_npx_olinkid}
\title{Help function checking whether df contains invalid Olink identifiers}
\usage{
check_npx_olinkid(df, col_names)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}. The output of \code{\link[=read_npx]{read_npx()}} is an \code{olink_npx}
object that carries the check log as an attribute. If the check log
indicates that the data still needs cleaning (e.g. invalid identifiers,
duplicate samples), the downstream function will automatically clean the
data before proceeding.}

\item{col_names}{A list of matched column names. This is the output of the
\var{check_npx_col_names} function.}
}
\value{
A character vector with invalid \var{OlinkID}.
}
\description{
This function checks if Olink identifiers (\var{OlinkID}) match the pattern
of a prefix "OID" followed by 5 integer numbers.
}
\author{
Masoumeh Sheikhi
}
\keyword{internal}
