% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx_qcwarn_assays}
\alias{check_npx_qcwarn_assays}
\title{Help function checking data for assay QC warnings.}
\usage{
check_npx_qcwarn_assays(df, col_names)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{col_names}{A list of matched column names. This is the output of the
\var{check_npx_col_names} function.}
}
\value{
A character vector containing \var{OlinkID} of assays with at least
one QC warning, otherwise a \emph{character(0)}.
}
\description{
Help function checking data for assay QC warnings.
}
