% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx_col_names.R
\name{check_npx_update_col_names}
\alias{check_npx_update_col_names}
\title{Update column names based on preferred names}
\usage{
check_npx_update_col_names(preferred_names)
}
\arguments{
\item{preferred_names}{A named character vector where names are expected column names and values are current column names.}
}
\value{
An updated dictionary of column names.
}
\description{
This function updates the column names of the data based on the preferred names provided.
}
\details{
This function checks if the preferred column names are valid and updates the column names
according to the provided preferences.
}
\author{
Klev Diamanti
 Masoumeh Sheikhi
}
