% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{check_olink_data_type}
\alias{check_olink_data_type}
\title{Help function checking that \var{data_type} is expected.}
\usage{
check_olink_data_type(x, broader_platform = NULL)
}
\arguments{
\item{x}{The name of the Olink data type. One of
"Ct", "NPX", and "Quantified".}

\item{broader_platform}{Name of the broader Olink platform. One of
"NGS" and "qPCR".}
}
\value{
\code{NULL} if data_type is expected, otherwise an error.
}
\description{
Help function checking that \var{data_type} is expected.
}
\author{
Klev Diamanti
}
