% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{check_olink_platform}
\alias{check_olink_platform}
\title{Help function checking that \var{olink_platform} is expected.}
\usage{
check_olink_platform(
  x,
  broad_platform = NULL,
  data_type = NULL,
  quant_type = NULL
)
}
\arguments{
\item{x}{Name of Olink platform. One of
"Explore 3072", "Explore HT", "Flex", "Focus", "Reveal", "Target 48", and "Target 96".}

\item{broad_platform}{Name of the \var{broad_platform} to filter for. If
\code{NULL} all broader platforms are considered, otherwise one of
"NGS" and "qPCR" is expected.}

\item{data_type}{Name of the \var{data_type} to filter the Olink for. If
\code{NULL} all data types are considered, otherwise one of
"Ct", "NPX", and "Quantified" is expected.}

\item{quant_type}{Name of the \var{quant_type} to filter for. If \code{NULL} all
quantification types are considered, otherwise one of
"absolute" and "relative" is expected.}
}
\value{
\code{NULL} if platform is expected, otherwise an error.
}
\description{
Help function checking that \var{olink_platform} is expected.
}
\author{
Klev Diamanti
}
