% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_olinkid.R
\name{check_olinkid}
\alias{check_olinkid}
\title{Help function checking whether df contains recognized OIDs}
\usage{
check_olinkid(df, col_names)
}
\arguments{
\item{df}{A tibble or an arrow object containing containing column "OlinkID"}

\item{col_names}{A list of matched column names.
This is the output of `check_npx_col_names` function.}
}
\value{
A character vector including the invalid Olink IDs
}
\description{
Help function checking whether df contains recognized OIDs
}
\author{
Simon Forsberg
 Masoumeh Sheikhi
}
