% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_osi}
\alias{check_osi}
\title{Utility function to check OSI values for validity}
\usage{
check_osi(df, check_log, osi_score)
}
\arguments{
\item{df}{An Olink dataset.}

\item{check_log}{Output log of check_npx()}

\item{osi_score}{Name of OSI column to check}
}
\value{
An Olink dataset with the OSI column checked and cleaned
}
\description{
Utility function to check OSI values for validity
}
\keyword{internal}
