% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_out_df_arg.R
\name{check_out_df_arg}
\alias{check_out_df_arg}
\title{Help function checking that the requested output class of the read_npx*
functions is acceptable.}
\usage{
check_out_df_arg(out_df)
}
\arguments{
\item{out_df}{The class of output data frame to be returned. Accepted values
are "tibble" and "arrow".}
}
\value{
`TRUE` if the argument is as expected, and an error otherwise.
}
\description{
Help function checking that the requested output class of the read_npx*
functions is acceptable.
}
\seealso{
[read_npx()]
  [read_npx_delim()]
  [read_npx_zip()]
  [read_npx_parquet()]
}
\author{
Klev Diamanti
}
