% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_excluded_assay}
\alias{clean_excluded_assay}
\title{Clean Excluded Assays from NPX Data Based on Normalization Column}
\usage{
clean_excluded_assay(df)
}
\arguments{
\item{df}{A data.frame or tibble containing NPX data.}
}
\value{
A filtered data.frame or tibble where failed assays
(Normalization matches "excluded") are removed.
}
\description{
Removes rows in an NPX dataset where the Normalization status indicates
exclusion. This includes any case variation of the word "excluded".
}
\examples{
clean_data <- clean_excluded_assay(npx_data)

}
\author{
Kang Dong
}
