% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_helper.R
\name{clean_heatmap_df}
\alias{clean_heatmap_df}
\title{convert df from long to wide}
\usage{
clean_heatmap_df(df, check_log, colnames)
}
\arguments{
\item{df}{Data frame in long format with SampleID, NPX, OlinkID, Assay and
columns of choice for annotations.}

\item{check_log}{output from check_npx on \code{df}}

\item{colnames}{Character. Determines how to label the columns.
Must be 'assay', 'oid', or 'both' (default 'both').}
}
\value{
wide version of df
}
\description{
convert df from long to wide
}
\keyword{internal}
