% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_sample_id_controls}
\alias{clean_sample_id_controls}
\title{Clean Control Samples Based on SampleID}
\usage{
clean_sample_id_controls(df, check_npx_out_list)
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{check_npx_out_list}{A list returned from \code{check_npx()} with column
name mapping.}
}
\value{
A filtered data.frame or tibble.
}
\description{
Removes rows from an NPX dataset where SampleID contains the term "control"
(case-insensitive).
}
\author{
Kang Dong
}
