% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_sample_type_controls}
\alias{clean_sample_type_controls}
\title{Clean Control Samples Based on SampleType}
\usage{
clean_sample_type_controls(df, check_npx_out_list)
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{check_npx_out_list}{A list returned from \code{check_npx()}
with column name mapping.}
}
\value{
A filtered data.frame or tibble.
}
\description{
Removes rows from an NPX dataset where SampleType indicates control samples.
These include: "SAMPLE_CONTROL", "PLATE_CONTROL", and "NEGATIVE_CONTROL".
}
\author{
Kang Dong
}
