% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_helper.R
\name{color_heatmap}
\alias{color_heatmap}
\title{Title}
\usage{
color_heatmap(
  df,
  check_log,
  pheatmap_args,
  variable_row_list,
  variable_col_list
)
}
\arguments{
\item{df}{}

\item{check_log}{}

\item{pheatmap_args}{}

\item{variable_row_list}{}

\item{variable_col_list}{}
}
\description{
Title
}
\keyword{internal}
