% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_utils.R
\name{convert_read_npx_output}
\alias{convert_read_npx_output}
\title{Help function converting the output dataset from read_npx* functions to
"tibble" or "ArrowObject".}
\usage{
convert_read_npx_output(df, out_df)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}. The output of \code{\link[=read_npx]{read_npx()}} is an \code{olink_npx}
object that carries the check log as an attribute. If the check log
indicates that the data still needs cleaning (e.g. invalid identifiers,
duplicate samples), the downstream function will automatically clean the
data before proceeding.}

\item{out_df}{The class of the output dataset. One of
"tibble" or "arrow". Defaults to
"tibble".}
}
\value{
The dataset in the requested class.
}
\description{
Help function converting the output dataset from read_npx* functions to
"tibble" or "ArrowObject".
}
\author{
Klev Diamanti
}
\keyword{internal}
