% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_helper.R
\name{create_pheatmap_args}
\alias{create_pheatmap_args}
\title{Title}
\usage{
create_pheatmap_args(
  df_wide,
  df,
  check_log,
  center_scale,
  cluster_rows,
  cluster_cols,
  na_col,
  show_rownames,
  show_colnames,
  annotation_legend,
  fontsize,
  variable_row_list,
  variable_col_list,
  colnames,
  ...
)
}
\arguments{
\item{df_wide}{}

\item{df}{}

\item{check_log}{}

\item{center_scale}{}

\item{cluster_rows}{}

\item{cluster_cols}{}

\item{na_col}{}

\item{show_rownames}{}

\item{show_colnames}{}

\item{annotation_legend}{}

\item{fontsize}{}

\item{variable_row_list}{}

\item{variable_col_list}{}

\item{colnames}{}

\item{...}{}
}
\description{
Title
}
\keyword{internal}
