% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_data_type.R
\name{detect_data_type}
\alias{detect_data_type}
\title{Help function checking whether the input data contains
a column named "NPX" or "Quantified_value".}
\usage{
detect_data_type(df)
}
\arguments{
\item{df}{An arrow object}
}
\value{
A character string indicating the detected data type
("NPX" or "Quantified_value").
}
\description{
Help function checking whether the input data contains
a column named "NPX" or "Quantified_value".
}
\author{
Simon Forsberg
 Masoumeh Sheikhi
}
