% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_plate_randomizer.R
\name{displayPlateDistributions}
\alias{displayPlateDistributions}
\title{Plot distributions of a given variable for all plates}
\usage{
displayPlateDistributions(data, fill.color)
}
\arguments{
\item{data}{tibble/data frame in long format returned from the olink_plate_randomizer function.}

\item{fill.color}{Column name to be used as coloring variable for wells.}
}
\description{
Displays a bar chart for each plate representing the distribution of the given grouping variable on each plate using ggplot and ggplot2::geom_bar.
}
\examples{
\donttest{randomized.manifest <- olink_plate_randomizer(manifest)}
\donttest{displayPlateDistributions(data=randomized.manifest,fill.color="Site")}

}
\keyword{ggplot}
\keyword{plates,}
\keyword{randomized}
