% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_plate_randomizer.R
\name{displayPlateLayout}
\alias{displayPlateLayout}
\title{Plot all plates colored by a variable}
\usage{
displayPlateLayout(data, fill.color, PlateSize = 96, include.label = FALSE)
}
\arguments{
\item{data}{tibble/data frame in long format returned from the olink_plate_randomizer function.}

\item{fill.color}{Column name to be used as coloring variable for wells.}

\item{PlateSize}{Integer. Either 96 or 48. 96 is default.}

\item{include.label}{Should the variable group be shown in the plot.}
}
\description{
Displays each plate in a facet with cells colored by the given variable using ggplot and ggplot2::geom_tile.
}
\examples{
\donttest{randomized.manifest <- olink_plate_randomizer(manifest)}
\donttest{displayPlateLayout(data = randomized.manifest, fill.color="Site")}
}
\keyword{ggplot}
\keyword{plates,}
\keyword{randomized}
