% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_params}
\alias{.check_params}
\title{Common parameters for check functions.}
\usage{
.check_params(x, error)
}
\arguments{
\item{x}{Variable to check.}

\item{error}{Scalar boolean to return an error instead of a \code{FALSE}
(\code{default = FALSE}).}
}
\value{
Boolean, \code{TRUE} or \code{FALSE}, if the variable is of the correct class;
If output is \code{FALSE} and \var{error} = \code{TRUE}, an error is thrown.
}
\description{
Common parameters for check functions.
}
\author{
Klev Diamanti
}
\keyword{internal}
