% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{.get_olink_data_details}
\alias{.get_olink_data_details}
\title{Common parameters for getter functions in this file.}
\usage{
.get_olink_data_details(broad_platform, platform_name, data_type, quant_type)
}
\arguments{
\item{broad_platform}{Name of the \var{broad_platform} to filter for. If
\code{NULL} all broader platforms are considered, otherwise one of
"NGS" and "qPCR" is expected.}

\item{platform_name}{Name of the \var{platform_name} to filter for. If \code{NULL}
all platforms are considered, otherwise one of
"Explore 3072", "Explore HT", "Flex", "Focus", "Reveal", "Target 48", and "Target 96" is expected.}

\item{data_type}{Name of the \var{data_type} to filter the Olink for. If
\code{NULL} all data types are considered, otherwise one of
"Ct", "NPX", and "Quantified" is expected.}

\item{quant_type}{Name of the \var{quant_type} to filter for. If \code{NULL} all
quantification types are considered, otherwise one of
"absolute" and "relative" is expected.}
}
\description{
Common parameters for getter functions in this file.
}
\author{
Klev Diamanti
}
