% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.read_npx_args}
\alias{.read_npx_args}
\title{Common parameters for read_npx-related functions.}
\usage{
.read_npx_args(
  filename,
  file,
  out_df,
  long_format,
  olink_platform,
  data_type,
  .ignore_files,
  quiet
)
}
\arguments{
\item{filename}{Path to Olink software output file in wide or long format.
Expecting extensions "xls" or "xlsx" for excel files, "csv" or "txt" for delim files, "parquet" for parquet files, and "zip" for compressed files.}

\item{file}{Path to Olink software output file in wide or long format.
Expecting extensions "xls" or "xlsx" for excel files, "csv" or "txt" for delim files, "parquet" for parquet files, and "zip" for compressed files.}

\item{out_df}{The class of the output dataset. One of
"tibble" or "arrow". (default =
"tibble")}

\item{long_format}{Boolean marking format of input file. One of \code{NULL}
(default) for auto-detection, or \code{TRUE} for long format and \code{FALSE} for
wide format files.}

\item{olink_platform}{Olink platform used to generate the input file.
One of \code{NULL} (default) for auto-detection,
or "Target 48", "Flex", "Target 96", "Explore 3072", "Explore HT", "Focus", and "Reveal".}

\item{data_type}{Quantification method of the input data. One of \code{NULL}
(default) for auto-detection, "Ct", "NPX", and "Quantified".}

\item{.ignore_files}{Character vector of files included in the zip-compressed
Olink software output files that should be ignored. Used only for
zip-compressed input files (default = \emph{c("README.txt")}).}

\item{quiet}{Boolean to print a confirmation message when reading the input
file. Applies to excel or delimited input only. \code{TRUE} (default) to not print
and \code{FALSE} to print.}
}
\value{
Dataset,
"tibble" or "ArrowObject", with Olink
data in long format.
}
\description{
Common parameters for read_npx-related functions.
}
\author{
Klev Diamanti
}
