% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_helper.R
\name{extract_ellipsis_arg}
\alias{extract_ellipsis_arg}
\title{Title}
\usage{
extract_ellipsis_arg(pheatmap_args, ...)
}
\arguments{
\item{pheatmap_args}{}

\item{...}{}
}
\description{
Title
}
\keyword{internal}
