% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_plate_randomizer.R
\name{generate_plate_holder}
\alias{generate_plate_holder}
\title{Create empty plate layout}
\usage{
generate_plate_holder(
  nplates,
  nspots,
  nsamples,
  plate_size,
  num_ctrl,
  rand_ctrl
)
}
\arguments{
\item{nplates}{number of plates}

\item{nspots}{number of spots on each plate}

\item{nsamples}{number of samples}

\item{plate_size}{size of plate}

\item{num_ctrl}{number of controls}

\item{rand_ctrl}{if controls are randomized}
}
\value{
plate layout including plates, rows, and columns of available wells
}
\description{
Create empty plate layout
}
\keyword{internal}
