% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_ext.R
\name{get_accepted_file_ext_summary}
\alias{get_accepted_file_ext_summary}
\title{Describes acceptable file extension for each file type.}
\usage{
get_accepted_file_ext_summary()
}
\value{
A scalar character vector with one sentence describing the acceptable
extensions of each file type.
}
\description{
Describes acceptable file extension for each file type.
}
