% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{get_all_olink_broader_platforms}
\alias{get_all_olink_broader_platforms}
\title{Get names of all broader Olink platforms.}
\usage{
get_all_olink_broader_platforms()
}
\value{
A character vector with names of all Olink broader platforms.
}
\description{
Get names of all broader Olink platforms.
}
\author{
Klev Diamanti
}
