% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{get_all_olink_data_types}
\alias{get_all_olink_data_types}
\title{Get names of all Olink quantification methods (data types).}
\usage{
get_all_olink_data_types()
}
\value{
A character vector with names of all Olink quantification methods
(data types).
}
\description{
Get names of all Olink quantification methods (data types).
}
\author{
Klev Diamanti
}
