% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{get_all_olink_quant_types}
\alias{get_all_olink_quant_types}
\title{Get names of all Olink quantification types.}
\usage{
get_all_olink_quant_types()
}
\value{
A character vector with names of all Olink quantification types.
}
\description{
Get names of all Olink quantification types.
}
\author{
Klev Diamanti
}
