% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_others.R
\name{get_alt_colnames}
\alias{get_alt_colnames}
\title{Gets alternative column names for a given column key.}
\usage{
get_alt_colnames(col_key)
}
\value{
A character vector of alternative column names corresponding to the
provided column key.
}
\description{
Gets alternative column names for a given column key.
}
\examples{
\donttest{
  OlinkAnalyze:::get_alt_colnames(col_key = "sample_id")
}
}
\keyword{internal}
