% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_checksum_file.R
\name{get_checksum_file}
\alias{get_checksum_file}
\title{Help function to get the name of the checksum file from a zip NPX.}
\usage{
get_checksum_file(files)
}
\arguments{
\item{files}{A vector of file names without any path prefix.}
}
\value{
The name of checksum file or NA if the file is not present.
}
\description{
Help function to get the name of the checksum file from a zip NPX.
}
\seealso{
[check_checksum()]
  [get_npx_file()]
}
\author{
Klev Diamanti
}
\keyword{MD5}
\keyword{NPX}
\keyword{SHA256}
\keyword{checksum}
\keyword{zip}
