% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_output.R
\name{get_df_output}
\alias{get_df_output}
\title{Prints options for \var{out_df}.}
\usage{
get_df_output()
}
\value{
A scalar character vector with the input options for \var{out_df}.
}
\description{
Prints options for \var{out_df}.
}
