% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_output.R
\name{get_df_output_print}
\alias{get_df_output_print}
\title{Prints class type output from read_npx* functions.}
\usage{
get_df_output_print()
}
\value{
A scalar character vector with the class type of outputs from
read_npx* functions.
}
\description{
Prints class type output from read_npx* functions.
}
