% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_others.R
\name{get_file_ext}
\alias{get_file_ext}
\title{Gets all file extensions based on the file format.}
\usage{
get_file_ext(name_sub = NULL)
}
\arguments{
\item{name_sub}{Substring of file format. One of
"excel", "delim", "parquet", "compressed", and "NULL". If \code{NULL} all
file extensions are returned.}
}
\value{
Character vector with accepted file extensions.
}
\description{
Gets all file extensions based on the file format.
}
