% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_others.R
\name{get_file_formats}
\alias{get_file_formats}
\title{Get all acceptable file formats.}
\usage{
get_file_formats()
}
\value{
A character vector with the acceptable file formats.
}
\description{
Get all acceptable file formats.
}
\author{
Klev Diamanti
}
