% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_separator.R
\name{get_file_separator}
\alias{get_file_separator}
\title{Help function to get the file separator.}
\usage{
get_file_separator(file)
}
\arguments{
\item{file}{Path to Olink software output in txt or csv.}
}
\value{
The file separator comma (;) or semicolon (;)
}
\description{
Help function to get the file separator.
}
