% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_npx_file_from_zip.R
\name{get_npx_file_from_zip}
\alias{get_npx_file_from_zip}
\title{Helper function to get the name of the NPX file from a zip NPX.}
\usage{
get_npx_file_from_zip(
  files,
  accepted_npx_files = c("xls", "xlsx", "csv", "txt", "parquet")
)
}
\arguments{
\item{files}{A vector of file names without any path prefix.}

\item{accepted_npx_files}{Vector of accepted extensions of NPX files.}
}
\value{
The name of the NPX file.
}
\description{
Helper function to get the name of the NPX file from a zip NPX.
}
