% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{get_olink_ngs_data_types}
\alias{get_olink_ngs_data_types}
\title{Get names of current data types for NGS Olink platforms.}
\usage{
get_olink_ngs_data_types()
}
\value{
A character vector with names of the quantification methods for NGS
Olink platforms.
}
\description{
Get names of current data types for NGS Olink platforms.
}
