% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{get_olink_ngs_platforms}
\alias{get_olink_ngs_platforms}
\title{Get names of current NGS Olink platforms.}
\usage{
get_olink_ngs_platforms()
}
\value{
A character vector with names of NGS Olink platforms.
}
\description{
Get names of current NGS Olink platforms.
}
