% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{get_olink_platforms}
\alias{get_olink_platforms}
\title{Get names of selected Olink platforms.}
\usage{
get_olink_platforms(broad_platform = NULL, data_type = NULL, quant_type = NULL)
}
\arguments{
\item{broad_platform}{Name of the \var{broad_platform} to filter for. If
\code{NULL} all broader platforms are considered, otherwise one of
"NGS" and "qPCR" is expected.}

\item{data_type}{Name of the \var{data_type} to filter the Olink for. If
\code{NULL} all data types are considered, otherwise one of
"Ct", "NPX", and "Quantified" is expected.}

\item{quant_type}{Name of the \var{quant_type} to filter for. If \code{NULL} all
quantification types are considered, otherwise one of
"absolute" and "relative" is expected.}
}
\value{
A character vector with names of Olink platforms filtered by
\var{broad_platform}, \var{data_type}, and \var{quant_type}.
}
\description{
Get names of selected Olink platforms.
}
\author{
Klev Diamanti
}
