% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{get_olink_quant_types}
\alias{get_olink_quant_types}
\title{Get names of selected Olink quantification types.}
\usage{
get_olink_quant_types(
  broad_platform = NULL,
  platform_name = NULL,
  data_type = NULL
)
}
\arguments{
\item{broad_platform}{Name of the \var{broad_platform} to filter the Olink
quantification types for. If \code{NULL} all relevant Olink quantification types
are returned, otherwise one of
"NGS" and "qPCR" is expected.}

\item{platform_name}{Name of the \var{platform_name} to filter the Olink
quantification types for. If \code{NULL} all relevant Olink quantification types
are returned, otherwise one of "Explore 3072", "Explore HT", "Flex", "Focus", "Target 48", and "Target 96"
is expected.}

\item{data_type}{Name of the \var{data_type} to filter the Olink
quantification types for. If \code{NULL} all relevant Olink quantification types
are returned, otherwise one of "Ct", "NPX", and "Quantified"
is expected.}
}
\value{
A character vector with names of Olink quantification types filtered
by \var{broad_platform}, \var{platform_name}, and \var{data_type}.
}
\description{
Get names of selected Olink quantification types.
}
\author{
Klev Diamanti
}
