% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_helper.R
\name{lst_styling}
\alias{lst_styling}
\title{Title}
\usage{
lst_styling(i, x, styling_location, styling)
}
\arguments{
\item{i}{}

\item{x}{}

\item{styling_location}{}

\item{styling}{}
}
\description{
Title
}
\keyword{internal}
