% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{make_cyt_graph}
\alias{make_cyt_graph}
\title{send topGO GO terms to cytoscape}
\usage{
make_cyt_graph(
  pvals,
  go_annot_list,
  test_stat = "ks",
  graph_name = "myGraph",
  firstSigNodes = 10
)
}
\description{
send topGO GO terms to cytoscape
}
