% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manifest}
\alias{manifest}
\title{Example Sample Manifest}
\format{
This dataset contains columns:
\describe{
\item{SubjectID}{Subject Identifier, A-Z}
\item{Visit}{Visit Number, 1-6}
\item{SampleID}{138 unique sample IDs}
\item{Site}{Site1 or Site2}
}
}
\usage{
manifest
}
\description{
Synthetic sample manifest to demonstrate use of functions in this package.
}
\details{
A tibble with 138 rows and 4 columns. This manifest contains 26 example
subjects, with 6 visits and 2 sites.
}
\keyword{datasets}
