% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping}
\alias{mapping}
\title{A file to map OlinkIDs in Explore 3072 to OlinkIDs in Explore HT}
\format{
This dataset  contains columns:
\describe{
  \item{Gene}{the gene name of the assay}
  \item{OlinkID}{the OlinkID in Explore HT}
  \item{OlinkID_Explore384}{the OlinkID in Explore 384}
  \item{UniProt}{the UniProt ID in Explore HT}
  \item{Block}{the block in Explore HT}
  \item{Panel_Explore384}{the panel in Explore 384}
  \item{Block_Explore384}{the block in Explore HT}
  }
}
\usage{
mapping
}
\description{
A file to map OlinkIDs in Explore 3072 to OlinkIDs in Explore HT
}
\details{
A tibble with 2920 rows and 7 columns.
}
\keyword{datasets}
