% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_utils.R
\name{mapping_file_id}
\alias{mapping_file_id}
\title{Identifying which mapping file to use}
\usage{
mapping_file_id(prod_uniq)
}
\arguments{
\item{prod_uniq}{Name of products (not_ref, ref)}
}
\value{
dataframe of mapping file to use for OlinkID mapping
(eHT_e3072_mapping, reveal_eht_mapping or reveal_e3072_mapping)
}
\description{
Identifying which mapping file to use
}
\keyword{internal}
