% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_2.R
\name{norm_internal_adjust_ref}
\alias{norm_internal_adjust_ref}
\title{Modify the reference dataset to be combined with the non-reference normalized
dataset}
\usage{
norm_internal_adjust_ref(ref_df, ref_name)
}
\arguments{
\item{ref_df}{The reference dataset to be used in normalization (required).}

\item{ref_name}{Project name of the reference dataset (required).}
}
\value{
Tibble or ArrowObject with the reference dataset with additional
columns "Project" and "Adj_factor".
}
\description{
Modify the reference dataset to be combined with the non-reference normalized
dataset
}
\author{
Klev Diamanti
}
