% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization.R
\name{norm_internal_rename_cols}
\alias{norm_internal_rename_cols}
\title{Update column names of non-reference dataset based on those of reference
dataset}
\usage{
norm_internal_rename_cols(ref_cols, not_ref_cols, not_ref_df)
}
\arguments{
\item{ref_cols}{Named list of column names identified in the reference
dataset.}

\item{not_ref_cols}{Named list of column names identified in the
non-reference dataset.}

\item{not_ref_df}{Non-reference dataset to be used in normalization.}
}
\value{
\code{not_ref_df} with updated column names.
}
\description{
This function handles cases when specific columns referring to the same thing
are named differently in \code{df1} and \code{df2} normalization datasets. It only
renames columns
panel_version, qc_warn, and assay_warn based on
their names in the reference dataset.#'
}
\author{
Klev Diamanti
}
