% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization.R
\name{norm_internal_update_maxlod}
\alias{norm_internal_update_maxlod}
\title{Update MaxLOD to the maximum MaxLOD across normalized datatsets.}
\usage{
norm_internal_update_maxlod(df, cols)
}
\arguments{
\item{df}{Normalized Olink dataset (required).}

\item{cols}{Named list of column names in the dataset (required).}
}
\value{
The same dataset as the input \var{df} with the column reflecting
MaxLOD updated.
}
\description{
Update MaxLOD to the maximum MaxLOD across normalized datatsets.
}
