% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npx_data1}
\alias{npx_data1}
\title{NPX Data in Long format.}
\format{
In addition to standard read_npx() columns, this dataset also
contains columns:
\describe{
\item{Subject}{Subject Identifier}
\item{Treatment}{ Treated or Untreated}
\item{Site}{Site indicator, 5 unique values}
\item{Time}{Baseline, Week.6 and Week.12}
\item{Project}{Project ID number}
}
}
\usage{
npx_data1
}
\description{
This is a synthetic dataset aiming to use-cases of functions from this
package.
}
\details{
A tibble with 29,440 rows and 17 columns.

\var{npx_data1} is an Olink NPX data file (tibble) in long format with 158
unique Sample identifiers (including 2 repeats each of control samples:
CONTROL_SAMPLE_AS 1 and CONTROL_SAMPLE_AS 2). The data also contains 1104
assays uniquely identified using OlinkID over 2 Olink Panels.
}
\keyword{datasets}
