% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_plot.R
\name{olink_autoencoder_plot}
\alias{olink_autoencoder_plot}
\title{Function to make an Autoencoder plot from the data}
\usage{
olink_autoencoder_plot(
  df,
  color_g = "QC_Warning",
  label_samples = FALSE,
  activation = "Tanh",
  max_mem_size = "5g",
  layers = 2,
  drop_assays = FALSE,
  drop_samples = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{df}{data frame in long format with Sample Id, NPX and column of choice for colors}

\item{color_g}{Character value indicating which column to use for colors (default QC_Warning)}

\item{label_samples}{Logical. If TRUE, points are replaced with SampleID (default FALSE)}

\item{activation}{Character string specifying the activation function in the neural network. Must be one of: "Tanh", "TanhWithDropout", "Rectifier", "RectifierWithDropout", "Maxout", "MaxoutWithDropout".}

\item{max_mem_size}{Character string specifying the maximum size, in bytes, of the memory allocation pool to H2O}

\item{layers}{Numeric vector specifying the sizes of the hidden layers. The number of layers must be odd, and the coding layer (assumed to be the middle one) must be 2. E.g. 2, c(x, 2, x), etc. (default 2, i.e. one hidden layer of size 2)}

\item{drop_assays}{Logical. All assays with any missing values will be dropped. Takes precedence over sample drop.}

\item{drop_samples}{Logical. All samples with any missing values will be dropped.}

\item{verbose}{Logical. Whether warnings about the number of samples and/or assays dropped or imputed should be printed to the console.}

\item{...}{coloroption passed to specify color order.}
}
\value{
An object of class "ggplot"
}
\description{
NOTE: This is a beta version.
Trains a neural network using H2O and plots the learned features, referred to as codings, in a 2 dimensional figure.
Unique sample names are required and imputation by the median is done for assays with missingness <10\% for multi-plate projects and <5\% for single plate projects.
}
\examples{
\donttest{
npx_data <- npx_data1 \%>\%
    mutate(SampleID = paste(SampleID, "_", Index, sep = ""))
olink_pca_plot(df=npx_data, color_g = "QC_Warning")}
}
\keyword{NPX,}
\keyword{UMAP}
