% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_boxplot_annova.R
\name{olink_boxplot_annova}
\alias{olink_boxplot_annova}
\title{Function which plots boxplots of a selected variable}
\usage{
olink_boxplot_annova(
  df,
  variable,
  olinkid_list,
  posthoc.results,
  verbose = F,
  ...
)
}
\arguments{
\item{df}{NPX data frame in long format with at least protein name (Assay), OlinkID (unique), UniProt and a grouping variable.}

\item{variable}{A character value indiciating which column to use as the x-axis grouping variable}

\item{olinkid_list}{Character vector indicating which proteins (OlinkIDs) to plot.}

\item{posthoc.results}{Results from olink_anova_posthoc function.}

\item{verbose}{Boolean. If the plots are shown as well as returned in the list (default is false).}

\item{...}{coloroption passed to specify color order}
}
\value{
A list of objects of class “ggplot” (the actual ggplot object is entry 1 in the list).
}
\description{
Generates faceted boxplots of NPX vs. grouping variable for a given list of proteins (OlinkIDs) using ggplot and ggplot2::geom_boxplot.
}
\examples{
\donttest{
# calculate the p-value for the ANOVA
anova_results_oneway <- olink_anova(df = npx_data1, 
                                   variable = 'Site')
# extracting the significant proteins
anova_results_oneway_significant <- anova_results_oneway \%>\%
 filter(Threshold == 'Significant') \%>\%
 pull(OlinkID)
anova_posthoc_oneway_results <- olink_anova_posthoc(df = npx_data1,
                                                   olinkid_list = anova_results_oneway_significant,
                                                   variable = 'Site',
                                                   effect = 'Site')
olink_boxplot_annova(npx_data1,
                    variable="Site",
                    olinkid_list=anova_results_oneway_significant,
                    posthoc.results= anova_posthoc_oneway_results,
                    verbose = F)}
}
