% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_boxplot_ttest.R
\name{olink_boxplot_ttest}
\alias{olink_boxplot_ttest}
\title{Function which plots boxplots of a selected variable}
\usage{
olink_boxplot_ttest(
  df,
  TtestResult,
  variable,
  olinkid_list,
  verbose = F,
  pair_id,
  ...
)
}
\arguments{
\item{df}{NPX data frame in long format with at least protein name (Assay), OlinkID (unique), UniProt and a grouping variable.}

\item{TtestResult}{The output dataframe from Olink_ttest function.}

\item{variable}{A character value indiciating which column to use as the x-axis grouping variable}

\item{olinkid_list}{Character vector indicating which proteins (OlinkIDs) to plot.}

\item{verbose}{Boolean. If the plots are shown as well as returned in the list (default is false).}

\item{pair_id}{Character value indicating which column indicates the paired sample identifier.}

\item{...}{coloroption passed to specify color order}
}
\value{
A list of objects of class “ggplot” (the actual ggplot object is entry 1 in the list).
}
\description{
Generates faceted boxplots of NPX vs. grouping variable for a given list of proteins (OlinkIDs) using ggplot and ggplot2::geom_boxplot.
}
\examples{
\donttest{
npx_df <- npx_data1 \%>\% filter(!grepl('control',SampleID, ignore.case = TRUE))
ttest_results <- olink_ttest(df=npx_df,
                             variable = 'Treatment',
                             alternative = 'two.sided')
significant_assays <- ttest_results \%>\%
    filter(Threshold == 'Significant') \%>\%
    pull(OlinkID)
olink_boxplot_ttest(npx_df, 
TtestResult = ttest_results, 
variable = 'Treatment',
olinkid_list = significant_assays)}
}
