% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_bridge_selector.R
\name{olink_bridge_selector}
\alias{olink_bridge_selector}
\alias{olink_bridgeselector}
\title{Bridge selection function}
\usage{
olink_bridge_selector(df, sample_missing_freq, n, check_log = NULL)

olink_bridgeselector(df, ..., n, check_log = NULL)
}
\arguments{
\item{df}{Tibble/data frame in long format such as produced by the
Olink Analyze read_npx function.}

\item{sample_missing_freq}{The threshold for sample wise missingness.}

\item{n}{Number of bridge samples to be selected.}

\item{check_log}{Optional named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
an attached \code{check_log} is used when present; otherwise \code{\link[=check_npx]{check_npx()}} will
be run internally using \code{df}.}

\item{...}{Additional arguments. Currently only accepts \code{sampleMissingFreq}
for backward compatibility. Please use \code{sample_missing_freq} instead of
\code{sampleMissingFreq} in the future.}
}
\value{
A "tibble" with sample IDs and mean NPX for a defined number of
bridging samples. Columns include:
\itemize{
\item{SampleID:} Sample ID
\item{PercAssaysBelowLOD:} Percent of Assays that are below LOD for
the sample
\item{MeanNPX:} Mean NPX for the sample
}
}
\description{
The bridge selection function will select a number of bridge samples based
on the input data. It selects samples with good detection that pass QC
and cover a good range of the data. If possible, Olink recommends 8-16
bridge samples. When running the selector, Olink recommends starting at
sample_missing_freq = 0.10 which represents a maximum of 10\\% data below LOD
per sample. If there are not enough samples output, increase to 20\\%. \cr\cr
The function accepts NPX Excel files with data < LOD replaced.
}
\details{
\code{olink_bridgeselector()} is a synonym of \code{olink_bridge_selector()} .
}
\examples{
\donttest{
  bridge_samples <- OlinkAnalyze::olink_bridge_selector(
    df = OlinkAnalyze::npx_data1,
    sample_missing_freq = 0.1,
    n = 20L
  )
}

}
