% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_normalization.R
\name{olink_bridging}
\alias{olink_bridging}
\title{Normalization of all proteins (by OlinkID).}
\usage{
olink_bridging(
  df1,
  df2,
  overlapping_samples,
  df1_project_nr = "P1",
  df2_project_nr = "P2",
  reference_project = "P1"
)
}
\arguments{
\item{df1}{First dataframe to be used in normalization (required).}

\item{df2}{Second dataframe to be used in normalization (required).}

\item{overlapping_samples}{Samples to be used for adjustment factor calculation (required).}

\item{df1_project_nr}{Project name of first dataset.}

\item{df2_project_nr}{Project name of second dataset.}

\item{reference_project}{Project name of reference_project. Needs to be the same as either df1_project_nr or df2_project_nr. The project to which the second project is adjusted to.}
}
\value{
A tibble of NPX data in long format containing normalized NPX values, including adjustment factors.
}
\description{
Normalizes an NPX data frame to another data frame using overlapping samples (bridge samples). By default, the older data frame will be used as reference.\cr\cr
The overlapping samples need to be named the same in the two data frames and adjustment is made using the median of the paired differences between the bridge samples in the two data frames.
The two data frames are inputs df1 and df2, the one being adjusted to is specified in the input reference_project, and the overlapping samples are specified in overlapping_samples.
}
\examples{
\donttest{

library(dplyr)

npx_df1 <- npx_data1 \%>\% dplyr::mutate(Project = 'P1')
npx_df2 <- npx_data2 \%>\% dplyr::mutate(Project = 'P2')

# Find overlapping samples, but exclude Olink control
overlap_samples <- intersect((npx_df1 \%>\%
                               dplyr::filter(!grepl("control", SampleID,
                                                     ignore.case=TRUE)))$SampleID,
                             (npx_df2 \%>\%
                               dplyr::filter(!grepl("control", SampleID,
                                                     ignore.case=TRUE)))$SampleID)
# Normalize
olink_normalization(df1 = npx_df1,
                    df2 = npx_df2,
                    overlapping_samples_df1 = overlap_samples,
                    df1_project_nr = 'P1',
                    df2_project_nr = 'P2',
                    reference_project = 'P1')
}
}
\keyword{Normalization}
