% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_class.R
\name{olink_check_log}
\alias{olink_check_log}
\title{Retrieve the check log from an Olink data object}
\usage{
olink_check_log(df)
}
\arguments{
\item{df}{A data object, typically from \code{\link[=read_npx]{read_npx()}}, \code{\link[=clean_npx]{clean_npx()}} or
\code{\link[=olink_normalization]{olink_normalization()}}.}
}
\value{
A named list (the check log) or \code{NULL} if no check log is attached.
}
\description{
Extracts the check log from an \code{olink_class} tibble or an ArrowObject with
embedded metadata. Returns \code{NULL} if no check log is found.
}
\examples{
\dontrun{
# read and check data
  npx_file <- system.file(
    "extdata",
    "npx_data_ext.parquet",
    package = "OlinkAnalyze"
  )

  npx_df <- OlinkAnalyze::read_npx(
    filename = npx_file
  )

  check_log <- OlinkAnalyze::check_npx(
    df = npx_df
  )

  # attach check_log to the data
  npx_obj <- OlinkAnalyze::new_olink_class(
    df = npx_df,
    check_log = check_log
  )

  # retrieve check_log
  OlinkAnalyze::olink_check_log(
    df = npx_obj
  )
}

}
