% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_color.R
\name{olink_color_discrete}
\alias{olink_color_discrete}
\title{Olink color scale for discrete ggplots}
\usage{
olink_color_discrete(..., alpha = 1, coloroption = NULL)
}
\arguments{
\item{...}{Optional. Additional arguments to pass to
\code{ggplot2::discrete_scale()}}

\item{alpha}{transparency (optional)}

\item{coloroption}{string, one or more of the following:
c("red", "orange", "yellow", "green", "teal", "turqoise", "lightblue",
"darkblue", "purple", "pink")}
}
\value{
No return value, called for side effects
}
\description{
Olink color scale for discrete ggplots
}
\examples{
\donttest{
ggplot2::ggplot(
  data = datasets::mtcars,
  mapping = ggplot2::aes(
    x = .data[["wt"]],
    y = .data[["mpg"]],
    color = as.factor(x = .data[["cyl"]])
  )
) +
  ggplot2::geom_point(
    size = 4L
  ) +
  OlinkAnalyze::olink_color_discrete() +
  ggplot2::theme_bw()

ggplot2::ggplot(
  data = datasets::mtcars,
  mapping = ggplot2::aes(
    x = .data[["wt"]],
    y = .data[["mpg"]],
    color = as.factor(x = .data[["cyl"]])
  )
) +
  ggplot2::geom_point(
    size = 4L
  ) +
  OlinkAnalyze::olink_color_discrete(
    coloroption = c("lightblue", "red", "green")
  ) +
  ggplot2::theme_bw()
}

}
