% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_color.R
\name{olink_color_gradient}
\alias{olink_color_gradient}
\title{Olink color scale for continuous ggplots}
\usage{
olink_color_gradient(..., alpha = 1, coloroption = NULL)
}
\arguments{
\item{...}{Optional. Additional arguments to pass to
\code{ggplot2::scale_color_gradientn()}}

\item{alpha}{transparency (optional)}

\item{coloroption}{string, one or more of the following:
c("red", "orange", "yellow", "green", "teal", "turqoise", "lightblue",
"darkblue", "purple", "pink")}
}
\value{
No return value, called for side effects
}
\description{
Olink color scale for continuous ggplots
}
\examples{
\donttest{
ggplot2::diamonds |>
  dplyr::filter(
    .data[["x"]] > 5
    & .data[["x"]] < 6
    & .data[["y"]] > 5
    & .data[["y"]] < 6
  ) |>
  dplyr::mutate(
    diff = sqrt(
      x = abs(
        x = .data[["x"]] - .data[["y"]]
      )
    ) * sign(
      x = .data[["x"]] - .data[["y"]]
    )
  ) |>
  ggplot2::ggplot(
    mapping = ggplot2::aes(
      x = .data[["x"]],
      y = .data[["y"]],
      colour = .data[["diff"]]
    )
  ) +
  ggplot2::geom_point() +
  ggplot2::theme_bw() +
  OlinkAnalyze::olink_color_gradient()
}

}
