% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{olink_consensus_clustering}
\alias{olink_consensus_clustering}
\title{consensus clustering}
\usage{
olink_consensus_clustering(df, clusteralg = "km", iters = 25, reps = 100)
}
\description{
identify what k should be set to
}
