% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_heatmaps.R
\name{olink_corrplot}
\alias{olink_corrplot}
\title{Create corrplot for olink proteomics data}
\usage{
olink_corrplot(
  df,
  t = FALSE,
  plot_labels = FALSE,
  GeneIDs = NULL,
  SampleIDs = NULL,
  verbose = TRUE,
  clusters = NA
)
}
\arguments{
\item{t}{(logical) whether to transpose prior to generation of correlation matrix -
by default will be sample vs sample}
}
\value{
pheatmap object
}
\description{
Create corrplot for olink proteomics data
}
