% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_color.R
\name{olink_fill_gradient}
\alias{olink_fill_gradient}
\title{Olink fill scale for continuous ggplots}
\usage{
olink_fill_gradient(..., alpha = 1, coloroption = NULL)
}
\arguments{
\item{alpha}{transparency (optional)}

\item{coloroption}{string, one or more of the following:
c('red', 'orange', 'yellow', 'green', 'teal', 'turqoise', 'gray', 'darkblue', 'purple', 'pink')}
}
\value{
None
}
\description{
Olink fill scale for continuous ggplots
}
\examples{
dsub <- subset(diamonds, x > 5 & x < 6 & y > 5 & y < 6) 
dsub$diff <- with(dsub, sqrt(abs(x-y))* sign(x-y)) 
ggplot(dsub, aes(x, y, colour=diff)) + 
geom_point() + 
 theme_bw() + 
 olink_fill_gradient()
}
