% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_format.R
\name{olink_format_oid_no_overlap}
\alias{olink_format_oid_no_overlap}
\title{Retrieve non-overlapping assays between two NPX datasets}
\usage{
olink_format_oid_no_overlap(lst_check)
}
\arguments{
\item{lst_check}{Normalization input list checks generated by
\code{olink_norm_input_check}.}
}
\value{
A combined "tibble" of Olink data in long format containing only the
non-overlapping assays from each input dataset.
}
\description{
For use in \code{olink_normalization_format} function. Generates a message stating
how many assays were not overlapping. Appends additional columns depending on
the normalization type to match normalized data output. For cross-product
normalization, splits any concatenated OlinkIDs.
}
\author{
Danai Topouza
Klev Diamanti
}
