% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_format.R
\name{olink_format_rm_ext_ctrl}
\alias{olink_format_rm_ext_ctrl}
\title{Remove negative controls and plate controls from dataset. For use in
olink_normalization_format function. Generates a message stating which
control samples were removed.}
\usage{
olink_format_rm_ext_ctrl(df, lst_check)
}
\arguments{
\item{df}{NPX dataset to be processed.}

\item{lst_check}{Normalization input list checks generated by
\code{olink_norm_input_check}.}
}
\value{
A "tibble" of Olink data in long format containing the input dataset
with negative controls and plate controls removed.
}
\description{
Remove negative controls and plate controls from dataset. For use in
olink_normalization_format function. Generates a message stating which
control samples were removed.
}
\author{
Danai G. Topouza
Klev Diamanti
}
