% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_heatmaps.R
\name{olink_heatmap}
\alias{olink_heatmap}
\title{Create heatmap for olink proteomics data}
\usage{
olink_heatmap(
  df,
  GeneIDs = NULL,
  SampleIDs = NULL,
  verbose = TRUE,
  label_rows = FALSE,
  label_cols = FALSE,
  row_clusters = NA,
  col_clusters = NA
)
}
\value{
pheatmap object
}
\description{
Create heatmap for olink proteomics data
}
