% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_heatmap.R
\name{olink_heatmap_plot}
\alias{olink_heatmap_plot}
\title{Function to plot a heatmap of the NPX data}
\usage{
olink_heatmap_plot(
  df,
  variable_list = NULL,
  scaledata = T,
  cluster_rows = T,
  cluster_cols = T,
  show_rownames = T,
  show_colnames = T,
  annotation_legend = T,
  fontsize = 10,
  na_col = "black",
  ...
)
}
\arguments{
\item{df}{data frame in long format with SampleID, NPX, OlinkID, Assay and columns of choice for annotations.}

\item{variable_list}{coloumns in df to be annotated.}

\item{scaledata}{Logical. If data should be scaled and centered across assays (default TRUE).}

\item{cluster_rows}{Logical. Determining if rows should be clustered (default TRUE).}

\item{cluster_cols}{Logical. Determining if columns should be clustered (default TRUE).}

\item{show_rownames}{Logical. Determining if row names are shown (default TRUE).}

\item{show_colnames}{Logical. Determining if column names are shown (default TRUE).}

\item{annotation_legend}{Logical. Determain if legend for annotations should be shown (default TRUE).}

\item{fontsize}{Integer. Fontsize.}

\item{na_col}{Color of cells with NA (default black).}

\item{...}{additional arguments used in pheatmap::pheatmap}
}
\description{
Generates a heatmap using pheatmap::pheatmap of all samples from NPX data.
}
\details{
The values are by default scaled across and centered in the heatmap. Columns
and rows are by default sorted by by dendrogram.
Unique sample names are required.
}
\keyword{Heatmap}
\keyword{NPX}
