% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_outlier_detection_utils.R
\name{olink_iqr}
\alias{olink_iqr}
\title{Compute inter-quartile range (IQR) of multiplied by a fixed value}
\usage{
olink_iqr(df, quant_col, iqr_group, iqr_sd)
}
\arguments{
\item{df}{Olink dataset}

\item{quant_col}{Character vector of name of quantification column}

\item{iqr_group}{Grouping for which to compute IQR for}

\item{iqr_sd}{Fixed value to multiply IQR with}
}
\value{
Input dataset with two additional columns, iqr and iqr_sd
}
\description{
Compute inter-quartile range (IQR) of multiplied by a fixed value
}
