% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{olink_lm}
\alias{olink_lm}
\title{Function which performs a standard linear model per protein}
\usage{
olink_lm(
  df,
  variable,
  outcome = "NPX",
  covariates = NULL,
  return.covariates = F,
  return.models = FALSE,
  verbose = T,
  reorder = NULL
)
}
\description{
Function which performs a standard linear model per protein
}
