% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mixed_model.R
\name{olink_lmer_main}
\alias{olink_lmer_main}
\title{Function which performs a linear mixed model per protein}
\usage{
olink_lmer_main(df, var1, var2, ID)
}
\arguments{
\item{df}{NPX data frame in long format with at least protein name (Assay), OlinkID, UniProt, 1-2 variables with at least 2 levels and subject ID.}

\item{var1}{Character value indicating the name of the column to be used as the first variable in the LME model}

\item{var2}{Character value indicating the name of the column to be used as the second variable in the LME model}

\item{ID}{Character value indicating the column of subject ID's. To be used as a random effect in the LME model.}
}
\value{
A data frame containing the results of fitting the linear mixed effects model to every protein by OlinkID.
}
\description{
Fits a linear mixed effects model for every protein (by OlinkID) using lmer. Model being fit will be NPX ~ var1 + var2 + var1:var2 + (1|ID)
}
\examples{
\donttest{p_results <- olink_lmer_main(df, var1 = "Visit", var2 = "Treatment", ID = "SubjectID")}
}
