% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_outlier_detection_utils.R
\name{olink_median}
\alias{olink_median}
\title{Compute median of quantified value}
\usage{
olink_median(df, quant_col, median_group)
}
\arguments{
\item{df}{Olink dataset}

\item{quant_col}{Character vector of name of quantification column}

\item{median_group}{Grouping for which to compute median for}
}
\value{
Input dataset with one additional columns, median
}
\description{
Compute median of quantified value
}
