% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_outlier_detection_utils.R
\name{olink_median_iqr_outlier}
\alias{olink_median_iqr_outlier}
\title{Compute outliers based on median +/- iqr_sd * IQR}
\usage{
olink_median_iqr_outlier(df, quant_col, group, iqr_sd)
}
\arguments{
\item{df}{Olink dataset}

\item{quant_col}{Character vector of name of quantification column}

\item{group}{Grouping for which to compute median for}

\item{iqr_sd}{Fixed value to multiply IQR with}
}
\value{
Boolean vector with length equal to the number of input rows
indicating outlier.
}
\description{
Compute outliers based on median +/- iqr_sd * IQR
}
