% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_non_parametric_one.R
\name{olink_non_parametric_one}
\alias{olink_non_parametric_one}
\title{Function which performs a Kruskal-Wallis Test or Friedman Test per protein
Performs an Kruskal-Wallis Test for each assay (by OlinkID) in every panel using kruskal.test.
Performs an Friedman Test for each assay (by OlinkID) in every panel using friedman_test.
The function handles factor variable. \cr\cr
Samples that have no variable information or missing factor levels are automatically removed from the analysis (specified in a messsage if verbose = T).
Character columns in the input dataframe are automatically converted to factors (specified in a message if verbose = T).
Numerical variables are not converted to factors.
If a numerical variable is to be used as a factor, this conversion needs to be done on the dataframe before the function call. \cr\cr
Inference is specified in a message if verbose = T. \cr
The formula notation of the final model is specified in a message if verbose = T. \cr\cr
Adjusted p-values are calculated by stats::p.adjust according to the Benjamini & Hochberg (1995) method (“fdr”).
The threshold is determined by logic evaluation of Adjusted_pval < 0.05.}
\usage{
olink_non_parametric_one(
  df,
  variable,
  outcome = "NPX",
  p_adjust_method = "BH",
  dependence = FALSE,
  verbose = T
)
}
\arguments{
\item{df}{NPX data frame in long format with at least protein name (Assay), OlinkID, UniProt, Panel and a factor with at least 3 levels.}

\item{variable}{Single character value.}

\item{outcome}{Character. The dependent variable. Default: NPX.}

\item{p_adjust_method}{Adjust P-value for Multiple Comparisons.}

\item{dependence}{Boolean. Default: FALSE. When the groups are independent, the kruskal-Wallis will run, when the groups are dependent, the Friedman test will run.}

\item{verbose}{Boolean. Default: True. If information about removed samples, factor conversion and final model formula is to be printed to the console.}
}
\value{
A tibble containing the Kruskal-Wallis Test results for every protein.
The tibble is arranged by ascending p-values.
}
\description{
Function which performs a Kruskal-Wallis Test or Friedman Test per protein
Performs an Kruskal-Wallis Test for each assay (by OlinkID) in every panel using kruskal.test.
Performs an Friedman Test for each assay (by OlinkID) in every panel using friedman_test.
The function handles factor variable. \cr\cr
Samples that have no variable information or missing factor levels are automatically removed from the analysis (specified in a messsage if verbose = T).
Character columns in the input dataframe are automatically converted to factors (specified in a message if verbose = T).
Numerical variables are not converted to factors.
If a numerical variable is to be used as a factor, this conversion needs to be done on the dataframe before the function call. \cr\cr
Inference is specified in a message if verbose = T. \cr
The formula notation of the final model is specified in a message if verbose = T. \cr\cr
Adjusted p-values are calculated by stats::p.adjust according to the Benjamini & Hochberg (1995) method (“fdr”).
The threshold is determined by logic evaluation of Adjusted_pval < 0.05.
}
