% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_non_parametric_one.R
\name{olink_non_parametric_one_posthoc}
\alias{olink_non_parametric_one_posthoc}
\title{Function which performs an posthoc test per protein.}
\usage{
olink_non_parametric_one_posthoc(
  df,
  olinkid_list = NULL,
  variable,
  outcome = "NPX",
  comparisons = NULL,
  p_adjust_method = "BH",
  verbose = T
)
}
\arguments{
\item{df}{NPX data frame in long format with at least protein name (Assay), OlinkID, UniProt, Panel and a factor with at least 3 levels.}

\item{olinkid_list}{Character vector of OlinkID's on which to perform post hoc analysis. If not specified, all assays in df are used.}

\item{variable}{Single character value or character array.}

\item{outcome}{Character. The dependent variable. Default: NPX.}

\item{comparisons}{A list of length-2 vectors specifying the groups of interest to be compared. For example to compare groups "A" vs "B" and "B" vs "C", the argument is as follow: comparisons = list(c("A", "B"), c("B", "C")).}

\item{p_adjust_method}{Adjust P-value for Multiple Comparisons.}

\item{verbose}{Boolean. Deafult: True. If information about removed samples, factor conversion and final model formula is to be printed to the console.}
}
\value{
Tibble of posthoc tests for specicified effect, arranged by ascending adjusted p-values.
}
\description{
Performs a post hoc test using emmeans::emmeans with Tukey p-value adjustment per assay (by OlinkID) for each panel at confidence level 0.95.
See \code{olink_kruskal} for details of input notation. \cr\cr
The function handles both factor and numerical variables.
The posthoc test for a numerical variable compares the difference in means of the outcome variable (default: NPX) for 1 standard deviation difference in the numerical variable, e.g.
mean NPX at mean(numerical variable) versus mean NPX at mean(numerical variable) + 1*SD(numerical variable).
}
\examples{
\donttest{
test_results <- olink_olink_non_parametric_one(df, "Group")
significant_assays <- test_results \%>\%
filter(Threshold == 'Significant') \%>\%
pull(OlinkID)

test_posthoc_results <- olink_non_parametric_one_posthoc(df,
variable = c("A"),
olinkid_list = significant_assays)}
}
