% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_utils.R
\name{olink_norm_format_get_nonoverlapping}
\alias{olink_norm_format_get_nonoverlapping}
\title{Retrieve non-overlapping assays between two NPX datasets. For use in
olink_normalization_format function. Generates a message stating how many
assays were not overlapping. Appends additional columns depending on the
normalization type to match normalized data output. For cross-product
normalization, splits any concatenated OlinkIDs.}
\usage{
olink_norm_format_get_nonoverlapping(
  lst_check = lst_check,
  df1 = df1,
  df1_project_nr = df1_project_nr,
  df2 = df2,
  df2_project_nr = df2_project_nr
)
}
\arguments{
\item{lst_check}{Normalization input list checks generated by
olink_norm_input_check()}

\item{df1}{First dataset to be used for normalization, pre-normalization.
Must match df1 used in olink_normalization product bridging.}

\item{df1_project_nr}{Project name of first dataset. Must match name used in
olink_normalization product bridging.}

\item{df2}{Second dataset to be used for normalization, pre-normalization.
Must match df2 used in olink_normalization product bridging.}

\item{df2_project_nr}{Project name of second dataset. Must match name used in
olink_normalization product bridging.}
}
\value{
A combined "tibble" of Olink data in long format containing only the
non-overlapping assays from each input dataset.
}
\description{
Retrieve non-overlapping assays between two NPX datasets. For use in
olink_normalization_format function. Generates a message stating how many
assays were not overlapping. Appends additional columns depending on the
normalization type to match normalized data output. For cross-product
normalization, splits any concatenated OlinkIDs.
}
\author{
Danai Topouza
}
