% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_utils.R
\name{olink_norm_format_rm_ext_ctrl}
\alias{olink_norm_format_rm_ext_ctrl}
\title{Remove negative controls and plate controls from dataset. For use in
olink_normalization_format function. Generates a message stating which
control samples were removed.}
\usage{
olink_norm_format_rm_ext_ctrl(df = df)
}
\arguments{
\item{df}{NPX dataset to be processed.}
}
\value{
A "tibble" of Olink data in long format containing the input dataset
with negative controls and plate controls removed.
}
\description{
Remove negative controls and plate controls from dataset. For use in
olink_normalization_format function. Generates a message stating which
control samples were removed.
}
\author{
Danai Topouxa
}
