% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_utils.R
\name{olink_norm_input_check_quant}
\alias{olink_norm_input_check_quant}
\title{Check quantification columns.}
\usage{
olink_norm_input_check_quant(quant_cols, quant_cols_set)
}
\arguments{
\item{quant_cols}{Named list of vector arrays with quantifications of the
datasets to be normalized.}

\item{quant_cols_set}{pre-ordered vector array of accepted quantification
column names.}
}
\value{
\var{quant_cols} with the selected quantification column.
}
\description{
This function is called from \var{olink_norm_input_check_df_cols} to resolve
ties and cases of multiple the quantification columns across datasets.
}
