% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_utils.R
\name{olink_norm_input_clean_assays}
\alias{olink_norm_input_clean_assays}
\title{Check \var{datasets} and \var{reference_medians} for unexpected Olink
identifiers or excluded assays}
\usage{
olink_norm_input_clean_assays(lst_df, reference_medians, lst_cols, norm_mode)
}
\arguments{
\item{lst_df}{Named list of datasets to be normalized.}

\item{reference_medians}{Dataset with columns "OlinkID" and "Reference_NPX".
Used for reference median normalization.}

\item{lst_cols}{Named list of vectors with the required column names for each
dataset in \var{lst_df}.}
}
\value{
A named list containing \var{lst_df} and \var{reference_medians}
stripped from unexpected Olink identifiers or excluded assays
}
\description{
Check \var{datasets} and \var{reference_medians} for unexpected Olink
identifiers or excluded assays
}
\author{
Klev Diamanti
}
