% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_utils.R
\name{olink_norm_input_norm_method}
\alias{olink_norm_input_norm_method}
\title{Check \var{datasets} and \var{reference_medians} for Olink identifiers not
shared across datasets.}
\usage{
olink_norm_input_norm_method(lst_df, lst_cols)
}
\arguments{
\item{lst_df}{Named list of datasets to be normalized.}

\item{lst_cols}{Named list of vectors with the required column names for each
dataset in \var{lst_df}.}
}
\value{
\code{NULL} if all assays are normalized with the same approach.
}
\description{
Check \var{datasets} and \var{reference_medians} for Olink identifiers not
shared across datasets.
}
\author{
Klev Diamanti;
Kathleen Nevola
}
