% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_utils.R
\name{olink_norm_input_ref_medians}
\alias{olink_norm_input_ref_medians}
\title{Check datasets of \var{reference_medians}}
\usage{
olink_norm_input_ref_medians(reference_medians)
}
\arguments{
\item{reference_medians}{Dataset with columns "OlinkID" and "Reference_NPX".
Used for reference median normalization.}
}
\value{
\code{NULL} otherwise error.
}
\description{
Check datasets of \var{reference_medians}
}
\author{
Klev Diamanti
}
