% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_utils.R
\name{olink_norm_reference_id}
\alias{olink_norm_reference_id}
\title{Identify reference project.}
\usage{
olink_norm_reference_id(lst_product, reference_project)
}
\arguments{
\item{lst_product}{Named character vector with the Olink product name that
each input dataset matches to.}

\item{reference_project}{Project name of reference_project. Should be one of
\var{df1_project_nr} or \var{df2_project_nr}. Indicates the project to which
the other project is adjusted to.}
}
\value{
Named character vector with \var{df1_project_nr} and
\var{df2_project_nr} marked as "ref" and "not_ref".
}
\description{
Identify reference project.
}
\author{
Kathy Nevola
Klev Diamanti
}
