% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_normalization.R
\name{olink_normalization_intensity}
\alias{olink_normalization_intensity}
\title{Intensity normalization of all proteins}
\usage{
olink_normalization_intensity(
  df1,
  df2,
  df1_project_nr = "P1",
  df2_project_nr = "P2",
  reference_project = "P1"
)
}
\arguments{
\item{df1}{First dataframe to be used in normalization (required).}

\item{df2}{Second dataframe to be used in normalization}

\item{df1_project_nr}{Project name of first dataset.}

\item{df2_project_nr}{Project name of second dataset.}

\item{reference_project}{Project name of reference_project. Needs to be the same as either df1_project_nr or df2_project_nr. The project to which the second project is adjusted to.}
}
\value{
A tibble of NPX data in long format containing normalized NPX values, including adjustment factors.
}
\description{
Normalizes two NPX data frames using the median difference per protein. This normalization method is referred to as intensity normalization.
}
\details{
df1 is adjusted to df2 based on the difference between the medians.
Each protein is adjusted so that the median value in each dataframe is the same.
By default, df1 will be used as reference.
}
\references{
https://www.olink.com/application/data-normalization-and-standardization/
}
\keyword{Normalization}
