% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_median_norm.R
\name{olink_normalization_median_ref}
\alias{olink_normalization_median_ref}
\title{Median normalization of all proteins between two NPX projects.}
\usage{
olink_normalization_median_ref(
  exploreht_df,
  explore3072_df,
  bridge_samples,
  exploreht_name = "reference",
  explore3072_name = "new"
)
}
\arguments{
\item{exploreht_df}{Name of the Explore HT project (default: reference).}

\item{explore3072_df}{Name of the Explore 3072 project (default: new).}

\item{bridge_samples}{Named list of 2 arrays containing, SampleID of shared
samples to be used for the calculation of adjustment factor. The
names of the two arrays should be DF1 and DF2 corresponding to projects HT
and 3072, respectively. Arrays should be of equal length and index of each
entry should correspond to the same sample. (required)}
}
\value{
A "tibble" of NPX data in long format containing normalized NPX
values, including adjustment factors and name of project.
}
\description{
Normalizes two NPX projects (data frames) using shared samples.\cr\cr
}
\details{
This function is a wrapper of olink_normalization.\cr\cr

In bridging normalization one of the projects is adjusted to another using
shared samples (bridge samples). It is not necessary for the shared
samples to be named the same in each project. Adjustment between the two
projects is made using the median of the paired differences between the
shared samples. The two data frames are inputs project_1_df and project_2_df,
the one being adjusted to is specified in the input project_ref_name and the
shared samples are specified in bridge_samples.\cr\cr
}
\keyword{Normalization;}
\keyword{median}
\keyword{normalization}
