% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_n.R
\name{olink_normalization_n_check}
\alias{olink_normalization_n_check}
\title{An internal function to perform checks on the input of the function
olink_normalization_n.}
\usage{
olink_normalization_n_check(norm_schema)
}
\arguments{
\item{norm_schema}{A tibble with more than 1 rows and (strictly) the
following columns: "order", "name", "data", "samples", "normalization_type",
"normalize_to". See above for details of the structure of the data frame. See
details in help for olink_normalization_n. (required)}
}
\value{
a character message. If the message is "TRUE" then all checks passed,
otherwise an error message will be printed.
}
\description{
An internal function to perform checks on the input of the function
olink_normalization_n.
}
