% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_n.R
\name{olink_normalization_project_name_check}
\alias{olink_normalization_project_name_check}
\title{An internal function to perform checks on the input project names in the
functions olink_normalization_bridge and olink_normalization_subset. The
function is expected to run all checks on project names to make sure that
normalization can be performed smoothly. It should work independently of the
function calling it.}
\usage{
olink_normalization_project_name_check(
  project_1_name,
  project_2_name,
  project_ref_name
)
}
\arguments{
\item{project_1_name}{Name of project 1 (required)}

\item{project_2_name}{Name of project 2 (required)}

\item{project_ref_name}{Name of reference project (required)}
}
\value{
a character message. If the message is "TRUE" then all checks passed,
otherwise an error message will be printed.
}
\description{
An internal function to perform checks on the input project names in the
functions olink_normalization_bridge and olink_normalization_subset. The
function is expected to run all checks on project names to make sure that
normalization can be performed smoothly. It should work independently of the
function calling it.
}
