% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_quantile_smoothing.R
\name{olink_normalization_qs}
\alias{olink_normalization_qs}
\title{Quantile smoothing normalization of all proteins between two NPX projects.}
\usage{
olink_normalization_qs(lst_df, ref_cols, bridge_samples)
}
\arguments{
\item{lst_df}{A named list of the 2 input datasets. First element should be
the reference dataset from Olink Explore HT and the second element should
originate from Olink Explore 3072. (required)}

\item{ref_cols}{A named list with the column names to use. Exported from
olink_norm_input_check. (required)}

\item{bridge_samples}{Character vector of samples to be used for the
quantile mapping. (required)}
}
\value{
A "tibble" of Olink data in long format containing both input
datatsets with the quantile normalized quantifications.
}
\description{
This function uses bridge samples to map quantiles of the non-reference
dataset to the ones of the reference dataset. Mapped quantiles are used to
transform the quantifications of the the non-reference dataset to the
reference.
}
\details{
We need to add some details from the tutorial/education material to describr
the function in a bit more detail.
}
\author{
Amrita Kar
Marianne Sandin
Masoumeh Sheikhi
Klev Diamanti
}
\keyword{;}
\keyword{Normalization;}
\keyword{Quantile}
\keyword{Smoothing}
