% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_color.R
\name{olink_pal}
\alias{olink_pal}
\title{Olink color panel for plotting}
\usage{
olink_pal(alpha = 1, coloroption = NULL)
}
\arguments{
\item{alpha}{transparency (optional)}

\item{coloroption}{string, one or more of the following:
c("red", "orange", "yellow", "green", "teal", "turqoise", "lightblue",
"darkblue", "purple", "pink")}
}
\value{
A character vector of palette hex codes for colors.
}
\description{
Olink color panel for plotting
}
\examples{
\donttest{
if (rlang::is_installed(pkg = c("scales"))) {
  # Color matrices
  scales::show_col(
    colours = OlinkAnalyze::olink_pal()(10L),
    labels = FALSE
  )
  scales::show_col(
    colours = OlinkAnalyze::olink_pal(
      coloroption = c("lightblue", "green")
    )(2L),
    labels = FALSE
  )

  # Contour plot
  filled.contour(
    x = datasets::volcano,
    color.palette = OlinkAnalyze::olink_pal(),
    asp = 1
  )
  filled.contour(
    x = datasets::volcano,
    color.palette = scales::hue_pal(),
    asp = 1
  )
}
}

}
\keyword{Olink}
\keyword{color}
\keyword{palette}
