% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_pathway_visualization.R
\name{olink_pathway_visualization}
\alias{olink_pathway_visualization}
\title{Creates bargraph of top/selected enrichment terms from GSEA or ORA results
from \code{olink_pathway_enrichment}}
\usage{
olink_pathway_visualization(
  enrich_results,
  method = "GSEA",
  keyword = NULL,
  number_of_terms = 20L
)
}
\arguments{
\item{enrich_results}{data frame of enrichment results from
\code{olink_pathway_enrichment}}

\item{method}{method used in \code{olink_pathway_enrichment} ("GSEA" (default)
or "ORA")}

\item{keyword}{(optional) keyword to filter enrichment results on. If not
specified, displays top terms.}

\item{number_of_terms}{number of terms to display (default is 20).}
}
\value{
A bargraph as a ggplot object.
}
\description{
Pathways are ordered by increasing p-value (unadjusted)
}
\examples{
\donttest{
if (rlang::is_installed(pkg = c("msigdbr", "clusterProfiler"))) {

  # Run olink_ttest or other stats test (see documentation )
  npx_df <- npx_data1 |>
    dplyr::filter(
      !grepl(
        pattern = "control",
        x = .data[["SampleID"]],
        ignore.case = TRUE
      )
    )

  check_log <- check_npx(df = npx_df)

  ttest_results <- OlinkAnalyze::olink_ttest(
    df = npx_df,
    variable = "Treatment",
    alternative = "two.sided",
    check_log = check_log
  )

  # Run olink_pathway_enrichment (see documentation)

  # GSEA
  gsea_results <- OlinkAnalyze::olink_pathway_enrichment(
    df = npx_df,
    test_results = ttest_results,
    check_log = check_log
  )

  # ORA
  ora_results <- OlinkAnalyze::olink_pathway_enrichment(
    df = npx_df,
    test_results = ttest_results,
    check_log = check_log,
    method = "ORA"
  )

  # Plot

  OlinkAnalyze::olink_pathway_visualization(
    enrich_results = gsea_results
  )

  OlinkAnalyze::olink_pathway_visualization(
    enrich_results = gsea_results,
    keyword = "immune"
  )

  OlinkAnalyze::olink_pathway_visualization(
    enrich_results = ora_results,
    method = "ORA",
    number_of_terms = 15L
  )
}
}

}
