% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_utils.R
\name{olink_product_identifier_norm}
\alias{olink_product_identifier_norm}
\title{Identify names of product for each project}
\usage{
olink_product_identifier_norm(lst_df, reference_project, lst_cols)
}
\arguments{
\item{lst_df}{Named list of datasets to be normalized.}

\item{reference_project}{Name of reference product}

\item{lst_cols}{Named list of vectors with the required column names for each
dataset in \var{lst_df}.}
}
\value{
list of lists containing named list of product (3k, Reveal, HT,
or other) and reference (ref or not_ref)
}
\description{
Identify names of product for each project
}
\author{
Kathy Nevola
}
