% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano_plot.R
\name{olink_volcano_plot}
\alias{olink_volcano_plot}
\title{Easy volcano plot with Olink theme}
\usage{
olink_volcano_plot(p.val_tbl, x_lab = "Estimate", olinkid_list = NULL)
}
\arguments{
\item{p.val_tbl}{a data frame of results generated by olink_ttest()}

\item{x_lab}{Optional. Character value to use as the X-axis label}

\item{olinkid_list}{Optional. Character vector of proteins (by OlinkID) to label in the plot. If not provided, default is to label all significant proteins.}
}
\value{
An object of class "ggplot"
}
\description{
Creates a volcano plot using the results of the olink_ttest function. The estimated difference is plotted on the x-axis and the p-value on the y-axis. Horizontal dotted line indicates a p-value of 0.05. Significant proteins are colored based on a Benjamini-Hochberg adjusted p-value cutoff of 0.05.
}
\examples{
\donttest{ttest_results <- olink_ttest(df=npx.data, variable = 'Group', alternative = 'two.sided')
volcano_plot(ttest_results)}
}
