% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_helper.R
\name{pheatmap_annotate_heatmap}
\alias{pheatmap_annotate_heatmap}
\title{Add annotations to pheatmap arguments}
\usage{
pheatmap_annotate_heatmap(
  df,
  check_log,
  colnames,
  pheatmap_args,
  variable_row_list,
  variable_col_list
)
}
\arguments{
\item{df}{Data frame in long format with SampleID, NPX, OlinkID, Assay and
columns of choice for annotations.}

\item{check_log}{output from check_npx on \code{df}}

\item{colnames}{Character. Determines how to label the columns.
Must be 'assay', 'oid', or 'both' (default 'both').}

\item{variable_row_list}{Columns in \code{df} to be annotated for rows in
the heatmap.}

\item{variable_col_list}{Columns in \code{df} to be annotated for columns in
the heatmap.}
}
\value{
updated pheatmap arguments
}
\description{
Add annotations to pheatmap arguments
}
\keyword{internal}
