% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_helper.R
\name{pheatmap_run}
\alias{pheatmap_run}
\title{Run the pheatmap using args}
\usage{
pheatmap_run(pheatmap_args)
}
\arguments{
\item{pheatmap_args}{}
}
\value{
ggplot of heatmap as generated by pheatmap
}
\description{
Run the pheatmap using args
}
\keyword{internal}
