% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_helper.R
\name{pheatmap_set_plot_theme}
\alias{pheatmap_set_plot_theme}
\title{Add theme to pheatmap}
\usage{
pheatmap_set_plot_theme(
  x,
  fontsize,
  col = "#737373",
  font1 = "Arial Regular",
  font2 = "Arial"
)
}
\arguments{
\item{x}{pheatmap plot}

\item{fontsize}{size of font}

\item{col}{color}

\item{font1}{font to use as default}

\item{font2}{secondary font to use}
}
\value{
pheatmap with updated theme
}
\description{
Add theme to pheatmap
}
\keyword{internal}
