% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_helper.R
\name{plot_heatmap_check_inputs}
\alias{plot_heatmap_check_inputs}
\title{Checking for needed packages and valid inputs}
\usage{
plot_heatmap_check_inputs(colnames, ...)
}
\arguments{
\item{colnames}{Character. Determines how to label the columns.
Must be 'assay', 'oid', or 'both' (default 'both').}

\item{...}{Additional arguments used in \code{pheatmap::pheatmap}}
}
\value{
Null or error/warnings
}
\description{
Checking for needed packages and valid inputs
}
\keyword{internal}
