% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_helper.R
\name{plot_heatmap_clean_df}
\alias{plot_heatmap_clean_df}
\title{remove low var assays and add colnames}
\usage{
plot_heatmap_clean_df(df, check_log, colnames)
}
\arguments{
\item{df}{Data frame in long format with SampleID, NPX, OlinkID, Assay and
columns of choice for annotations.}

\item{check_log}{output from check_npx on \code{df}}

\item{colnames}{Character. Determines how to label the columns.
Must be 'assay', 'oid', or 'both' (default 'both').}
}
\value{
df w/o low var assays with added colnames
}
\description{
remove low var assays and add colnames
}
\keyword{internal}
