% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised.R
\name{plot_rf_interactions}
\alias{plot_rf_interactions}
\title{plot random forest interactions}
\usage{
plot_rf_interactions(
  long,
  rf,
  vars,
  variable = "grouped_severity",
  num_interactions = 15,
  make_the_plot = TRUE
)
}
\description{
plot random forest interactions
}
