% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_capture.R
\name{print_and_capture}
\alias{print_and_capture}
\title{Capture the output of printing an object}
\usage{
print_and_capture(x)
}
\arguments{
\item{x}{printable object}
}
\value{
string representation of the provided object
}
\description{
Capture the output of printing an object
}
\examples{

OlinkAnalyze:::print_and_capture(npx_data1)

}
