% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_plate_randomizer.R
\name{product_to_plate_size}
\alias{product_to_plate_size}
\title{Check product name and set plate size accordingly}
\usage{
product_to_plate_size(product)
}
\arguments{
\item{product}{(String) Name of product (needs to match one of names in
accepted_olink_platforms$name)}
}
\value{
(Integer) Corresponding plate size per
accepted_olink_platforms$plate_size
}
\description{
If plate size is not provided, function will use
accepted_olink_products tibble to map the product name to the plate size
}
\keyword{internal}
