% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read NPX data.R
\name{read_multitab_NPX}
\alias{read_multitab_NPX}
\title{Multitab NPX data parsing}
\usage{
read_multitab_NPX(filename, sample_manifest = NULL, pheno = NULL, num_tabs = 1)
}
\description{
Multitab NPX data parsing
}
