% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flex.R
\name{read_npflex}
\alias{read_npflex}
\title{Read in flex data}
\usage{
read_npflex(filename)
}
\arguments{
\item{filename}{where the file is located}
}
\value{
tibble of data
}
\description{
Called by read_NPX
}
