% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_delim.R
\name{read_npx_delim}
\alias{read_npx_delim}
\alias{read_npx_csv}
\title{Help function to read NPX, Ct or absolute quantification data from semicolon-
or comma-delimited Olink software output files in R.}
\usage{
read_npx_delim(file, out_df = "arrow", sep = NULL)

read_npx_csv(file, out_df = "arrow", sep = NULL)
}
\arguments{
\item{file}{Path to Olink software output delimited file in wide or long
format. Expecting file extensions `csv` or `txt`.}

\item{out_df}{The class of output data frame. One of `tibble` (default) or
`arrow` for ArrowObject.}

\item{sep}{Character separator of delimited input file. One of `NULL` for
auto-detection (default), `,` for comma or `;` for semicolon. Used only for
delimited output files from Olink software.}
}
\value{
Tibble or ArrowObject with Olink data in long format.
}
\description{
The function can handle delimited files in long and wide format.
}
\seealso{
[read_npx()]
  [read_npx_parquet()]
  [read_npx_zip()]
  [read_npx_excel()]
  [read_npx_delim_wide()]
  [read_npx_delim_long()]
  [read_npx_format()]
}
\author{
Klev Diamanti;
  Christoffer Cambronero;
  Kathleen Nevola
}
