% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_delim.R
\name{read_npx_delim_long}
\alias{read_npx_delim_long}
\title{Help function to read in Olink data in long format from comma or semicolon
delimited files.}
\usage{
read_npx_delim_long(file, sep = NULL)
}
\arguments{
\item{file}{Path to Olink software output in txt or csv.}

\item{sep}{The separator of the file: NULL (autodetect), comma (,) or
semicolon (;).}
}
\value{
An R6 class ArrowObject or tibble.
}
\description{
Help function to read in Olink data in long format from comma or semicolon
delimited files.
}
\author{
Klev Diamanti;
  Christoffer Cambronero;
  Kathleen Nevola
}
