% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_delim.R
\name{read_npx_delim_wide}
\alias{read_npx_delim_wide}
\title{Help function to read NPX, Ct or absolute quantification data from semicolon-
or comma-delimited Olink software output files in R.}
\usage{
read_npx_delim_wide(file, sep = NULL)
}
\arguments{
\item{file}{Path to Olink software output delimited file in wide format.
Expecting file extensions `csv` or `txt`.}

\item{sep}{Character separator of delimited input file. One of `NULL` for
auto-detection (default), `,` for comma or `;` for semicolon. Used only for
delimited output files from Olink software.}
}
\value{
An R6 class ArrowObject or tibble.
}
\description{
The function can handle delimited files in wide format.
}
\seealso{
\code{\link{read_npx_delim_long}}
  \code{\link{read_npx_delim}}
}
\author{
Klev Diamanti
}
