% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_excel.R
\name{read_npx_excel_format}
\alias{read_npx_excel_format}
\title{Help function to determine the wide/long excel file format.}
\usage{
read_npx_excel_format(file, long_format = NULL, quant_methods_excel)
}
\arguments{
\item{file}{The input excel file.}

\item{quant_methods_excel}{Character vector with the Olink protein
quantification methods.}

\item{in_long_format}{Boolean marking if input file is in long (TRUE) or wide
(FALSE) format.}
}
\value{
A list of two variables: is_long_format and data_cells The first one
is a boolean noting the file format and the second one is the character
vector from the input file where the format was determined from.
An error is returned if the format cannot be determined, unless the user has
defined the file format from the input.
}
\description{
Help function to determine the wide/long excel file format.
}
