% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_excel.R
\name{read_npx_excel_platform}
\alias{read_npx_excel_platform}
\title{Help function to determine the Olink platform from the excel file.}
\usage{
read_npx_excel_platform(
  file,
  olink_platform = NULL,
  is_long_format,
  olink_platforms_excel
)
}
\arguments{
\item{file}{The input excel file.}

\item{is_long_format}{Boolean marking if input file is in long (TRUE) or wide
(FALSE) format.}

\item{olink_platforms_excel}{Data frame from the global variable
accepted_olink_platforms.}

\item{in_olink_platform}{The Olink platform used to generate the input file.
Expecting "Target 96", "Target 48", "Flex" or "Focus".}
}
\value{
The name of the Olink platform detected automatically or set by the
user.
}
\description{
Help function to determine the Olink platform from the excel file.
}
