% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_excel.R
\name{read_npx_excel_quant}
\alias{read_npx_excel_quant}
\title{Help function to determine the the type of the quantification of the data in
the excel file.}
\usage{
read_npx_excel_quant(
  file,
  data_type = NULL,
  data_cells,
  quant_methods_expected
)
}
\arguments{
\item{file}{The input excel file.}

\item{data_type}{The quantification in which the data comes in. Expecting one
of NPX, Quantified or Ct.}

\item{data_cells}{A character vector with the contents of the cells
indicating the quantification method.}

\item{quant_methods_expected}{Character vector with the Olink protein
quantification methods.}
}
\value{
The quantification method detected in the file.
An error is returned if the format cannot be determined.
}
\description{
Help function to determine the the type of the quantification of the data in
the excel file.
}
\author{
Klev Diamanti
}
