% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_utils.R
\name{read_npx_format_colnames}
\alias{read_npx_format_colnames}
\title{Help function checking whether a dataset contains NA or empty strings on
its column names}
\usage{
read_npx_format_colnames(df, file)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{file}{Path to Olink software output file in wide or long format.
Expecting extensions "xls" or "xlsx" for excel files, "csv" or "txt" for delim files, "parquet" for parquet files, and "zip" for compressed files.}
}
\value{
Error is file contains problematic column names. \code{NULL} otherwise.
}
\description{
Help function checking whether a dataset contains NA or empty strings on
its column names
}
\author{
Klev Diamanti
}
