% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_format.R
\name{read_npx_format_get_format}
\alias{read_npx_format_get_format}
\title{Help function to determine the format (wide or long) of the input dataset.}
\usage{
read_npx_format_get_format(df_top_n, file, long_format = NULL)
}
\arguments{
\item{df_top_n}{A tibble containing the first \var{read_n} rows of the input
dataset.}

\item{file}{Path to Olink software output file in wide or long format.
Expecting file extensions
"xls", "xlsx", "csv", or "txt".}

\item{long_format}{Boolean marking format of input file. One of \code{TRUE} for
long format and \code{FALSE} for wide format files. Defaults to \code{NULL} for
auto-detection.}
}
\value{
A list with two elements:
\itemize{
\item A scalar boolean (\var{is_long_format}) marking if the input file is in
long (\code{TRUE}) or wide (\code{FALSE}) format.
\item A character vector (\var{data_cells}) from the input file which allows
detection of the quantification method. Used in function
\code{\link{read_npx_format_get_quant}}.
}
}
\description{
The function uses the first \var{read_n} rows of the input dataset to
determine the format of the input file.

The user can provide the file format as an input argument \var{long_format}.
If the user did not provide an input with regards to the file format, then it
is auto-determined is this function. If file format was provided as input,
then the function cross-checks that its auto-detection matches the user
input. If not, it will throw a warning, and accept user's input as the
correct answer.
}
\seealso{
\code{\link{read_npx_format}}
\code{\link{read_npx_format_read}}
\code{\link{read_npx_format_get_platform}}
\code{\link{read_npx_format_get_quant}}
}
\author{
Klev Diamanti
}
