% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_format.R
\name{read_npx_format_get_platform}
\alias{read_npx_format_get_platform}
\title{Help function to determine Olink platform from the input file in wide format.}
\usage{
read_npx_format_get_platform(
  df,
  file,
  olink_platform = NULL,
  olink_platforms_wide
)
}
\arguments{
\item{df}{A tibble containing the first \var{read_n} rows of the input Olink
file.}

\item{file}{Path to Olink software output file in wide or long format.
Expecting file extensions `csv`, `txt`, `xls`, or `xlsx`.}

\item{olink_platform}{Olink platform used to generate the input file.
One of `NULL` (default), `Target 96`, `Target 48`, `Flex` or `Focus`.}

\item{olink_platforms_wide}{Tibble with rows corresponding to qPCR platforms
extracted from \var{accepted_olink_platforms} from the local environment.}
}
\value{
The name of the Olink platform. One of `Target 96`, `Target 48`,
`Flex` or `Focus`.
}
\description{
This function uses the panel name from Olink software files in wide format to
determine the qPCR platform that was used for the project that this dataset
represents.
}
\seealso{
\code{\link{read_npx_format}}
  \code{\link{read_npx_format_read}}
  \code{\link{read_npx_format_get_format}}
  \code{\link{read_npx_format_get_quant}}
}
\author{
Klev Diamanti
}
