% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_format.R
\name{read_npx_format_read}
\alias{read_npx_format_read}
\title{Help function that reads the first two rows of an Olink delimited or excel
file.}
\usage{
read_npx_format_read(file, sep = NULL, read_n = 3L)
}
\arguments{
\item{file}{Path to Olink software output in excel, txt or csv.}

\item{sep}{The separator of the file: NULL (autodetect), comma (,) or
semicolon (;). Ignored if file is in excel format.}

\item{read_n}{Number of top rows to read.}
}
\value{
A list with one tibble (df_top_n) and one ArrowObject (df). The
ArrowObject contains the full dataset and the tibble the 3 rows that will
allow us to determine the characteristics of the input file.
}
\description{
This function extracts the first two rows of an Olink output file to detect
the Olink platform that generated the file, the data type and and file
format. The input file format may be wide or long, and the input file type
can be excel (xls/xlsx) or delimited (csv/txt).

In case of long format input, the function will return the row of the column
names and the first row with data.

In case of wide format inpiut, the function will return the first two rows of
the inpit file.
}
\author{
Klev Diamanti
}
