% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_format.R
\name{read_npx_get_format}
\alias{read_npx_get_format}
\title{Help function to determine the wide or long file format.}
\usage{
read_npx_get_format(df, file, long_format = NULL, quant_methods)
}
\arguments{
\item{df}{Tibble containing the first three rows of the input Olink file.}

\item{file}{The input Olink file.}

\item{long_format}{Boolean marking if input file is in long (TRUE) or wide
(FALSE) format.}

\item{quant_methods}{Character vector with the Olink protein
quantification methods.}
}
\value{
A list of two variables: is_long_format and data_cells The first one
is a boolean noting the file format and the second one is the character
vector from the input file where the format was determined from.
An error is returned if the format cannot be determined, unless the user has
defined the file format from the input.
}
\description{
Help function to determine the wide or long file format.
}
\author{
Klev Diamanti
}
