% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_format.R
\name{read_npx_get_platform}
\alias{read_npx_get_platform}
\title{Help function to determine the Olink platform from the input Olink wide file.}
\usage{
read_npx_get_platform(df, file, olink_platform = NULL, olink_platforms_wide)
}
\arguments{
\item{df}{Tibble containing the first three rows of the input Olink file in
wide format.}

\item{file}{The input excel file.}

\item{olink_platform}{The Olink platform used to generate the input file in
wide format. Expecting "Target 96", "Target 48", "Flex" or "Focus".}

\item{olink_platforms_wide}{Data frame from the global variable
accepted_olink_platforms.}
}
\value{
The name of the Olink platform detected automatically or set by the
user.
}
\description{
Help function to determine the Olink platform from the input Olink wide file.
}
\author{
Klev Diamanti
}
