% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide_utils.R
\name{read_npx_wide_assay_nrow}
\alias{read_npx_wide_assay_nrow}
\title{Help function to determine the number of rows with assay information in Olink
excel long or wide format files.}
\usage{
read_npx_wide_assay_nrow(file, olink_platform = NULL, data_type = NULL)
}
\arguments{
\item{file}{The input excel file.}

\item{olink_platform}{The Olink platform used to generate the input file.
Expecting "Target 96", "Target 48", "Flex" or "Focus".}

\item{data_type}{The quantification in which the data comes in. Expecting one
of NPX, Quantified or Ct.}
}
\value{
Integer number of rows containing info about assays in Olink excel
wide files.
}
\description{
Help function to determine the number of rows with assay information in Olink
excel long or wide format files.
}
\author{
Klev Diamanti
}
