% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_bottom}
\alias{read_npx_wide_bottom}
\title{Help function that converts df_bottom into long format.}
\usage{
read_npx_wide_bottom(
  df,
  file,
  olink_platform,
  data_type,
  col_names,
  format_spec,
  df_plate_panel
)
}
\arguments{
\item{df}{The bottom data frame from a split Olink wide excel file.}

\item{file}{The input excel file.}

\item{olink_platform}{The Olink platform used to generate the input file.
Expecting "Target 96", "Target 48", "Flex" or "Focus".}

\item{data_type}{The quantification in which the data comes in. Expecting one
of NPX, Quantified or Ct.}

\item{col_names}{Names list of character vector with the names of the columns
containing Olink data on assays and internal control assays.}

\item{format_spec}{A one-row data frame filtered from olink_wide_excel_spec
with the Olink wide excel file specifications.}

\item{df_plate_panel}{Data frame with unique combinations of panels and
plates from the combination of top and middle data frames.}
}
\value{
The bottom matrix in long format.
}
\description{
Help function that converts df_bottom into long format.
}
\author{
Klev Diamanti
}
