% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_bottom_t}
\alias{read_npx_wide_bottom_t}
\title{Help function that converts df_bottom into long format.}
\usage{
read_npx_wide_bottom_t(df, file, data_type, col_split)
}
\arguments{
\item{df}{The bottom data frame from a split Olink wide excel file.}

\item{file}{The input excel file.}

\item{data_type}{The quantification in which the data comes in. Expecting one
of NPX or Quantified.}

\item{col_split}{The column the splits the Olink wide excel file into left
and right.}
}
\value{
The bottom matrix in long format.
}
\description{
Help function that converts df_bottom into long format.
}
