% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_check_middle}
\alias{read_npx_wide_check_middle}
\title{Help function to check df_mid for potential inconsistencies.}
\usage{
read_npx_wide_check_middle(
  df,
  file,
  data_type,
  assay_cols,
  plate_cols,
  qc_warn_cols,
  int_ctrl_cols
)
}
\arguments{
\item{df}{The middle data frame from a split Olink wide excel file.}

\item{file}{The input excel file.}

\item{data_type}{The quantification in which the data comes in. Expecting one
of NPX or Quantified.}

\item{assay_cols}{Character vector with the names of the columns containing
Olink assays.}

\item{plate_cols}{Character vector with the names of the columns containing
"Plate ID".}

\item{qc_warn_cols}{Character vector with the names of the columns containing
"QC Warning".}

\item{int_ctrl_cols}{Character vector with the names of the columns
containing "Internal Controls".}
}
\value{
Nothing or an error if inconsistencies are detected.
}
\description{
Help function to check df_mid for potential inconsistencies.
}
\author{
Klev Diamanti
}
