% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_check_top}
\alias{read_npx_wide_check_top}
\title{Additional checks of the top matrix of Olink dataset in wide format.}
\usage{
read_npx_wide_check_top(df, file, format_spec)
}
\arguments{
\item{df}{Top matrix of Olink datasets in wide format \var{df_top}.}

\item{file}{Path to Olink software output file in wide or long format.
Expecting file extensions
xls, xlsx, csv, or txt. # nolint}

\item{format_spec}{A tibble derived from \var{olink_wide_spec} in the local
environment containing the expected format of the Olink wide file based on
the \var{olink_platform} and \var{data_type}.}
}
\value{
NULL unless an inconsistency is spotted.
}
\description{
Additional checks of the top matrix of Olink dataset in wide format.
}
\seealso{
\code{\link{read_npx_wide_top}}
}
\author{
Klev Diamanti
}
