% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_panel_version}
\alias{read_npx_wide_panel_version}
\title{Help function to extract Panel_Version from Panel column.}
\usage{
read_npx_wide_panel_version(df)
}
\arguments{
\item{df}{A tibble containing the column `Panel`.}
}
\value{
Same tibble as input with additional column Panel_Version.
}
\description{
Help function to extract Panel_Version from Panel column.
}
