% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_split_col}
\alias{read_npx_wide_split_col}
\title{Help function to determine the number of columns with assay measurements in
Olink excel wide format files.}
\usage{
read_npx_wide_split_col(df, file)
}
\arguments{
\item{df}{The top data frame from a split Olink wide excel file.}

\item{file}{The input excel file.}
}
\value{
The index at which the columns of the Olink wide file are split into
the assay measurements and other metadata.
}
\description{
Help function to determine the number of columns with assay measurements in
Olink excel wide format files.
}
\author{
Klev Diamanti
}
