% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_zip.R
\name{read_npx_zip}
\alias{read_npx_zip}
\title{Helper function to read NPX zip-compressed files with checksum.}
\usage{
read_npx_zip(file, .ignore_files = c("README.txt"))
}
\arguments{
\item{file}{Path to Olink Software output zip file.}

\item{.ignore_files}{Vector of files to ignore.}
}
\value{
A "tibble" in long format.
}
\description{
Helper function to read NPX zip-compressed files with checksum.
}
