% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_all_na_cols}
\alias{remove_all_na_cols}
\title{Utility function removing columns with all values NA from a dataset.}
\usage{
remove_all_na_cols(df)
}
\arguments{
\item{df}{An Olink dataset.}
}
\value{
The input Olink dataset without all-NA columns.
}
\description{
Utility function removing columns with all values NA from a dataset.
}
\author{
Klev Diamanti
}
