% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_go}
\alias{run_go}
\title{run and plot GO analysis using topGO}
\usage{
run_go(
  pvals,
  go_annot_list,
  p_label = 2.7,
  size_label = c(0.9, 1.5),
  plt_title = NULL,
  return_res = FALSE,
  test_stat = "ks"
)
}
\description{
run and plot GO analysis using topGO
}
