% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_helper.R
\name{run_pheatmap}
\alias{run_pheatmap}
\title{Run the pheatmap using args}
\usage{
run_pheatmap(pheatmap_args)
}
\arguments{
\item{pheatmap_args}{}
}
\value{
ggplot of heatmap as generated by pheatmap
}
\description{
Run the pheatmap using args
}
\keyword{internal}
