% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_theme.R
\name{set_plot_theme}
\alias{set_plot_theme}
\title{Function to set plot theme}
\usage{
set_plot_theme(font = "Arial")
}
\arguments{
\item{font}{Font family to use for text elements. Default: "Arial".}
}
\value{
No return value, used as theme for ggplots
}
\description{
This function sets a coherent plot theme for functions.
}
\examples{
\donttest{
ggplot2::ggplot(
  data = datasets::mtcars,
  mapping = ggplot2::aes(
    x = .data[["wt"]],
    y = .data[["mpg"]],
    color = as.factor(x = .data[["cyl"]])
  )
) +
  ggplot2::geom_point(
    size = 4L
  ) +
  OlinkAnalyze::set_plot_theme()

ggplot2::ggplot(
  data = datasets::mtcars,
  mapping = ggplot2::aes(
    x = .data[["wt"]],
    y = .data[["mpg"]],
    color = as.factor(x = .data[["cyl"]])
  )
) +
  ggplot2::geom_point(
    size = 4L
  ) +
  OlinkAnalyze::set_plot_theme(
    font = ""
  )
}
}
