% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_helper.R
\name{set_plot_theme_pheatmap}
\alias{set_plot_theme_pheatmap}
\title{Title}
\usage{
set_plot_theme_pheatmap(
  x,
  fontsize,
  col = "#737373",
  font1 = "Arial Regular",
  font2 = "Arial"
)
}
\arguments{
\item{x}{}

\item{fontsize}{}

\item{col}{}

\item{font1}{}

\item{font2}{}
}
\description{
Title
}
\keyword{internal}
