% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{simple_lm_de}
\alias{simple_lm_de}
\title{simple standard differential expression analysis plot}
\usage{
simple_lm_de(
  long,
  case_control_to_remove = NA,
  variable = "case.control",
  covariates = NULL,
  logp_label = 6,
  fc_label = c(-0.5, 1),
  plot_xlab = "Log2 Fold Change (P-N)",
  plot_title = "plasma case/control"
)
}
\description{
simple standard differential expression analysis plot
}
