% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_npx.R
\name{tbl_sum.olink_npx}
\alias{tbl_sum.olink_npx}
\title{Provide a custom header for \code{olink_npx} objects}
\usage{
\method{tbl_sum}{olink_npx}(x, ...)
}
\arguments{
\item{x}{An \code{olink_npx} object.}

\item{...}{Additional arguments passed to the tibble method.}
}
\value{
A named character vector of summary items.
}
\description{
Provide a custom header for \code{olink_npx} objects
}
\keyword{internal}
