% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read NPX data.R
\name{uniprot_to_gene}
\alias{uniprot_to_gene}
\title{Map protein IDs to gene IDs.}
\usage{
uniprot_to_gene(df)
}
\arguments{
\item{df}{meta_dat dataframe}

\item{tab}{(optional) tab number}
}
\value{
Gene IDs
}
\description{
Map protein IDs to gene IDs.
}
