% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeable.R
\name{bridgeable}
\alias{bridgeable}
\title{Function to find bridgeable assays from 3K to HT}
\usage{
bridgeable(data_Explore384, data_HT)
}
\arguments{
\item{data_Explore384}{A Explore384 NPX file}

\item{data_HT}{An Explore HT NPX file}
}
\value{
A "tibble" in long format. Columns include:
\itemize{
\item{OlinkIDconcat:} Concatenated Olink Explore HT and Explore 3072 IDs
\item{OlinkID_HT:} Olink ID HT
\item{OlinkID_E3072:} Olink ID Explore 3072
\item{Assay:} Assay HT
\item{Bridgeable:} Bridging Flag
}
Additional columns may be present or missing depending on the platform
}
\description{
Imports an Explore 384 and a Explore HT NPX file exported from Olink Software.
No alterations to the output format is allowed.
}
\examples{
\donttest{
results <- OlinkAnalyze:::bridgeable(data_Explore384 = OlinkAnalyze:::data_3k,
                                     data_HT = OlinkAnalyze:::data_ht)
}

}
\author{
Amrita Kar

Marianne Sandin

Danai G. Topouza
}
\keyword{Bridging}
\keyword{NPX}
