% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all_na_assays.R
\name{check_all_na_assays}
\alias{check_all_na_assays}
\title{Help function to identify Olink assays with all quantified values NA.}
\usage{
check_all_na_assays(df, col_names)
}
\arguments{
\item{df}{A tibble or an arrow object containing columns "OlinkID" and
either "NPX", "Quantified_value" or "Ct"}

\item{col_names}{A list of matched column names, the output of `check_npx_col_names` function.}
}
\value{
A character vector containing
Olink ID of assays with all quantified values NA,
otherwise returns `character(0)`.
}
\description{
Help function to identify Olink assays with all quantified values NA.
}
\author{
Simon Forsberg
 Masoumeh Sheikhi
}
