% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_checksum.R
\name{check_checksum}
\alias{check_checksum}
\title{Help function comparing the checksum reported by Olink software to the
checksum of the delivered NPX file.}
\usage{
check_checksum(checksum_file, npx_file)
}
\arguments{
\item{checksum_file}{The plain file that contains the checksum output from
Olink software. The file should contain "MD5" or "SHA256" in the file name.}

\item{npx_file}{The NPX file accompanying the checksum file.}
}
\value{
A string or NA. If the function return NA then everything worked
fine, otherwise there was some sort of error.
}
\description{
Help function comparing the checksum reported by Olink software to the
checksum of the delivered NPX file.
}
\author{
Klev Diamanti
}
