% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_columns}
\alias{check_columns}
\title{Check presence of columns in dataset.}
\usage{
check_columns(df, col_list)
}
\arguments{
\item{df}{An Olink dataset (tibble or ArrowObject).}

\item{col_list}{A list of character vectors.}
}
\value{
Nothing or an error message if any column is missing.
}
\description{
Check if the input dataset (tibble or ArrowObject) \var{df} contains columns
specified in \var{col_list}. \var{col_list} supports both exact matches of
column names and alternative column names. In the latter case, alternative
column names are elements of a character vector, and exactly one of the
elements is required to be present.
}
\details{
\var{col_list} contains a collection of character vectors. If a character
vector is scalar (length = 1), the element of the vector is expected to be
present among the column names of \var{df}. When an element of \var{col_list}
contains more than one elements, the function will check whether the column
names of \var{df} include \strong{at least one} of the elements of that
vector.
}
\examples{
\dontrun{
tmp_data <- dplyr::tibble(
  "A" = c(1,2,3),
  "B" = c(T,T,F),
  "C" = c("A", "B", "C"),
  "D" = c(F,F,T)
)

# OK
check_columns(df = tmp_data,
              col_list = list("A", "B"))

# ERROR: E is missing
check_columns(df = tmp_data,
              col_list = list("A", "E"))

# ERROR: E and F are missing
check_columns(df = tmp_data,
              col_list = list("A", "E", "F"))

# OK
check_columns(df = tmp_data,
              col_list = list("A", c("B", "C")))

# OK
check_columns(df = tmp_data,
              col_list = list("A", c("B", "E")))

# ERROR: c(F, E) are missing
check_columns(df = tmp_data,
              col_list = list("A", c("F", "E")))

# ERROR: c(F, E) and c(M, N) are missing
check_columns(df = tmp_data,
              col_list = list("A", c("F", "E"), c("M", "N")))
}

}
\author{
Klev Diamanti
Albin Lundin
Lei Conze
Pascal Pucholt
Gil Henriques
}
