% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_NPX_data.R
\name{check_darid}
\alias{check_darid}
\title{Check DarIDs
DarIDs D.07, 08, 10, and 14 need to be exported with
Panel Data Archive Version 1.5 or later}
\usage{
check_darid(df)
}
\arguments{
\item{df}{NPX data}
}
\value{
warning messages
}
\description{
Check DarIDs
DarIDs D.07, 08, 10, and 14 need to be exported with
Panel Data Archive Version 1.5 or later
}
\keyword{internal}
