% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_darid}
\alias{check_darid}
\title{Help function checking for DARID and PanelDataArchiveVersion combinations}
\usage{
check_darid(df, col_names)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}
}
\value{
A warning message if any invalid combinations are found.
}
\description{
DarIDs D.07, 08, 10, and 14 need to exported with Panel Data Archive Version
1.5 or later. This function identifies cases where \var{DataAnalysisRefID}
are paired with earlier \var{PanelDataArchiveVersion}.
}
\author{
Kathleen Nevola
Kang Dong
Klev Diamanti
}
\keyword{internal}
