% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_sample_ids.R
\name{check_duplicate_sample_ids}
\alias{check_duplicate_sample_ids}
\title{Help function checking for duplicate sample IDs in data}
\usage{
check_duplicate_sample_ids(df, col_names)
}
\arguments{
\item{df}{A tibble or an arrow object containing
the columns "SampleID" and "OlinkID".}

\item{col_names}{A list of matched column names.
This is the output of `check_npx_col_names` function.}
}
\description{
Help function checking for duplicate sample IDs in data
}
\author{
Masoumeh Sheikhi
}
