% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_delim.R
\name{check_field_separator}
\alias{check_field_separator}
\title{Help function to check and get the file's field separator.}
\usage{
check_field_separator(file, sep = NULL)
}
\arguments{
\item{file}{Path to Olink software output in txt or csv.}

\item{sep}{The separator of the file: NULL (autodetect), comma (,) or
semicolon (;).}
}
\value{
The file separator comma (;) or semicolon (;)
}
\description{
Help function to check and get the file's field separator.
}
\author{
Klev Diamanti
}
