% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Help function checking if file exists.}
\usage{
check_file_exists(file, error = FALSE)
}
\arguments{
\item{file}{Path to file.}

\item{error}{Return error or a boolean (default = FALSE).}
}
\value{
Boolean if the file exists or not, and an error if "error = TRUE".
}
\description{
Check \strong{only one file at a time} if it exists.
}
\author{
Klev Diamanti
}
