% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Help function checking if file exists.}
\usage{
check_file_exists(file, error = FALSE)
}
\arguments{
\item{file}{Path to Olink software output file in wide or long format.
Expecting extensions "xls" or "xlsx" for excel files, "csv" or "txt" for delim files, "parquet" for parquet files, and "zip" for compressed files.}

\item{error}{Scalar boolean to return an error instead of a \code{FALSE}
(\code{default = FALSE}).}
}
\value{
\code{TRUE} if the file exists, and \code{FALSE} if not; error if the file does
not exist and \code{error = TRUE}.
}
\description{
Check \strong{one file at a time} if it exists.
}
\author{
Klev Diamanti
}
