% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_exists.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Help function checking if file exists.}
\usage{
check_file_exists(file, error = FALSE)
}
\arguments{
\item{file}{Path to the file.}

\item{error}{Boolean to return error or a boolean (default).}
}
\value{
Boolean if the file exists or not, and an error if \code{error = TRUE}.
}
\description{
Only one file at a time!
}
\author{
Klev Diamanti
}
