% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file.R
\name{check_file_extension}
\alias{check_file_extension}
\title{Help function checking if file extension is acceptable.}
\usage{
check_file_extension(file)
}
\arguments{
\item{file}{Path to Olink software output file in wide or long format.
Expecting extensions "xls" or "xlsx" for excel files, "csv" or "txt" for delim files, "parquet" for parquet files, and "zip" for compressed files.}
}
\value{
The type of the file extension based on the global variable
\var{accepted_npx_file_ext}.
}
\description{
Use variable \var{accepted_npx_file_ext} to check if extension of the input
file is acceptable. Expecting one of
""xls" or "xlsx" for excel files, "csv" or "txt" for delim files, "parquet" for parquet files, and "zip" for compressed files".
}
\author{
Klev Diamanti
}
