% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_is_arrow_object.R
\name{check_is_arrow_object}
\alias{check_is_arrow_object}
\title{Help function to check if a variable is an R6 ArrowObject.}
\usage{
check_is_arrow_object(var)
}
\arguments{
\item{var}{Variable to check.}
}
\value{
An error if the variable is not an R6 ArrowObject.
}
\description{
Help function to check if a variable is an R6 ArrowObject.
}
