% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check_is_arrow_object}
\alias{check_is_arrow_object}
\title{Help function checking if a variable is an R6 ArrowObject.}
\usage{
check_is_arrow_object(df, error = FALSE)
}
\arguments{
\item{df}{Variable to check.}

\item{error}{Scalar boolean to return an error instead of a \code{TRUE}/\code{FALSE}
(\code{default = FALSE}).}
}
\value{
\code{TRUE} if the variable is an ArrowObject, and \code{FALSE} if not; error
if the variable is not an ArrowObject and \code{error = TRUE}.
}
\description{
Help function checking if a variable is an R6 ArrowObject.
}
\seealso{
\code{\link{check_is_tibble}}
\code{\link{check_is_dataset}}
}
\author{
Klev Diamanti
}
