% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check_is_arrow_object}
\alias{check_is_arrow_object}
\title{Help function checking if a data frame is an R6 ArrowObject.}
\usage{
check_is_arrow_object(df, error = FALSE)
}
\arguments{
\item{df}{Data frame to check.}

\item{error}{`TRUE` to return an error if \var{df} is not an arrow object;
`FALSE` (default) to return a boolean.}
}
\value{
A scalar boolean if the object is an arrow object or not, and an
error if \var{error} = `TRUE`.
}
\description{
Help function checking if a data frame is an R6 ArrowObject.
}
\seealso{
\code{\link{check_is_tibble}}
  \code{\link{check_is_arrow_or_tibble}}
}
\author{
Klev Diamanti
}
