% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boolean.R
\name{check_is_boolean}
\alias{check_is_boolean}
\title{Help function checking if a variable is a boolean vector.}
\usage{
check_is_boolean(bool, error = FALSE)
}
\arguments{
\item{bool}{Input to check.}

\item{error}{Boolean to return error or a boolean (default).}
}
\value{
Boolean if the variable is a boolean vector or not, and an error if
`error = TRUE`.
}
\description{
Help function checking if a variable is a boolean vector.
}
\seealso{
[check_is_character()]
  [check_is_integer()]
  [check_is_numeric()]
}
\author{
Klev Diamanti
}
