% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boolean.R
\name{check_is_boolean}
\alias{check_is_boolean}
\title{Help function checking if a variable is a boolean vector.}
\usage{
check_is_boolean(bool, error = FALSE)
}
\arguments{
\item{bool}{Boolean to check.}

\item{error}{Return error or a boolean (default = FALSE).}
}
\value{
Boolean if the variable is a boolean vector or not, and an error if
"error = TRUE".
}
\description{
Help function checking if a variable is a boolean vector.
}
\seealso{
\code{\link{check_is_character}}
\code{\link{check_is_integer}}
\code{\link{check_is_numeric}}
}
\author{
Klev Diamanti
}
