% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_character.R
\name{check_is_character}
\alias{check_is_character}
\title{Help function checking if a variable is a character vector.}
\usage{
check_is_character(string, error = FALSE)
}
\arguments{
\item{string}{String variable to check.}

\item{error}{Return error or a boolean (default = FALSE).}
}
\value{
Boolean if the variable is a character vector or not, and an error if
"error = TRUE".
}
\description{
Help function checking if a variable is a character vector.
}
\seealso{
\code{\link{check_is_boolean}}
\code{\link{check_is_integer}}
\code{\link{check_is_numeric}}
}
\author{
Klev Diamanti
}
