% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_is_character.R
\name{check_is_character}
\alias{check_is_character}
\title{Help function checking if a variable is a character vector.}
\usage{
check_is_character(string)
}
\arguments{
\item{string}{Input to check.}
}
\value{
An error if the input is not a character vector.
}
\description{
Help function checking if a variable is a character vector.
}
