% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check_is_dataset}
\alias{check_is_dataset}
\title{Help function checking if a data frame is a tibble or an arrow object.}
\usage{
check_is_dataset(df, error = FALSE)
}
\arguments{
\item{df}{Data frame to check.}

\item{error}{Return error or a boolean (default = FALSE).}
}
\value{
A scalar boolean if the object is a tibble or an arrow object or not,
and an error if "error = TRUE".
}
\description{
Help function checking if a data frame is a tibble or an arrow object.
}
\seealso{
\code{\link{check_is_arrow_object}}
\code{\link{check_is_tibble}}
}
\author{
Klev Diamanti
}
