% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_integer.R
\name{check_is_integer}
\alias{check_is_integer}
\title{Help function checking if a variable is an integer vector.}
\usage{
check_is_integer(int, error = FALSE)
}
\arguments{
\item{int}{Integer to check.}

\item{error}{Return error or a boolean (default = FALSE).}
}
\value{
Boolean if the variable is an integer vector or not, and an error if
"error = TRUE".
}
\description{
Help function checking if a variable is an integer vector.
}
\seealso{
\code{\link{check_is_boolean}}
\code{\link{check_is_character}}
\code{\link{check_is_numeric}}
}
\author{
Klev Diamanti
}
