% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_is_integer.R
\name{check_is_integer}
\alias{check_is_integer}
\title{Help function checking if a variable is a integer vector.}
\usage{
check_is_integer(int, error = FALSE)
}
\arguments{
\item{int}{Input to check.}

\item{error}{Boolean to return error or a boolean (default).}
}
\value{
An error if the input is not a character vector.
}
\description{
Help function checking if a variable is a integer vector.
}
