% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_list.R
\name{check_is_list}
\alias{check_is_list}
\title{Help function to check if the variable is a list.}
\usage{
check_is_list(lst, error = FALSE)
}
\arguments{
\item{lst}{List to check.}

\item{error}{Return error or a boolean (default = FALSE).}
}
\value{
Boolean if the object is a list or not, and an error if
"error = TRUE".
}
\description{
Help function to check if the variable is a list.
}
\author{
Klev Diamanti
}
