% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_is_numeric.R
\name{check_is_numeric}
\alias{check_is_numeric}
\title{Help function checking if a variable is a numeric vector.}
\usage{
check_is_numeric(num, error = FALSE)
}
\arguments{
\item{num}{Input to check.}

\item{error}{Boolean to return error or a boolean (default).}
}
\value{
An error if the input is not a character vector.
}
\description{
Help function checking if a variable is a numeric vector.
}
