% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_is_numeric}
\alias{check_is_numeric}
\title{Help function checking if a variable is a vector of numerics.}
\usage{
check_is_numeric(num, error = FALSE)
}
\arguments{
\item{num}{Variable to check.}

\item{error}{Scalar boolean to return an error instead of a \code{TRUE}/\code{FALSE}
(\code{default = FALSE}).}
}
\value{
\code{TRUE} if the variable is a numeric vector, and \code{FALSE} if not;
error if the variable is not a numeric vector and \code{error = TRUE}.
}
\description{
Help function checking if a variable is a vector of numerics.
}
\seealso{
\code{\link{check_is_boolean}}
\code{\link{check_is_character}}
\code{\link{check_is_integer}}
}
\author{
Klev Diamanti
}
