% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boolean.R
\name{check_is_scalar_boolean}
\alias{check_is_scalar_boolean}
\title{Help function checking if a variable is a scalar boolean.}
\usage{
check_is_scalar_boolean(bool, error = FALSE)
}
\arguments{
\item{bool}{Variable to check.}

\item{error}{Scalar boolean to return an error instead of a \code{TRUE}/\code{FALSE}
(\code{default = FALSE}).}
}
\value{
\code{TRUE} if the variable is a boolean vector of length 1, and \code{FALSE}
if not; error if the variable is not boolean vector of length 1, and
\code{error = TRUE}.
}
\description{
Help function checking if a variable is a scalar boolean.
}
\seealso{
\code{\link{check_is_scalar_character}}
\code{\link{check_is_scalar_integer}}
\code{\link{check_is_scalar_numeric}}
}
\author{
Klev Diamanti
}
