% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_character.R
\name{check_is_scalar_character}
\alias{check_is_scalar_character}
\title{Help function checking if a variable is a character vector of length 1.}
\usage{
check_is_scalar_character(string, error = FALSE)
}
\arguments{
\item{string}{String variable to check.}

\item{error}{Return error or a boolean (default = FALSE).}
}
\value{
Boolean if the variable is a character vector of length 1 or not, and
an error if "error = TRUE".
}
\description{
Help function checking if a variable is a character vector of length 1.
}
\seealso{
\code{\link{check_is_scalar_boolean}}
\code{\link{check_is_scalar_integer}}
\code{\link{check_is_scalar_numeric}}
}
\author{
Klev Diamanti
}
