% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_character.R
\name{check_is_scalar_character}
\alias{check_is_scalar_character}
\title{Help function checking if a variable is a scalar character.}
\usage{
check_is_scalar_character(string, error = FALSE)
}
\arguments{
\item{string}{Variable to check.}

\item{error}{Scalar boolean to return an error instead of a \code{FALSE}
(\code{default = FALSE}).}
}
\value{
\code{TRUE} if the variable is a character vector of length 1, and \code{FALSE}
if not; error if the variable is not a character vector of length 1, and
\code{error = TRUE}.
}
\description{
Help function checking if a variable is a scalar character.
}
\seealso{
\code{\link{check_is_scalar_boolean}}
\code{\link{check_is_scalar_integer}}
\code{\link{check_is_scalar_numeric}}
}
\author{
Klev Diamanti
}
