% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_is_scalar_integer.R
\name{check_is_scalar_integer}
\alias{check_is_scalar_integer}
\title{Help function checking if a variable is an integer vector of length 1.}
\usage{
check_is_scalar_integer(int, error = FALSE)
}
\arguments{
\item{int}{Input to check.}

\item{error}{Boolean to return error or a boolean (default).}
}
\value{
An error if the input is not an integer vector of length 1
}
\description{
Help function checking if a variable is an integer vector of length 1.
}
