% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_is_scalar_numeric}
\alias{check_is_scalar_numeric}
\title{Help function checking if a variable is a scalar numeric}
\usage{
check_is_scalar_numeric(num, error = FALSE)
}
\arguments{
\item{num}{Variable to check.}

\item{error}{Scalar boolean to return an error instead of a \code{FALSE}
(\code{default = FALSE}).}
}
\value{
\code{TRUE} if the variable is a numeric vector of length 1, and \code{FALSE}
if not; error if the variable is not a numeric vector of length 1, and
\code{error = TRUE}.
}
\description{
Help function checking if a variable is a scalar numeric
}
\seealso{
\code{\link{check_is_scalar_boolean}}
\code{\link{check_is_scalar_character}}
\code{\link{check_is_scalar_integer}}
}
\author{
Klev Diamanti
}
