% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_is_scalar_numeric}
\alias{check_is_scalar_numeric}
\title{Help function checking if a variable is a numeric vector of length 1.}
\usage{
check_is_scalar_numeric(num, error = FALSE)
}
\arguments{
\item{num}{Input to check.}

\item{error}{Boolean to return error or a boolean (default).}
}
\value{
Boolean if the variable is a numeric vector of length 1 or not, and
an error if `error = TRUE`.
}
\description{
Help function checking if a variable is a numeric vector of length 1.
}
\seealso{
[check_is_scalar_boolean()]
  [check_is_scalar_character()]
  [check_is_scalar_integer()]
}
\author{
Klev Diamanti
}
