% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check_is_tibble}
\alias{check_is_tibble}
\title{Help function checking if a variable is a tibble dataset.}
\usage{
check_is_tibble(df, error = FALSE)
}
\arguments{
\item{df}{Variable to check.}

\item{error}{Scalar boolean to return an error instead of a \code{TRUE}/\code{FALSE}
(\code{default = FALSE}).}
}
\value{
\code{TRUE} if the variable is a tibble, and \code{FALSE} if not; error if the
variable is not a tibble and \code{error = TRUE}.
}
\description{
Help function checking if a variable is a tibble dataset.
}
\seealso{
\code{\link{check_is_arrow_object}}
\code{\link{check_is_dataset}}
}
\author{
Klev Diamanti
}
