% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check_is_tibble}
\alias{check_is_tibble}
\title{Help function checking if a data frame is a tibble data frame.}
\usage{
check_is_tibble(df, error = FALSE)
}
\arguments{
\item{df}{Data frame to check.}

\item{error}{`TRUE` to return an error if \var{df} is not a tibble; `FALSE`
(default) to return a boolean.}
}
\value{
A scalar boolean if the object is a tibble or not, and an error if
\var{error} = `TRUE`.
}
\description{
Help function checking if a data frame is a tibble data frame.
}
\seealso{
\code{\link{check_is_arrow_object}}
  \code{\link{check_is_data_frame}}
  \code{\link{check_is_arrow_or_data_frame}}
}
\author{
Klev Diamanti
}
