% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_library_installed.R
\name{check_library_installed}
\alias{check_library_installed}
\title{Help function to check if suggested libraries are installed when required.}
\usage{
check_library_installed(libraries, error = FALSE)
}
\arguments{
\item{libraries}{Character vector of R libraries.}

\item{error}{Return error or a boolean (default = FALSE).}
}
\value{
Boolean if the library is installed or not, and an error if
"error = TRUE".
}
\description{
Checks whether a collection of R libraries is available.
}
\author{
Klev Diamanti
}
