% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_library_installed.R
\name{check_library_installed}
\alias{check_library_installed}
\title{Help function to check is suggested libraries are installed when required.}
\usage{
check_library_installed(libraries)
}
\arguments{
\item{libraries}{A character vector of R libraries.}
}
\value{
An error if one of the libraries is missing.
}
\description{
Help function to check is suggested libraries are installed when required.
}
