% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_r_library.R
\name{check_library_installed}
\alias{check_library_installed}
\title{Help function checking if one or more R libraries are installed.}
\usage{
check_library_installed(libraries, error = FALSE)
}
\arguments{
\item{libraries}{Character vector of R libraries.}

\item{error}{Scalar boolean to return an error instead of a \code{FALSE}
(\code{default = FALSE}).}
}
\value{
\code{TRUE} if the libraries are installed, and \code{FALSE} if not; error if
the libraries are not installed and \code{error = TRUE}.
}
\description{
Help function checking if one or more R libraries are installed.
}
\author{
Klev Diamanti
}
