% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx_all_na_assays}
\alias{check_npx_all_na_assays}
\title{Help function to identify Olink assays with all quantified values \emph{NA}}
\usage{
check_npx_all_na_assays(df, col_names)
}
\arguments{
\item{df}{A tibble or an arrow object containing columns \emph{OlinkID} and
the quantification column Ct, NPX or Quantified}

\item{col_names}{A list of matched column names. This is the output of the
\emph{check_npx_col_names} function.}
}
\value{
A character vector containing OlinkIDs of assays with quantified
values \emph{NA} for all samples, otherwise returns \emph{character(0)}.
}
\description{
This function checks if there are assays with the quantified values for all
samples \emph{NA}.
}
\author{
Simon Forsberg;
Masoumeh Sheikhi
}
