% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx_all_na_sample}
\alias{check_npx_all_na_sample}
\title{Help function to identify Olink samples with all quantified values \emph{NA}}
\usage{
check_npx_all_na_sample(df, col_names)
}
\arguments{
\item{df}{A tibble or an arrow object containing columns \var{SampleID} and
the quantification column
"Ct", "NPX", or "Quantified".}

\item{col_names}{A list of matched column names. This is the output of the
\var{check_npx_col_names} function.}
}
\value{
A character vector containing \var{SampleID} of samples with
quantified values \emph{NA} for all assays, otherwise returns
\emph{character(0)}.
}
\description{
This function checks if there are samples with the quantified values for all
assays \emph{NA}.
}
\details{
We have added the tags importFrom for "dbplyr" and "duckdb" because
"devtools::check()" would complain with a note that the two libraries are
imported but never used. To avoid that we used solutions taken from here:
\enumerate{
\item https://github.com/hadley/r-pkgs/issues/203
\item https://github.com/pbs-software/pbs-modelling/issues/95
}
}
\author{
Simon Forsberg;
Masoumeh Sheikhi
Klev Diamanti
}
