% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx_col_names.R
\name{check_npx_col_names}
\alias{check_npx_col_names}
\title{Check, update and define column names used in downstream analyses.}
\usage{
check_npx_col_names(df, preferred_names = NULL)
}
\arguments{
\item{df}{A tibble or and arrow object from \code{read_npx}.}

\item{preferred_names}{A named character vector where names are internal
column names and values are column names to be selected from current data
frame. Read the \code{description} for further information.}
}
\value{
A list of matched column names based on the preferred names.
}
\description{
OlinkAnalyze uses pre-defined names of columns of data frames to perform
downstream analyses. At the same time, different Olink platforms are capable
of exporting data with different quantification approaches. This function
aims to instruct each function of OlinkAnalyze on the column it should be
using for the downstream analysis. This should be seamless for data exported
from Olink Software and imported to R using the read_npx function.

However, in certain cases the columns of interest might be named differently.
This function allows assigning custom-named columns of a data frame to
internally expected variables that will in turn instruct Olink Analyze
functions to use them for downstream analysis. For example, if one has
transformed the existing NPX value and has stored the result in a new column
called NPX_2, then they can assign this new name to the internal variable
\code{quant} to inform the package that in the downstream analysis \code{NPX_2} should
be used. See example 1.

Similarly, in case of multiple matches (e.g. the data frame contains both
columns \code{NPX} and \code{Quantified_value}) the ties will need to be resolved by
the user using the argument \code{preferred_names} from this function.  See
example 2.

The argument \code{preferred_names} is a named character vector with internal
column names as names and column names of the current data set as values.
Names of the input vector can be one or more of the following: \code{sample_id},
\code{sample_type}, \code{olink_id}, \code{plate_id}, \code{qc_warning}, \code{lod}, \code{quant}.
}
\examples{
\donttest{
file <- system.file("extdata",
                    "npx_data_ext.parquet",
                     package = "OlinkAnalyze")
df <- read_npx(filename = file)

# run df as is
OlinkAnalyze:::check_npx_col_names(df = df)

# SampleType missing
df |>
  dplyr::select(
    -dplyr::all_of("SampleType")
  ) |>
  OlinkAnalyze:::check_npx_col_names()

# use PCNormalizedNPX instead on NPX
OlinkAnalyze:::check_npx_col_names(
  df = df,
  preferred_names = c("quant" = "PCNormalizedNPX")
)

# use PCNormalizedNPX instead on NPX, and PlateLOD instead of LOD
df |>
  dplyr::mutate(
    LOD = 1L,
    PlateLOD = 2L
  ) |>
  OlinkAnalyze:::check_npx_col_names(
    preferred_names = c("quant" = "PCNormalizedNPX",
                        "lod" = "PlateLOD")
  )
}

}
\author{
Klev Diamanti
Masoumeh Sheikhi
}
