% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx_duplicate_sample_ids}
\alias{check_npx_duplicate_sample_ids}
\title{Help function checking for duplicate sample identifiers in data.}
\usage{
check_npx_duplicate_sample_ids(df, col_names)
}
\arguments{
\item{df}{A tibble or an arrow object containing the columns \emph{SampleID}
and \emph{OlinkID}.}

\item{col_names}{A list of matched column names. This is the output of the
\emph{check_npx_col_names} function.}
}
\value{
A character vector of duplicate SampleIDs found in the data.
}
\description{
This function checks if there are duplicate sample identifiers for any assay.
}
\author{
Masoumeh Sheikhi
}
