% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx_olinkid}
\alias{check_npx_olinkid}
\title{Help function checking whether df contains invalid Olink IDs}
\usage{
check_npx_olinkid(df, col_names)
}
\arguments{
\item{df}{A tibble or an arrow object containing containing column "OlinkID"}

\item{col_names}{A list of matched column names.
This is the output of `check_npx_col_names` function.}
}
\value{
A character vector including invalid Olink IDs
}
\description{
Help function checking whether df contains invalid Olink IDs
}
\author{
Masoumeh Sheikhi
}
