% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx_olinkid}
\alias{check_npx_olinkid}
\title{Help function checking whether df contains invalid Olink identifiers}
\usage{
check_npx_olinkid(df, col_names)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{col_names}{A list of matched column names. This is the output of the
\var{check_npx_col_names} function.}
}
\value{
A character vector with invalid \var{OlinkID}.
}
\description{
This function checks if Olink identifiers (\var{OlinkID}) match the pattern
of a prefix "OID" followed by 5 integer numbers.
}
\author{
Masoumeh Sheikhi
}
