% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{check_npx_update_col_names}
\alias{check_npx_update_col_names}
\title{Update column names to be used in downstream analyses}
\usage{
check_npx_update_col_names(preferred_names)
}
\arguments{
\item{preferred_names}{A named character vector where names are internal
column names and values are column names to be selected from the input data
frame. Read the \emph{description} for further information.}
}
\value{
\var{column_name_dict} updated based on \var{preferred_names}.
}
\description{
OlinkAnalyze uses pre-defined names of columns of data frames to perform
downstream analyses. However, in certain cases the columns of interest might
be named differently. The aim of this function is to assign custom-named
columns of a data frame to internally expected variables that will in turn
enable analysis of Olink data. For example, if one wished to #' use the
column \var{PCNormalizedNPX} for their analysis instead of the column
\var{NPX}, then they can assign this new name to the internal variable
\var{quant} to inform the package that in the downstream analysis
\var{PCNormalizedNPX} should be used.

This function takes as input a named character vector with internal column
names as names and column names of the current data set as values. Names of
the input vector can be one or more of the following:
`r cli::ansi_collapse(x = names(column_name_dict))`
}
\author{
Klev Diamanti
Masoumeh Sheikhi
}
