% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{check_olink_broader_platform}
\alias{check_olink_broader_platform}
\title{Help function checking that the broader Olink platform is expected.}
\usage{
check_olink_broader_platform(x)
}
\arguments{
\item{x}{Name of the broader Olink platform. One of
"NGS" and "qPCR".}
}
\value{
\code{NULL} if broader Olink platform is expected, otherwise an error.
}
\description{
Help function checking that the broader Olink platform is expected.
}
\author{
Klev Diamanti
}
