% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_utils.R
\name{check_olink_broader_platform}
\alias{check_olink_broader_platform}
\title{Help function checking that the broader Olink platform is acceptable.}
\usage{
check_olink_broader_platform(x)
}
\arguments{
\item{x}{Name of the broader Olink platform. One of \code{qPCR} or \code{NGS}.}
}
\value{
Nothing if broader Olink platform is ok, otherwise an error.
}
\description{
Help function checking that the broader Olink platform is acceptable.
}
