% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{check_olink_data_type}
\alias{check_olink_data_type}
\title{Help function checking that \var{data_type} is expected.}
\usage{
check_olink_data_type(
  x,
  broad_platform = NULL,
  platform_name = NULL,
  quant_type = NULL
)
}
\arguments{
\item{x}{The name of the Olink data type. One of
"Ct", "NPX", and "Quantified".}

\item{broad_platform}{Name of the \var{broad_platform} to filter for. If
\code{NULL} all broader platforms are considered, otherwise one of
"NGS" and "qPCR" is expected.}

\item{platform_name}{Name of the \var{platform_name} to filter for. If \code{NULL}
all platforms are considered, otherwise one of
"Explore 3072", "Explore HT", "Flex", "Focus", "Reveal", "Target 48", and "Target 96" is expected.}

\item{quant_type}{Name of the \var{quant_type} to filter for. If \code{NULL} all
quantification types are considered, otherwise one of
"absolute" and "relative" is expected.}
}
\value{
\code{NULL} if quantification method (data type) is expected, otherwise an error.
}
\description{
Help function checking that \var{data_type} is expected.
}
\author{
Klev Diamanti
}
