% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_utils.R
\name{check_olink_data_type}
\alias{check_olink_data_type}
\title{Help function checking that the Olink data_type is acceptable.}
\usage{
check_olink_data_type(x, broader_platform = NULL)
}
\arguments{
\item{x}{The name of the Olink data type. One of \code{NPX}, \code{Quantified} or \code{Ct}.}

\item{broader_platform}{Name of the broader Olink platform. One of \code{qPCR} or
\code{NGS}.}
}
\value{
Nothing if data_type is ok, otherwise an error.
}
\description{
Help function checking that the Olink data_type is acceptable.
}
