% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{check_olink_platform}
\alias{check_olink_platform}
\title{Help function checking that \var{olink_platform} is expected.}
\usage{
check_olink_platform(x, broader_platform = NULL)
}
\arguments{
\item{x}{Name of Olink platform. One of
"Explore 3072", "Explore HT", "Flex", "Focus", "Target 48", and "Target 96".}

\item{broader_platform}{Name of the broader Olink platform. One of
"NGS" and "qPCR".}
}
\value{
\code{NULL} if platform is expected, otherwise an error.
}
\description{
Help function checking that \var{olink_platform} is expected.
}
\author{
Klev Diamanti
}
