% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_utils.R
\name{check_olink_platform}
\alias{check_olink_platform}
\title{Help function checking that the olink_platform is acceptable.}
\usage{
check_olink_platform(x, broader_platform = NULL)
}
\arguments{
\item{x}{The name of the Olink platform. One of \verb{Explore 3072}, \verb{Explore HT},
\verb{Target 96}, \verb{Target 48}, \code{Flex} or \code{Focus}.}

\item{broader_platform}{Name of the broader Olink platform. One of \code{qPCR} or
\code{NGS}.}
}
\value{
Nothing if platform is ok, otherwise an error.
}
\description{
Help function checking that the olink_platform is acceptable.
}
