% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_utils.R
\name{check_olink_platform}
\alias{check_olink_platform}
\title{Help function checking that the olink_platform is acceptable.}
\usage{
check_olink_platform(x, broader_platform = NULL)
}
\arguments{
\item{x}{The name of the Olink platform. One of
Target 48, Flex, Target 96, Explore 3072, Explore HT or Focus.}

\item{broader_platform}{Name of the broader Olink platform. One of
qPCR and NGS. # nolint}
}
\value{
\code{NULL} if platform is ok, otherwise an error.
}
\description{
Help function checking that the olink_platform is acceptable.
}
\author{
Klev Diamanti
}
