% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_utils.R
\name{check_olink_platform}
\alias{check_olink_platform}
\title{Help function checking that the olink_platform is acceptable.}
\usage{
check_olink_platform(x, broader_platform = NULL)
}
\arguments{
\item{x}{The name of the Olink platform. One of `Explore 3072`, `Explore HT`,
`Target 96`, `Target 48`, `Flex` or `Focus`.}

\item{broader_platform}{Name of the broader Olink platform. One of `qPCR` or
`NGS`.}
}
\value{
Nothing if platform is ok, otherwise an error.
}
\description{
Help function checking that the olink_platform is acceptable.
}
