% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_olink_vars.R
\name{check_olink_platform}
\alias{check_olink_platform}
\title{Help function checking with the olink_platform is acceptable}
\usage{
check_olink_platform(x, broader_platform = NULL)
}
\arguments{
\item{x}{The name of the Olink platform. One of "Target 48", "Flex",
"Target 96", "Explore 3072", "Explore HT" or "Focus".}

\item{broader_platform}{Name of the broader Olink platform. On of "qPCR" or
"NGS".}
}
\value{
Nothing if platform is ok, otherwise an error.
}
\description{
Help function checking with the olink_platform is acceptable
}
