% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_olinkid.R
\name{check_olinkid}
\alias{check_olinkid}
\title{Help function checking whether df contains recognized OIDs}
\usage{
check_olinkid(df)
}
\arguments{
\item{df}{An arrow object}
}
\value{
A character vector indicating the invalid OlinkIDs
}
\description{
Help function checking whether df contains recognized OIDs
}
\author{
Simon Forsberg
 Masoumeh Sheikhi
}
