% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_output.R, R/read_npx_utils.R
\name{check_out_df_arg}
\alias{check_out_df_arg}
\title{Help function checking that the requested output class of the read_npx*
functions is acceptable.}
\usage{
check_out_df_arg(out_df)

check_out_df_arg(out_df)
}
\arguments{
\item{out_df}{The class of output data frame. One of `tibble` (default) or
`arrow` for ArrowObject.}
}
\value{
An error if the argument is not as expected.

An error if the argument is not as expected.
}
\description{
Help function checking that the requested output class of the read_npx*
functions is acceptable.

Help function checking that the requested output class of the read_npx*
functions is acceptable.
}
\author{
Klev Diamanti
}
