% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_utils.R
\name{check_out_df_arg}
\alias{check_out_df_arg}
\title{Help function checking that the requested output class of the read_npx*
functions is acceptable.}
\usage{
check_out_df_arg(out_df)
}
\arguments{
\item{out_df}{The class of the output dataset. One of
"tibble" or "arrow". (default =
"tibble")}
}
\value{
Error if \var{out_df} is not one of
"tibble" and "arrow".
}
\description{
Help function checking that the requested output class of the read_npx*
functions is acceptable.
}
\author{
Klev Diamanti
}
