% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_assay_na}
\alias{clean_assay_na}
\title{Help function removing assays with all quantified values \code{NA}.}
\usage{
clean_assay_na(df, check_log, remove_assay_na = TRUE, verbose = FALSE)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{check_log}{A named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{\link[=check_npx]{check_npx()}} will be run internally using \code{df}.}

\item{remove_assay_na}{Logical. If \code{FALSE}, skips filtering assays with all
quantified values \code{NA}. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise messages.}
}
\description{
This function filters out rows from a \code{tibble} or \code{arrow} object where the
assay identifier (one of
"OlinkID", "OID", "olinkid", "oid", or "olink_id")
matches those listed in \code{check_npx_log$assay_na}, which contains assays
composed entirely of \code{NA} values in their quantification column.
}
\author{
Kang Dong
Klev Diamanti
}
