% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_assay_na}
\alias{clean_assay_na}
\title{Help function removing assays with all quantified values \code{NA}.}
\usage{
clean_assay_na(df, check_npx_log, verbose = FALSE)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loads from \code{read_npx()}, including a
column identified by
"OlinkID", "OID", "olinkid", "oid", or "olink_id".}

\item{check_npx_log}{A named list generated by the function \code{check_npx()},
containing:
\itemize{
\item \code{assay_na}: a character vector of assay identifiers with all quantified
values \code{NA}.
\item \code{col_names$olink_id}: the column name of the assay identifier in the
dataset.
}}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise CLI
messages.}
}
\value{
A \code{tibble} or \code{arrow} object with rows corresponding to assays with
all quantified values \code{NA} removed.
}
\description{
This function filters out rows from a \code{tibble} or \code{arrow} object where the
assay identifier (one of
"OlinkID", "OID", "olinkid", "oid", or "olink_id")
matches those listed in \code{check_npx_log$assay_na}, which contains assays
composed entirely of \code{NA} values in their quantification column.
}
\author{
Kang Dong
}
