% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_assay_type}
\alias{clean_assay_type}
\title{Help function removing control assays based on assay type.}
\usage{
clean_assay_type(df, check_log, remove_control_assay = TRUE, verbose = FALSE)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{check_log}{A named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{\link[=check_npx]{check_npx()}} will be run internally using \code{df}.}

\item{remove_control_assay}{If \code{FALSE}, all internal control assays are
retained. If \code{TRUE}, all internal control assays are removed. Alternatively,
a character vector with one or more of
"assay", "inc", "det", "ext", and "amp" indicating the assay
types to remove.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise messages.}
}
\description{
This function filters out internal control assays (\code{ext_ctrl}, \code{inc_ctrl},
\code{amp_ctrl}) from the dataset, unless user specified to retain them. The
function uses column mapping provided by \code{check_log}.
}
\author{
Kang Dong
Klev Diamanti
}
