% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_assay_type}
\alias{clean_assay_type}
\title{Remove Internal Control Assays by Assay Type}
\usage{
clean_assay_type(
  df,
  check_npx_log,
  control_assay_types = c("ext_ctrl", "inc_ctrl", "amp_ctrl"),
  out_df = "tibble"
)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded from \code{read_npx()}, containing
an assay type column.}

\item{check_npx_log}{A list generated by \code{check_npx()} that includes:
\itemize{
\item \code{col_names$assay_type}: the name of the column holding assay type
information.
}}

\item{control_assay_types}{A character vector of internal control assay
types to remove. Default is \code{c("ext_ctrl", "inc_ctrl", "amp_ctrl")}.}

\item{out_df}{The class of output data frame. Options are \code{"tibble"} or
\code{"arrow"}. Default: \code{"tibble"}.}
}
\value{
A filtered tibble or arrow object with internal control assay rows removed.
}
\description{
This function filters out rows from an NPX data where the assay type
matches any of the specified control assay types (e.g., \code{ext_ctrl},
\code{inc_ctrl}, \code{amp_ctrl}). Matching is case-insensitive and uses regular
expressions via \code{stringr::str_detect()}. Informative CLI messages report the
number and type of rows removed.
}
\examples{
\dontrun{

df <- dplyr::tibble(
  SampleID = c("A", "B", "C", "D"),
  OlinkID = rep("OID12345", 4L),
  NPX = rep(1, 4L),
  PlateID = rep("plate1", 4L),
  QC_Warning = rep("Pass", 4L),
  LOD = rep(1, 4L),
  AssayType = c("assay",
                "ext_ctrl",
                "inc_ctrl",
                "amp_ctrl")
)

log <- OlinkAnalyze::check_npx(df)

# Example 1: Run clean_assay_type() as default. Internal control assay are removed
clean_df <- OlinkAnalyze::clean_assay_type(df, log)

# Example 2: Set control_assay_types = NULL. Return unchanged data
clean_df <- OlinkAnalyze::clean_assay_type(df, log, control_assay_types = NULL)
}
}
\author{
Kang Dong
}
