% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_assay_type}
\alias{clean_assay_type}
\title{Help function removing control assays based on assay type.}
\usage{
clean_assay_type(
  df,
  check_npx_log,
  remove_control_assay = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded from \code{read_npx()}.}

\item{check_npx_log}{A named list generated by the function \code{check_npx()},
containing:
\itemize{
\item \code{col_names$olink_id}: the column name of the assay identifier.
\item \code{col_names$assay_type}: the column name of the assay type. \strong{If
column is missing then function will return the original dataset.}
}}

\item{remove_control_assay}{If \code{FALSE}, all internal control assays are
retained. If \code{TRUE}, all internal control assays are removed. Alternatively,
a character vector with one or more of
"assay", "inc", "det", "ext", and "amp" indicating the assay
types to remove.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise CLI
messages.}
}
\value{
A \code{tibble} or \code{arrow} object with rows corresponding to control
assays removed.
}
\description{
This function filters out internal control assays (\code{ext_ctrl}, \code{inc_ctrl},
\code{amp_ctrl}) from the dataset, unless user specified to retain them. The
function uses column mapping provided by \code{check_npx_log}.
}
\author{
Kang Dong
Klev Diamanti
}
