% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_assay_type}
\alias{clean_assay_type}
\title{Help function cleaning Internal Control Assays by Assay Type}
\usage{
clean_assay_type(
  df,
  check_npx_log,
  keep_control_assay = FALSE,
  out_df = "tibble"
)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded from \code{read_npx()}.}

\item{check_npx_log}{A list generated by \code{check_npx()} that includes:
\itemize{
\item \code{col_names$assay_type}: the name of the column holding assay type
information.
}}

\item{keep_control_assay}{Logical; if \code{TRUE}, internal control assays are
retained and no filtering is applied. Defaults to \code{FALSE}.}

\item{out_df}{Output format: either "tibble" (default) or "arrow".}
}
\value{
A cleaned tibble or arrow object with control assays removed.
}
\description{
This function filters out internal control assays (\code{ext_ctrl}, \code{inc_ctrl},
\code{amp_ctrl}) from the NPX data, unless user specified to retain them. The
function uses column mapping provided in the \code{check_npx_log} and returns
the \code{tibble} or \code{arrow} object as specified by the user.
}
\author{
Kang Dong
}
