% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_assay_warning}
\alias{clean_assay_warning}
\title{Help function cleaning assays flagged with assay warning}
\usage{
clean_assay_warning(df, check_npx_log, out_df = "tibble")
}
\arguments{
\item{df}{A data frame or tibble containing NPX data.}

\item{check_npx_log}{A list generated by \code{check_npx()} that includes:
\itemize{
\item \code{col_names$assay_warning}: the name of the column indicating QC status.
}}

\item{out_df}{A string indicating the desired output format. Passed to
\code{convert_read_npx_output()}. Common values include \code{"tibble"} and \code{"arrow"}.}
}
\value{
A cleaned tibble or arrow object with warning assay removed.
}
\description{
Removes assays from the NPX data frame where the Assay warning column (as
identified in \code{check_npx_log}) has a value of \code{"WARN"}. The function is used
to clean assay-level QC warnings from the dataset before analysis.
}
