% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_assay_warning}
\alias{clean_assay_warning}
\title{Help function removing instances of assays flagged with warnings.}
\usage{
clean_assay_warning(
  df,
  check_log,
  remove_assay_warning = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{check_log}{A named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{\link[=check_npx]{check_npx()}} will be run internally using \code{df}.}

\item{remove_assay_warning}{Logical. If \code{FALSE}, retains assays flagged as
\code{WARN} in assay warning. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise messages.}
}
\description{
The function is used to remove assay-level QC warnings from the dataset
before analysis. It uses the column marking assay QC warnings identified by
\code{check_log} to remove assays flagged as \code{WARN} in the dataset.
}
\author{
Kang Dong
Klev Diamanti
}
