% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_assay_warning}
\alias{clean_assay_warning}
\title{Help function removing instances of assays flagged with warnings.}
\usage{
clean_assay_warning(
  df,
  check_npx_log,
  remove_assay_warning = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{A tibble or and arrow object from \code{\link{read_npx}}.}

\item{check_npx_log}{A named list generated by the function \code{check_npx()},
containing:
\itemize{
\item \code{col_names$olink_id}: the column name of the assay identifier in the
dataset.
\item \code{col_names$assay_warning}: the name of the column indicating assay QC
status.
}}

\item{remove_assay_warning}{Logical. If \code{FALSE}, retains assays flagged as
\code{WARN} in assay warning. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise messages.}
}
\value{
A \code{tibble} or \code{arrow} object with rows corresponding assay with
warning flags removed.
}
\description{
The function is used to remove assay-level QC warnings from the dataset
before analysis. It uses the column marking assay QC warnings identified by
\code{check_npx_log} to remove assays flagged as \code{WARN} in the dataset.
}
\author{
Kang Dong
Klev Diamanti
}
