% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_col_class}
\alias{clean_col_class}
\title{Help function converting types of columns to the expected ones.}
\usage{
clean_col_class(df, check_log, convert_df_cols = TRUE, verbose = FALSE)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{check_log}{A named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{\link[=check_npx]{check_npx()}} will be run internally using \code{df}.}

\item{convert_df_cols}{Logical. If \code{FALSE}, retains columns of \code{df} as are.
Defaults to \code{TRUE}, were columns required for downstream analysis are
converted to the expected format.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise messages.}
}
\description{
This function checks for mismatches between actual and expected column
classes in the input data frame and coerces those columns to the expected
class using information from \code{check_npx_log$col_class}.
}
\author{
Kang Dong
Klev Diamanti
}
