% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_col_class}
\alias{clean_col_class}
\title{Help function converting types of columns to the expected ones.}
\usage{
clean_col_class(df, check_npx_log, convert_df_cols = TRUE, verbose = FALSE)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{check_npx_log}{A named list generated by the function \code{check_npx()},
containing:
\itemize{
\item \code{col_names$col_class}: a data frame containing column names that need
to be converted to a different type.
}}

\item{convert_df_cols}{Logical. If \code{FALSE}, retains columns of \code{df} as are.
Defaults to \code{TRUE}, were columns required for downstream analysis are
converted to the expected format.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise messages.}
}
\value{
A \code{tibble} or \code{arrow} object with types of selected columns converted
to the expected type.
}
\description{
This function checks for mismatches between actual and expected column
classes in the input data frame and coerces those columns to the expected
class using information from \code{check_npx_log$col_class}.
}
\author{
Kang Dong
Klev Diamanti
}
