% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_control_sample_id}
\alias{clean_control_sample_id}
\title{Clean Control Samples Based on Sample ID}
\usage{
clean_control_sample_id(
  df,
  check_npx_log,
  control_sample_ids = NULL,
  out_df = "tibble"
)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded using \code{read_npx()}, including
a sample ID column.}

\item{check_npx_log}{A list generated by \code{check_npx()} that includes:
\itemize{
\item \code{col_names$sample_id}: the column name in \code{df} identifying sample IDs.
}}

\item{out_df}{Output format of the returned data frame.
Options: \code{"tibble"} or \code{"arrow"}. Default is \code{"tibble"}.}

\item{control_sample_idA}{character vector of control sample IDs,
e.g., \code{c("control_a", "control_b")}. Default is \code{NULL}, in which case the
data is returned unchanged.}
}
\value{
A filtered data table with control samples removed.
}
\description{
This function removes rows from NPX data where the sample ID column (as
defined in \code{check_npx_log}) matches any of the provided control sample IDs.
This is useful for filtering out technical or customer control samples prior
to downstream analysis.
}
\examples{
\dontrun{
# Example 1: use npx_data1 to check that clean_control_sample_id() works
log <- OlinkAnalyze::npx_data1 |>
  OlinkAnalyze::check_npx() |>
  suppressWarnings()

out <- OlinkAnalyze::clean_control_sample_id(npx_data1,
  check_npx_log = log,
  control_sample_id = c("CONTROL_SAMPLE_AS 1", "CONTROL_SAMPLE_AS 2")
)

# Example 2: check default setting of clean_control_sample_id(). Return data
# unchanged.
out <- OlinkAnalyze::clean_control_sample_id(npx_data1,
  check_npx_log = log,
  control_sample_id = NULL
)
}
}
\author{
Kang Dong
}
