% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_control_sample_id}
\alias{clean_control_sample_id}
\title{Remove Control Samples Based on Sample ID Pattern}
\usage{
clean_control_sample_id(
  df,
  check_npx_log,
  control_sample_id = NULL,
  out_df = "tibble"
)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded using \code{read_npx()}, including
a sample ID column.}

\item{check_npx_log}{A list generated by \code{check_npx()} that includes:
\itemize{
\item \code{col_names$sample_id}: the column name in \code{df} identifying sample IDs.
}}

\item{control_sample_id}{A character vector of regex strings
(case-insensitive) used to match control sample IDs, e.g., c("control").
Default is: \code{NULL}.}

\item{out_df}{Output format of the returned data frame.
Options: \code{"tibble"} or \code{"arrow"}. Default is \code{"tibble"}.}
}
\value{
A filtered data frame with control sample IDs removed.
\dontrun{
# Example 1: use npx_data1 to check that clean_control_sample_id() works
log <- OlinkAnalyze::npx_data1 |>
  OlinkAnalyze::check_npx() |>
  suppressWarnings()

out <- OlinkAnalyze::clean_control_sample_id(npx_data1,
                                             check_npx_log = log,
                                             control_sample_id = "control")

# Example 2: check default setting of clean_control_sample_id(). Return data
unchanged.
out <- OlinkAnalyze::clean_control_sample_id(npx_data1,
                                             check_npx_log = log,
                                             control_sample_id = NULL)
}
}
\description{
This function removes rows from a NPX data where the \code{sample_id} matches a
regular expression pattern indicating customer control samples. This is useful
for filtering out technical controls or customer control samples based on their
SampleID.
}
\author{
Kang Dong
}
