% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_control_sample_id}
\alias{clean_control_sample_id}
\title{Help function removing a set of control samples from the dataset.}
\usage{
clean_control_sample_id(
  df,
  check_log,
  control_sample_ids = NULL,
  verbose = FALSE
)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{check_log}{A named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{\link[=check_npx]{check_npx()}} will be run internally using \code{df}.}

\item{control_sample_ids}{character vector of sample identifiers of control
samples. Default \code{NULL}, to mark no samples to be removed.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise messages.}
}
\description{
This function removes rows from NPX data where the sample identifiers, as
defined in \code{check_npx_log}, match samples provided in
\var{control_sample_ids}. Primary goal of the function is to serve for
filtering out technical replicates or control samples prior to downstream
analysis.
}
\examples{
\dontrun{
# use npx_data1 to check that clean_control_sample_id() works
log <- OlinkAnalyze::check_npx(
  df = OlinkAnalyze::npx_data1
) |>
  suppressWarnings() |>
  suppressMessages()

out <- OlinkAnalyze:::clean_control_sample_id(
  df = npx_data1,
  check_npx_log = log,
  control_sample_id = c("CONTROL_SAMPLE_AS 1", "CONTROL_SAMPLE_AS 2")
)
}

}
\author{
Kang Dong
Klev Diamanti
}
