% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_duplicate_sample_id}
\alias{clean_duplicate_sample_id}
\title{Remove samples with duplicate SampleIDs}
\usage{
clean_duplicate_sample_id(df, check_npx_log, out_df = "tibble")
}
\arguments{
\item{df}{A data frame loaded from \code{read_npx()}, containing a column
specified by \code{check_npx_log$col_names$sample_id}.}

\item{check_npx_log}{A list returned by \code{check_npx()}, containing:
\itemize{
\item \code{sample_id_dups}: a character vector of duplicate SampleIDs to be removed.
\item \code{col_names$sample_id}: the name of the column in \code{df} that holds
the SampleIDs.
}}

\item{out_df}{Output format of the cleaned data.
Options: \code{"tibble"} (default) or \code{"arrow"}.}
}
\value{
A filtered data frame with duplicated SampleIDs removed.
}
\description{
This function filters out rows from a data frame where the sample identifier
(e.g., \code{SampleID}) appears more than once. The duplicate SampleIDs are
identified in \code{check_npx_log$sample_id_dups}, which is generated by the
\code{check_npx()} function. Uses CLI messages to inform users of exclusions.
}
\examples{
# df <- read_npx("example.parquet")
# log <- check_npx(df)
# clean_df <- clean_duplicate_sample_id(df, log)
}
\author{
Kang Dong
}
