% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_duplicate_sample_id}
\alias{clean_duplicate_sample_id}
\title{Remove samples with duplicate SampleIDs}
\usage{
clean_duplicate_sample_id(df, check_npx_log, out_df = "tibble")
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded from \code{read_npx()}, containing a column
specified by \code{check_npx_log$col_names$sample_id}.}

\item{check_npx_log}{A list returned by \code{check_npx()}, containing:
\itemize{
\item \code{sample_id_dups}: a character vector of duplicate SampleIDs to be removed.
\item \code{col_names$sample_id}: the name of the column in \code{df} that holds
the SampleIDs.
}}

\item{out_df}{Output format of the cleaned data.
Options: \code{"tibble"} (default) or \code{"arrow"}.}
}
\value{
A filtered tibble or arrow object with duplicated SampleIDs removed.
}
\description{
This function filters out rows from a NPX data where the sample identifier
(e.g., \code{SampleID}) appears more than once. The duplicate SampleIDs are
identified in \code{check_npx_log$sample_id_dups}, which is generated by the
\code{check_npx()} function. Uses CLI messages to inform users of exclusions.
}
\examples{
\dontrun {
# Example 1: Exclude 1 sample with duplicate SampleID
df <- dplyr::tibble(
  SampleID = c("A", "B", "C", "C"),
  OlinkID = c(rep("OID12345", 2L), rep("OID12345", 2L)),
  SampleType = rep("SAMPLE", 4L),
  NPX = c(rep(1.0, 2L), rep(2.0, 2L)),
  PlateID = rep("plate1", 4L),
  QC_Warning = rep("Pass", 4L),
  LOD = rep(1.0, 4L)
)
log <- OlinkAnalyze::check_npx(df)
clean_df <- OlinkAnalyze::clean_duplicate_sample_id(df, log)
}
}
\author{
Kang Dong
}
