% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_duplicate_sample_id}
\alias{clean_duplicate_sample_id}
\title{Help function removing samples with duplicate identifiers.}
\usage{
clean_duplicate_sample_id(
  df,
  check_npx_log,
  out_df = "tibble",
  verbose = FALSE
)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded from \code{read_npx()}, including a
column identified by
"SampleID", "sampleid", or "sample_id".}

\item{check_npx_log}{A named list generated by the function \code{check_npx()},
containing:
\itemize{
\item \code{sample_id_dups}: a character vector of duplicated sample identifiers.
\item \code{col_names$sample_id}: the column name of the sample identifier in the
dataset.
}}

\item{out_df}{The class of the output dataset. One of
"tibble" and "arrow". (default = "tibble")}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise CLI
messages.}
}
\value{
A \code{tibble} or \code{arrow} object with rows corresponding to samples with
duplicated identifiers removed.
}
\description{
This function filters out rows from a \code{tibble} or \code{arrow} object where the
sample identifier (one of
"SampleID", "sampleid", or "sample_id")
matches values listed in \code{check_npx_log$sample_id_dups}, which identifies
samples with duplicated identifiers.
}
\author{
Kang Dong
}
