% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_helper.R
\name{clean_heatmap_df}
\alias{clean_heatmap_df}
\title{Title}
\usage{
clean_heatmap_df(df, check_log, colnames)
}
\arguments{
\item{df}{}

\item{check_log}{}

\item{colnames}{}
}
\description{
Title
}
\keyword{internal}
