% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_invalid_oid}
\alias{clean_invalid_oid}
\title{Help function removing assays with invalid identifiers.}
\usage{
clean_invalid_oid(df, check_log, remove_invalid_oid = TRUE, verbose = FALSE)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{check_log}{A named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{\link[=check_npx]{check_npx()}} will be run internally using \code{df}.}

\item{remove_invalid_oid}{Logical. If \code{FALSE}, skips filtering assays with
invalid identifiers. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise messages.}
}
\description{
This function filters out rows from a \code{tibble} or \code{arrow} object where the
assay identifier (one of
"OlinkID", "OID", "olinkid", "oid", or "olink_id")
matches values listed in \code{check_npx_log$oid_invalid}, which identifies
invalid or malformed assay identifiers.
}
\author{
Kang Dong
Klev Diamanti
}
