% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_invalid_oid}
\alias{clean_invalid_oid}
\title{Help function removing assays with invalid identifiers.}
\usage{
clean_invalid_oid(df, check_npx_log, out_df = "tibble", verbose = FALSE)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded from \code{read_npx()}, including a
column identified by
"OlinkID", "OID", "olinkid", "oid", or "olink_id".}

\item{check_npx_log}{A named list generated by the function \code{check_npx()},
containing:
\itemize{
\item \code{oid_invalid}: a character vector of invalid assay identifiers.
\item \code{col_names$olink_id}: the column name of the assay identifier in the
dataset.
}}

\item{out_df}{The class of the output dataset. One of
"tibble" and "arrow". (default = "tibble")}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise CLI
messages.}
}
\value{
A \code{tibble} or \code{arrow} object with rows corresponding to assays with
invalid identifiers removed.
}
\description{
This function filters out rows from a \code{tibble} or \code{arrow} object where the
assay identifier (one of
"OlinkID", "OID", "olinkid", "oid", or "olink_id")
matches values listed in \code{check_npx_log$oid_invalid}, which identifies
invalid or malformed assay identifiers.
}
\author{
Kang Dong
}
