% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_nonunique_uniprot}
\alias{clean_nonunique_uniprot}
\title{Help function converting types of columns to the expected ones.}
\usage{
clean_nonunique_uniprot(
  df,
  check_log,
  convert_nonunique_uniprot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{check_log}{A named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{\link[=check_npx]{check_npx()}} will be run internally using \code{df}.}

\item{convert_nonunique_uniprot}{Logical. If \code{FALSE}, retains non-unique
OlinkID - UniProt mapping. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise messages.}
}
\description{
This function checks the non-unique "OlinkID - UniProt" mappings, as defined
in \code{check_log}.  it selects the first UniProt ID per OlinkID and replaces
the original UniProt column with the unified mapping.
}
\author{
Kang Dong
Klev Diamanti
}
