% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_npx}
\alias{clean_npx}
\title{Clean Olink NPX Data Frame}
\usage{
clean_npx(
  df,
  check_log = NULL,
  preferred_names = NULL,
  keep_controls = NULL,
  control_sample_ids = NULL,
  out_df = "tibble",
  verbose = TRUE
)
}
\arguments{
\item{df}{A \code{tibble} or Arrow \code{Table} loaded from \code{read_npx()}.}

\item{check_log}{A list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{check_npx()}will be run internally using \code{df} and \code{preferred_names}.}

\item{preferred_names}{An optional named list to supply preferred column
names to \code{check_npx()}, if \code{check_log} is not provided.}

\item{keep_controls}{Character string indicating whether to retain any
controls:
\itemize{
\item \code{"sample"}: keep only sample-level controls,
\item \code{"assay"}: keep only assay-level controls,
\item \code{"both"}: keep all controls,
\item \code{NULL} (default): remove all controls.
}}

\item{control_sample_ids}{A character vector of \code{SampleID}s to remove.
Typically used to remove customer-provided or technical control samples.}

\item{out_df}{Output format: \code{"tibble"} (default) or \code{"arrow"}.}

\item{verbose}{Logical. If \code{TRUE} (default), prints step-wise CLI messages.}
}
\value{
A cleaned NPX dataset in either tibble or Arrow Table format,
depending on \code{out_df}.
}
\description{
This function applies a series of cleaning steps to an NPX data based on
the results from \code{\link[=check_npx]{check_npx()}}. It removes problematic samples and assays
to prepare a clean NPX data for downstream analysis, including duplicates
samples, controls samples, internal control assays, and samples or assays
with QC flag. Additionally, an instruction message is printed when user plan
to analyze absolute quantification value.

The cleaning pipeline performs the following steps:
\enumerate{
\item \strong{Remove invalid Olink IDs}: Assays flagged as having invalid Olink IDs.
\item \strong{Remove assays with all NA values}: Assays without quantifiable data.
\item \strong{Remove duplicate Sample IDs}: Ensures uniqueness across samples.
\item \strong{Remove control samples}:
\itemize{
\item Based on \code{SampleType} (e.g., \code{"SAMPLE_CONTROL"}, \code{"PLATE_CONTROL"},
\code{"NEGATIVE_CONTROL"}).
\item Based on \code{SampleID} matching known control sample IDs.
}
\item \strong{Remove samples failing QC}: Samples with QC status \code{'FAIL'}.
\item \strong{Remove internal control assays}: Based on \code{AssayType} (e.g.,
\code{"ext_ctrl"}, \code{"inc_ctrl"}, \code{"amp_ctrl"}).
\item \strong{Remove assays flagged with assay warnings}.
}

#' If the dataset includes absolute quantification, a message is shown
recommending log2 transformation for downstream analysis.
}
\author{
Kang Dong
}
