% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_qc_warning}
\alias{clean_qc_warning}
\title{Clean Samples That Failed QC}
\usage{
clean_qc_warning(df, check_npx_out_list)
}
\arguments{
\item{df}{A data.frame or tibble containing NPX data.}

\item{check_npx_out_list}{A list returned from \code{check_npx()}
containing standardized column names.}
}
\value{
A filtered data.frame or tibble with samples failing QC removed.
}
\description{
Removes samples from NPX data where the SampleQC column equals "FAIL"
(case-insensitive). Also logs a summary of all SampleQC statuses present
in the dataset.
}
\examples{
clean_data <- clean_qc_warning(npx_data, check_npx_out_list)

}
\author{
Kang Dong
}
