% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_qc_warning}
\alias{clean_qc_warning}
\title{Help function removing instances of samples that failed QC.}
\usage{
clean_qc_warning(df, check_log, remove_qc_warning = TRUE, verbose = FALSE)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{check_log}{A named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{\link[=check_npx]{check_npx()}} will be run internally using \code{df}.}

\item{remove_qc_warning}{Logical. If \code{FALSE}, retains samples flagged as
\code{FAIL} in QC warning. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise messages.}
}
\description{
This function uses the column marking QC warnings identified by
\code{check_npx_log} to remove samples flagged \code{FAIL} in the dataset.
}
\author{
Kang Dong
Klev Diamanti
}
