% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_qc_warning}
\alias{clean_qc_warning}
\title{Help function removing instances of samples that have failed QC}
\usage{
clean_qc_warning(df, check_npx_log, out_df = "tibble", verbose = FALSE)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded from \code{read_npx()}, including a
column identified by
"QC_Warning", "SampleQC", "qc_warning", or "sample_qc".}

\item{check_npx_log}{A named list generated by the function \code{check_npx()},
containing:
\itemize{
\item \code{col_names$sample_id}: the column name of the sample identifier in the
dataset.
\item \code{col_names$qc_warning}: the name of the column indicating QC status.
}}

\item{out_df}{The class of the output dataset. One of
"tibble" and "arrow". (default = "tibble")}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise CLI
messages.}
}
\value{
A \code{tibble} or \code{arrow} object with rows corresponding samples failed
QC removed.
}
\description{
This function uses the column marking QC warnings identified by
\code{check_npx_log} to samples flagged \code{FAIL} in the dataset.
}
\author{
Kang Dong
}
