% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_qc_warning}
\alias{clean_qc_warning}
\title{Help function cleaning Samples with QC Status 'FAIL'}
\usage{
clean_qc_warning(df, check_npx_log, out_df = "tibble")
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded using \code{read_npx()}, containing
a QC status column.}

\item{check_npx_log}{A list generated by \code{check_npx()} that includes:
\itemize{
\item \code{col_names$qc_warning}: the name of the column indicating QC status.
}}

\item{out_df}{A string specifying the output format of the data frame.
Options are \code{"tibble"} or \code{"arrow"}. Default is \code{"tibble"}.}
}
\value{
A cleaned tibble or arrow object with failed QC sample removed.
}
\description{
This function use the QC warning column identified by \code{check_npx_log} and
removes samples flagged \code{qc_warning} = "FAIL" from the NPX data. The function
returns the cleaned NPX data in the the desired output format (\code{"tibble"} or
\code{"arrow"}).
}
\author{
Kang Dong
}
