% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_sample_qc}
\alias{clean_sample_qc}
\title{Remove Samples That Failed QC}
\usage{
clean_sample_qc(df, check_npx_log, out_df = "tibble")
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded using \code{read_npx()}, containing
a QC status column.}

\item{check_npx_log}{A list generated by \code{check_npx()} that includes:
\itemize{
\item \code{col_names$qc_warning}: the name of the column indicating QC status.
}}

\item{out_df}{A string specifying the output format of the data frame.
Options are \code{"tibble"} or \code{"arrow"}. Default is \code{"tibble"}.}
}
\value{
A filtered tibble or arrow object with samples that
failed QC removed.
}
\description{
This function removes rows from an NPX data where the QC warning column
contains the term \code{"fail"} (case-insensitive). These rows typically represent
samples that failed quality control checks. The function uses
\code{stringr::str_detect()} for pattern matching and returns the cleaned data in
the desired output format (\code{"tibble"} or \code{"arrow"}).
}
\examples{
\dontrun{
# Example 1: Remove 1 sample that failed sample QC

df <- dplyr::tibble(
  SampleID = c("A", "B", "C", "D"),
  OlinkID = rep("OID12345", 4L),
  NPX = rep(1, 4L),
  PlateID = rep("plate1", 4L),
  QC_Warning = c(rep("Pass", 3L), "FAIL"),
  LOD = rep(1, 4L)
)

log <- OlinkAnalyze::check_npx(df)
clean_df <- OlinkAnalyze::clean_sample_qc(df, log)
}
}
\author{
Kang Dong
}
