% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_sample_type}
\alias{clean_sample_type}
\title{Help function removing control samples based on sample type.}
\usage{
clean_sample_type(
  df,
  check_npx_log,
  keep_control_sample = FALSE,
  out_df = "tibble",
  verbose = FALSE
)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded from \code{read_npx()}.}

\item{check_npx_log}{A named list generated by the function \code{check_npx()},
containing:
\itemize{
\item \code{col_names$sample_id}: the column name of the sample identifier in the
dataset.
\item \code{col_names$sample_type}: the column name of the sample type in the
dataset. \strong{If column is missing then function will return the original
dataset.}
}}

\item{keep_control_sample}{Logical. If \code{TRUE}, control samples are retained
and no filtering is attempted. Defaults to \code{FALSE}.}

\item{out_df}{The class of the output dataset. One of
"tibble" and "arrow". (default = "tibble")}

\item{verbose}{Logical. If \code{FALSE} (default), silences step-wise CLI
messages.}
}
\value{
A \code{tibble} or \code{arrow} object with rows corresponding to control
samples removed.
}
\description{
This function filters out rows from a dataset where the sample type column
matches known control sample types: \code{"SAMPLE_CONTROL"}, \code{"PLATE_CONTROL"} or
\code{"NEGATIVE_CONTROL"}. If \code{keep_control_sample} is set to \code{TRUE}, or if the
sample type column is present in the \code{check_npx_log}, the function returns
the original data unchanged.
}
\author{
Kang Dong
}
