% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_sample_type}
\alias{clean_sample_type}
\title{Remove control samples based on sample type}
\usage{
clean_sample_type(
  df,
  check_npx_log,
  control_sample_types = c("SAMPLE_CONTROL", "PLATE_CONTROL", "NEGATIVE_CONTROL"),
  out_df = "tibble"
)
}
\arguments{
\item{df}{A data frame loaded from \code{read_npx()}.}

\item{check_npx_log}{A list returned from \code{check_npx()}, containing:
\itemize{
\item \code{col_names$sample_type}: name of the column in \code{df} that identifies
sample type.
}}

\item{control_sample_types}{A character vector of control sample types to
exclude. Regular expression matching is used.Default is c("SAMPLE_CONTROL",
"PLATE_CONTROL", "NEGATIVE_CONTROL").}

\item{out_df}{Output format: either "tibble" (default) or "arrow".}
}
\value{
A cleaned data frame with control sample removed.
}
\description{
This function filters out rows from an NPX data frame where the sample type
column matches one of the defined control types using regular expressions
(case-insensitive).
}
\examples{
clean_df <- clean_sample_type(df, check_npx_log)
}
\author{
Kang Dong
}
