% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_npx.R
\name{clean_sample_type}
\alias{clean_sample_type}
\title{Remove control samples based on sample type}
\usage{
clean_sample_type(
  df,
  check_npx_log,
  control_sample_types = c("SAMPLE_CONTROL", "PLATE_CONTROL", "NEGATIVE_CONTROL"),
  out_df = "tibble"
)
}
\arguments{
\item{df}{A \code{tibble} or \code{arrow} object loaded from \code{read_npx()}.}

\item{check_npx_log}{A list returned from \code{check_npx()}, containing:
\itemize{
\item \code{col_names$sample_type}: name of the column in \code{df} that identifies
sample type.
}}

\item{control_sample_types}{A character vector of control sample types to
exclude. Regular expression matching is used. Default is c("SAMPLE_CONTROL",
"PLATE_CONTROL", "NEGATIVE_CONTROL").}

\item{out_df}{Output format: either "tibble" (default) or "arrow".}
}
\value{
A cleaned tibble or arrow object with control sample removed.
}
\description{
This function filters out rows from an NPX data where the sample type
column matches one of the defined control types using regular expressions
(case-insensitive).
}
\examples{
\dontrun {
# Example 1: Run clean_sample_type() with default setting to remove
control samples
df <- dplyr::tibble(
  SampleID = c("A", "B", "C", "D"),
  OlinkID = rep("OID12345", 4L),
  SampleType = c("SAMPLE",
                 "SAMPLE_CONTROL",
                 "PLATE_CONTROL",
                 "NEGATIVE_CONTROL"),
  NPX = rep(1, 4L),
  PlateID = rep("plate1", 4L),
  QC_Warning = rep("Pass", 4L),
  LOD = rep(1, 4L)
)
log <- OlinkAnalyze::check_npx(df)
clean_df <- OlinkAnalyze::clean_sample_type(df, log)

# Example 2: Run control_sample_types = NULL to keep control samples, and
return the unchanged data
clean_df <- OlinkAnalyze::clean_sample_type(df, log, control_sample_types = NULL)
}
}
\author{
Kang Dong
}
