% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_helper.R
\name{color_heatmap}
\alias{color_heatmap}
\title{add colors to pheatmap arguments}
\usage{
color_heatmap(
  df,
  check_log,
  pheatmap_args,
  variable_row_list,
  variable_col_list
)
}
\arguments{
\item{df}{Data frame in long format with SampleID, NPX, OlinkID, Assay and
columns of choice for annotations.}

\item{check_log}{output from check_npx on \code{df}}

\item{variable_row_list}{Columns in \code{df} to be annotated for rows in
the heatmap.}

\item{variable_col_list}{Columns in \code{df} to be annotated for columns in
the heatmap.}
}
\value{
updated pheatmap_args
}
\description{
add colors to pheatmap arguments
}
\keyword{internal}
