% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_helper.R
\name{create_pheatmap_args}
\alias{create_pheatmap_args}
\title{create list of arguments to pass to pheatmap function}
\usage{
create_pheatmap_args(
  df_wide,
  df,
  check_log,
  center_scale,
  cluster_rows,
  cluster_cols,
  na_col,
  show_rownames,
  show_colnames,
  annotation_legend,
  fontsize,
  variable_row_list,
  variable_col_list,
  colnames,
  ...
)
}
\arguments{
\item{df_wide}{wide version of df}

\item{df}{Data frame in long format with SampleID, NPX, OlinkID, Assay and
columns of choice for annotations.}

\item{check_log}{output from check_npx on \code{df}}

\item{center_scale}{Logical. If data should be centered and scaled across
assays (default \code{TRUE}).}

\item{cluster_rows}{Logical. Determining if rows should be clustered
(default \code{TRUE}).}

\item{cluster_cols}{Logical. Determining if columns should be clustered
(default \code{TRUE}).}

\item{na_col}{Color of cells with \code{NA} (default black)}

\item{show_rownames}{Logical. Determining if row names are shown
(default \code{TRUE}).}

\item{show_colnames}{Logical. Determining if column names are shown
(default \code{TRUE}).}

\item{annotation_legend}{Logical. Determining if legend for annotations
should be shown (default \code{TRUE}).}

\item{fontsize}{Fontsize (default 10)}

\item{variable_row_list}{Columns in \code{df} to be annotated for rows in
the heatmap.}

\item{variable_col_list}{Columns in \code{df} to be annotated for columns in
the heatmap.}

\item{colnames}{Character. Determines how to label the columns.
Must be 'assay', 'oid', or 'both' (default 'both').}

\item{...}{Additional arguments used in \code{pheatmap::pheatmap}}
}
\value{
list of arguments for pheatmap
}
\description{
create list of arguments to pass to pheatmap function
}
\keyword{internal}
