% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_params}
\alias{.check_params}
\title{Common parameters for check functions.}
\usage{
.check_params(error)
}
\arguments{
\item{error}{Scalar boolean to return an error instead of a \code{FALSE}
(\code{default = FALSE}).}
}
\description{
Common parameters for check functions.
}
\author{
Klev Diamanti
}
