% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.downstream_fun_args}
\alias{.downstream_fun_args}
\title{Common parameters for downstream anaysis functions.}
\usage{
.downstream_fun_args(df)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}
}
\description{
Common parameters for downstream anaysis functions.
}
\author{
Klev Diamanti
}
