% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_helper.R
\name{extract_ellipsis_arg}
\alias{extract_ellipsis_arg}
\title{extract ellipsis arguments and add to pheatmap arguments}
\usage{
extract_ellipsis_arg(pheatmap_args, ...)
}
\arguments{
\item{pheatmap_args}{pheatmap argument list}

\item{...}{additional arguments to be passed to pheatmap function}
}
\value{
updated pheatmap arguments list with ellipsis variables
}
\description{
extract ellipsis arguments and add to pheatmap arguments
}
\keyword{internal}
