% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_others.R
\name{get_alt_colnames}
\alias{get_alt_colnames}
\title{Gets alternative column names for a given column key.}
\usage{
get_alt_colnames(col_key)
}
\arguments{
\item{col_key}{Column key for which to retrieve alternative column names. One
of "sample_id", "sample_type", "assay_type", "olink_id", "uniprot", "assay", "panel", "block", "plate_id", "panel_version", "lod", "quant", "ext_npx", "count", "qc_warning", "assay_warn", "normalization", and "qc_version".}
}
\value{
A character vector of alternative column names corresponding to the
provided column key.
}
\description{
Gets alternative column names for a given column key.
}
\examples{
\donttest{
  OlinkAnalyze:::get_alt_colnames(col_key = "sample_id")
}
}
\keyword{internal}
