% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_delim.R
\name{get_field_separator}
\alias{get_field_separator}
\title{Help function to get the separator of a delimited file from Olink software.}
\usage{
get_field_separator(file)
}
\arguments{
\item{file}{Path to Olink software output delimited file in wide format.
Expecting file extensions
"csv" and "txt".}
}
\value{
The file delimiter
";", ",", "	", "|", or ":".
}
\description{
This function uses the first line of the provided file to determine the
separator of the file.

\strong{Note:} The function does not allow presence of commas and semicolons
on the same line.
}
\seealso{
\code{\link{read_npx_delim}}
}
\author{
Klev Diamanti
}
