% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_zip.R
\name{get_npx_file}
\alias{get_npx_file}
\title{Help function to get the name of the NPX file from a compresseed NPX.}
\usage{
get_npx_file(files, compressed_file_ext = c("zip"))
}
\arguments{
\item{files}{A vector of file names without any path prefix.}

\item{compressed_file_ext}{Character vector of file extensions for
compressed files.}
}
\value{
The name of the NPX file.
}
\description{
Help function to get the name of the NPX file from a compresseed NPX.
}
\seealso{
[check_checksum()]
  [get_checksum_file()]
}
\author{
Klev Diamanti
}
\keyword{NPX}
\keyword{zip}
