% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{get_olink_broader_platforms}
\alias{get_olink_broader_platforms}
\title{Get names of current broader Olink platforms.}
\usage{
get_olink_broader_platforms()
}
\value{
A character vector with names of the broader Olink platforms.
}
\description{
Get names of current broader Olink platforms.
}
