% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{get_olink_data_types}
\alias{get_olink_data_types}
\title{Get names of selected Olink quantification methods (data types).}
\usage{
get_olink_data_types(
  broad_platform = NULL,
  platform_name = NULL,
  quant_type = NULL
)
}
\arguments{
\item{broad_platform}{Name of the \var{broad_platform} to filter for. If
\code{NULL} all broader platforms are considered, otherwise one of
"NGS" and "qPCR" is expected.}

\item{platform_name}{Name of the \var{platform_name} to filter for. If \code{NULL}
all platforms are considered, otherwise one of
"Explore 3072", "Explore HT", "Flex", "Focus", "Reveal", "Target 48", and "Target 96" is expected.}

\item{quant_type}{Name of the \var{quant_type} to filter for. If \code{NULL} all
quantification types are considered, otherwise one of
"absolute" and "relative" is expected.}
}
\value{
A character vector with names of Olink quantification methods (data
types) filtered by \var{broad_platform}, \var{platform_name}, and
\var{quant_type}.
}
\description{
Get names of selected Olink quantification methods (data types).
}
\author{
Klev Diamanti
}
