% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_olink_platforms.R
\name{get_olink_data_types}
\alias{get_olink_data_types}
\title{Get names of current Olink quantification methods.}
\usage{
get_olink_data_types(broader_platform = NULL)
}
\arguments{
\item{broader_platform}{Name of the \var{broader_platform} to filter the
Olink platforms for. If set to \code{NULL}, all Olink platforms are returned.}
}
\value{
A character vector with names of Olink quantification methods.
}
\description{
Get names of current Olink quantification methods.
}
