% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_helper.R
\name{lst_styling}
\alias{lst_styling}
\title{Function to edit grob}
\usage{
lst_styling(i, x, styling_location, styling)
}
\arguments{
\item{i}{index of attribute}

\item{x}{plot}

\item{styling_location}{list of locations for different attributes}

\item{styling}{grob altered parameters for theme}
}
\value{
updated plot with updated grob
}
\description{
Function to edit grob
}
\keyword{internal}
