% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{make_cyt_graph}
\alias{make_cyt_graph}
\title{send topGO GO terms to cytoscape}
\usage{
make_cyt_graph(
  pvals,
  go_annot_list,
  test_stat = "ks",
  graph_name = "myGraph",
  firstSigNodes = 10
)
}
\description{
can be useful to visualise the topGO terms in a network - so here's a function
for exporting results to cytoscape
}
\details{
there's actually a built in function for creating a plot based on this network,
but it was not suitable as there were so many terms
}
