% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npx_data1}
\alias{npx_data1}
\title{NPX Data in Long format}
\format{
In addition to standard read_NPX() columns, this dataset also contains columns:
\describe{
\item{Subject}{Subject Identifier}
\item{Treatment}{ Treated or Untreated}
\item{Site}{Site indicator, 5 unique values}
\item{Time}{Baseline, Week.6 and Week.12}
\item{Project}{Project ID number}
}
}
\usage{
npx_data1
}
\description{
Data is generated randomly to demonstrate use of functions in this package.
}
\details{
A tibble with 29,440 rows and 17 columns.
Dataset npx_data1 is an Olink NPX data file (tibble) in long format with 158 unique Sample ID's
(including 2 repeats each of control samples: CONTROL_SAMPLE_AS 1 CONTROL_SAMPLE_AS 2).
The data also contains 1104 assays (uniquely identified using OlinkID) over 2 Panels.
}
\keyword{datasets}
