% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_anova.R
\name{olink_anova}
\alias{olink_anova}
\title{Function which performs an ANOVA per protein}
\usage{
olink_anova(df, variable)
}
\arguments{
\item{df}{NPX data frame in long format with atleast protein name (Assay), OlinkID, UniProt and a factor with at least 3 levels.}

\item{variable}{Character value indicating which column should be used as the grouping variable. Needs have  3 or more levels}
}
\value{
A data frame containing the ANOVA results for every protein.
}
\description{
Performs an ANOVA test on every protein (by OlinkID) for a given grouping variable
}
\examples{
\donttest{anova_results <- olink_anova(df = npx.data, variable = 'Group')}
}
