% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_anova.R
\name{olink_anova_posthoc}
\alias{olink_anova_posthoc}
\title{Function which performs an ANOVA posthoc test per protein}
\usage{
olink_anova_posthoc(df, variable, olinkid_list)
}
\arguments{
\item{df}{tibble/data frame in long format with atleast protein name (Assay), OlinkID, and a variable with at least 3 levels.}

\item{variable}{Variable to performed the ANOVA on.}

\item{olinkid_list}{Vector of significant proteins (OlinkID)}
}
\description{
This function allows you to perform an ANOVA posthoc test per protein (by OlinkID).
}
\examples{
\donttest{anova_results <- olink_anova(df, "Group")
significant_assays <- anova_results \%>\%
filter(Threshold == 'Significant') \%>\%
pull(OlinkID)
anova_posthoc_results <- olink_anova_posthoc(df, "Group", significant_assay)}
}
