% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_bridge_selector.R
\name{olink_bridge_selector}
\alias{olink_bridge_selector}
\alias{olink_bridgeselector}
\alias{\itemize{}
\alias{\item{SampleID:}}
\alias{Sample}
\alias{ID}
\alias{\item{perc_assays_below_lod:}}
\alias{Percent}
\alias{of}
\alias{Assays}
\alias{that}
\alias{are}
\alias{below}
\alias{LOD}
\alias{for}
\alias{the}
\alias{sample}
\alias{\item{MeanNPX:}}
\alias{Mean}
\alias{NPX}
\alias{}}
\title{Bridge selection function}
\usage{
olink_bridge_selector(df, sample_missing_freq, n)

olink_bridgeselector(df, ..., n)
}
\arguments{
\item{df}{Tibble/data frame in long format such as produced by the
Olink Analyze read_npx function.}

\item{sample_missing_freq}{The threshold for sample wise missingness.}

\item{n}{Number of bridge samples to be selected.}
}
\value{
A "tibble" with sample IDs and mean NPX for a defined number of
bridging samples. Columns include:
}
\description{
The bridge selection function will select a number of bridge samples based
on the input data. It selects samples with good detection that pass QC
and cover a good range of the data. If possible, Olink recommends 8-16
bridge samples. When running the selector, Olink recommends starting at
sample_missing_freq = 0.10 which represents a maximum of 10\\% data below LOD
per sample. If there are not enough samples output, increase to 20\\%. \cr\cr
The function accepts NPX Excel files with data < LOD replaced.
}
\details{
\code{olink_bridgeselector()} is a synonym of \code{olink_bridge_selector()} .
}
\examples{
\donttest{bridge_samples <- olink_bridge_selector(npx_data1,
sample_missing_freq = 0.1, n = 20)}

}
