% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_bridgeselector.R
\name{olink_bridgeselector}
\alias{olink_bridgeselector}
\title{Bridge selection function}
\usage{
olink_bridgeselector(df, sampleMissingFreq, n)
}
\arguments{
\item{df}{Tibble/data frame in long format such as produced by the olinkr read_NPX function.}

\item{sampleMissingFreq}{The threshhold for sample wise missingness.}

\item{n}{Number of bridge samples to be selected.}
}
\value{
Tibble with sample ID:s and mean NPX for a defined number of bridging samples.
}
\description{
The bridge selection function will select a number of bridge samples based on the input data. It select samples with
good detection, which passes QC and cover a good range of the data. If possible olink recommends 8-16 bridge samples.
Bridge sample selection strategy: Output 2x (number of bridging samples) for final selection (by customer).
When running the selector, start at 10\% missingness. If there are not enough samples output, increase to 20\%.
If still not enough, consider outputting e.g. 1.5x (number of bridging samples) for final selection. \cr\cr
The function accepts NPX Excel files with data < LOD replaced.
}
\examples{
\donttest{bridge_samples <- olink_bridgeselector(npx_data1, sampleMissingFreq = 0.1, n = 20)}
}
