% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_class.R
\name{olink_check_log}
\alias{olink_check_log}
\title{Retrieve the check log from an Olink data object}
\usage{
olink_check_log(x)
}
\arguments{
\item{x}{A data object, typically from \code{\link[=read_npx]{read_npx()}} or \code{\link[=clean_npx]{clean_npx()}}.}
}
\value{
A named list (the check log) or \code{NULL} if no check log is attached.
}
\description{
Extracts the check log from an \code{olink_class} tibble or an ArrowObject with
embedded metadata. Returns \code{NULL} if no check log is found.
}
\examples{
\dontrun{
# read and check data
npx_file <- system.file("extdata",
                        "npx_data_ext.parquet",
                        package = "OlinkAnalyze")
npx_df <- read_npx(filename = npx_file)
check_log <- check_npx(df = npx_df)

# attach check_log to the data
npx_obj <- new_olink_class(data = npx_df, check_log = check_log)

# retrieve check_log
olink_check_log(npx_obj)
}

}
