% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_color.R
\name{olink_fill_discrete}
\alias{olink_fill_discrete}
\title{Olink fill scale for discrete ggplots}
\usage{
olink_fill_discrete(..., alpha = 1, coloroption = NULL)
}
\arguments{
\item{...}{Optional. Additional arguments to pass to ggplot2::discrete_scale()}

\item{alpha}{transparency (optional)}

\item{coloroption}{string, one or more of the following:
c('red', 'orange', 'yellow', 'green', 'teal', 'turqoise', 'lightblue', 'darkblue', 'purple', 'pink')}
}
\value{
No return value, called for side effects
}
\description{
Olink fill scale for discrete ggplots
}
\examples{

library(ggplot2)

dsub <- subset(diamonds, x > 5 & x < 6 & y > 5 & y < 6)
dsub$diff <- with(dsub, sqrt(abs(x - y)) * sign(x - y))

ggplot(dsub, aes(x, y, colour = diff)) +
  geom_point() +
  theme_bw() +
  olink_fill_discrete()

}
