% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mixed_model.R
\name{olink_lmer_plot}
\alias{olink_lmer_plot}
\title{Function which performs a linear mixed model plot per significant protein}
\usage{
olink_lmer_plot(df, var1, var2, ID, olinkid_list,
  number_of_proteins_per_plot = 6, ...)
}
\arguments{
\item{df}{NPX data frame in long format with at least protein name, OlinkID, UniProt, 1-2 variables with at least 2 levels and subject ID.}

\item{var1}{Character value indicating the name of the column to be used as the first variable in the LME model}

\item{var2}{SCharacter value indicating the name of the column to be used as the second variable in the LME model}

\item{ID}{Character value indicating the column of subject ID's. To be used as a random effect in the LME model}

\item{olinkid_list}{Character vector indicating which proteins (by OlinkID) to create figures for}

\item{number_of_proteins_per_plot}{Number plots to include in the array of point-range plots. Defaults to 6 plots per figure}
}
\value{
A list of objects of class "ggplot"
}
\description{
Creates an array of point-range plots for a given list of proteins (by OlinkID).
}
\examples{
\donttest{lme_results <- olink_lmer_main(df=npx.data, var1 = "Visit", var2 = "Treatment", ID = "SubjectID")
significant_proteins <- lme_results \%>\%
 filter(Threshold == 'Significant') \%>\%
 pull(OlinkID)
olink_lmer_plot(df=npx.data,  var1 = "Visit", var2 = "Treatment", ID = "Subject", olinkid_list = significant_proteins}
}
