% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_lod.R
\name{olink_lod}
\alias{olink_lod}
\title{Calculate LOD using Negative Controls or Fixed LOD}
\usage{
olink_lod(data, lod_file_path = NA, lod_method = "NCLOD")
}
\arguments{
\item{data}{npx data file}

\item{lod_file_path}{location of lod file from Olink (only needed if lod_method = "FixedLOD")}

\item{lod_method}{method for calculating LOD using either "FixedLOD" or negative controls ("NCLOD")}
}
\value{
A dataframe with 2 additional columns, LOD and PCNormalizedLOD. When Normalization = "Plate Control", LOD and PCNormalizedLOD are equivalent
}
\description{
Calculate LOD using Negative Controls or Fixed LOD
}
