% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_utils.R
\name{olink_norm_product_id}
\alias{olink_norm_product_id}
\title{Identify names of product for each project}
\usage{
olink_norm_product_id(lst_df, lst_cols)
}
\arguments{
\item{lst_df}{Named list of datasets to be normalized.}

\item{lst_cols}{Named list of vectors with the required column names for each
dataset in \var{lst_df}.}
}
\value{
Named character vector with the Olink product name that  each input
datatset matches to.
}
\description{
Identify names of product for each project
}
\author{
Kathy Nevola
Klev Diamanti
}
