% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_normalization.R
\name{olink_normalization}
\alias{olink_normalization}
\title{Function to normalize two datasets together}
\usage{
olink_normalization(df1, df2, overlapping_sample, df1_project_nr = "P1",
  df2_project_nr = "P2", centering = T)
}
\arguments{
\item{df1}{First dataframe to be used in normalization}

\item{df2}{Second dataframe to be used in normalization}

\item{overlapping_sample}{Character vector of sample ID's to be used for bridge normalization. Sample ID's must be present in both df1 and df2. If not provided, intensity normalization will be performed.}

\item{df1_project_nr}{Project name of first dataset.}

\item{df2_project_nr}{Project name of second dataset.}

\item{centring}{If TRUE, medians are adjusted to the centerpoint between the median for df1 and the median for df2. If FALSE, df2 is adjusted to have the same median as df1 (default TRUE)}
}
\value{
A data frame of NPX data in long format containing both df1 and df2 and normalized NPX values.
}
\description{
Median normalizes all proteins (by OlinkID) in two dataframes. If using bridging samples, adjustment is made using the median of the paired differences between the two data frames. If intensity normalizing, adjustment is done using the difference of the medians.
}
\examples{
\donttest{
data_1 <- read_NPX("~/NPX data1.xlsx")
data_2 <- read_NPX("~/NPX data2.xlsx")
olink_normalization(data_1,data_2,df1_project_nr='Project 1', df2_project_nr='Project 2',centering = T)}
}
\keyword{Normalization}
