% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_normalization.R
\name{olink_normalization_bridge}
\alias{olink_normalization_bridge}
\title{Normalization of all proteins using reference samples}
\usage{
olink_normalization_bridge(
  df1,
  df2,
  overlapping_samples,
  df1_project_nr = "P1",
  df2_project_nr = "P2",
  reference_project = "P1"
)
}
\arguments{
\item{df1}{First dataframe to be used in normalization}

\item{df2}{Second dataframe to be used in normalization}

\item{overlapping_samples}{Samples to be used for adjustment factor calculations. These samples have to be present in both df1 and df2.}

\item{df1_project_nr}{Project name of first dataset.}

\item{df2_project_nr}{Project name of second dataset.}

\item{reference_project}{Project name of reference_project. Needs to be the same as either df1_project_nr or df2_project_nr. The project to which the second project is adjusted to.}
}
\value{
A tibble of NPX data in long format containing normalized NPX values, including adjustment factors.
}
\description{
Normalizes two NPX data frames using reference samples. This normalization method is referred to as bridge normalization, or bridging for short.
}
\details{
df1 is adjusted to df2 using overlapping samples (bridge samples).
Each protein is adjusted using the median of the paired differences between the bridge samples in the two data frames.
The overlapping samples need to be named the same in both dataframes, and the dataframe being adjusted to is specified in the input reference_project.
By default, df1 will be used as reference.
}
\references{
https://www.olink.com/application/data-normalization-and-standardization/
}
\keyword{Normalization}
