% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_quantile_smoothing.R
\name{olink_normalization_qs}
\alias{olink_normalization_qs}
\title{Quantile smoothing normalization of all proteins between two NPX projects.}
\usage{
olink_normalization_qs(
  exploreht_df,
  explore3072_df,
  bridge_samples,
  exploreht_name = "reference",
  explore3072_name = "new"
)
}
\arguments{
\item{exploreht_df}{Data frame of the first project, must be Explore HT
data (required). (default: reference)}

\item{explore3072_df}{Data frame of the second project, must be Explore
3072 data (required). (default : new)}

\item{bridge_samples}{Named list of 2 arrays containing, SampleID of shared
samples to be used for the calculation of adjustment factor. The
names of the two arrays should be DF_ht and DF_3k corresponding to projects
HT and 3072, respectively. Arrays should be of equal length and index of each
entry should correspond to the same sample. (required)}

\item{exploreht_name}{Name of Explore HT project (default: reference)}

\item{explore3072_name}{Name of Explore 3072  project (default: new)}
}
\value{
A "tibble" of NPX data in long format containing qs normalized NPX
values, including adjustment factors and name of project.
}
\description{
Normalizes two NPX projects (data frames) using shared samples.\cr\cr
}
\author{
Amrita Kar

Marianne Sandin

Masoumeh Sheikhi
}
\keyword{;}
\keyword{Normalization;}
\keyword{Quantile}
\keyword{Smoothing}
