% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_normalization_quantile_smoothing.R
\name{olink_normalization_qs}
\alias{olink_normalization_qs}
\title{Quantile smoothing normalization of all proteins between two NPX projects.}
\usage{
olink_normalization_qs(lst_df, ref_cols, bridge_samples)
}
\arguments{
\item{lst_df}{A named list of the 2 input datasets. First element should be
the reference dataset from Olink Explore HT and the second element should
originate from Olink Explore 3072. (required)}

\item{ref_cols}{A named list with the column names to use. Exported from
olink_norm_input_check. (required)}

\item{bridge_samples}{Character vector of samples to be used for the
quantile mapping. (required)}
}
\value{
A "tibble" of Olink data in long format containing both input
datatsets with the quantile normalized quantifications.
}
\description{
This function uses bridge samples to map quantiles of the non-reference
dataset to the ones of the reference dataset. Mapped quantiles are used to
transform the quantifications of the the non-reference dataset to the
reference.
}
\details{
Individual Olink NPX projects are generally normalized using
either plate control normalization or intensity normalization methods.
Since NPX is a relative measurement, in the case when a study is
separated into multiple projects, an additional normalization step is
needed to allow the data to be comparable across projects.
The \code{olink_normalization_qs()} performs quantile smoothing bridging
normalization between datasets from two Olink products (for example
Olink Explore 3072 and Olink Explore HT).
}
\examples{
\dontrun{
\donttest{\
# Bridge samples
bridge_samples <- intersect(
x = unique(data_ht$SampleID),
y = unique(data_3k$SampleID)) |>
  (\(x) x[!grepl("CONTROL", x)])()

# Run the internal function olink_norm_input_check()
ref_col <- olink_norm_input_check (
df1 = data_ht, df2 = data_3k,
overlapping_samples_df1 = bridge_samples,
overlapping_samples_df2 = NULL,
df1_project_nr = "P1",
df2_project_nr = "P2",
reference_project = "P1",
reference_medians = NULL
)

# Named list of input datasets
lst_df <- list (
P1 = ref_col$ref_df,
P2 = ref_col$not_ref_df
)

ref_cols <- ref_col$ref_cols

qs_result <- olink_normalization_qs (
lst_df = lst_df,
ref_cols = ref_cols,
bridge_samples = bridge_samples
)
}
}


}
\author{
Amrita Kar
Marianne Sandin
Masoumeh Sheikhi
Klev Diamanti
}
\keyword{;}
\keyword{Normalization;}
\keyword{Quantile}
\keyword{Smoothing}
