% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osi_dist_plot.R
\name{olink_osi_dist_plot}
\alias{olink_osi_dist_plot}
\title{Title}
\usage{
olink_osi_dist_plot(df, check_log = NULL, osi_score = NULL)
}
\arguments{
\item{df}{data frame with OSI data present}

\item{check_log}{check log from check NPX}

\item{osi_score}{OSI column to graph, one of OSISummary,
OSITimeToCentrifugation, or OSIPreparationTemperature}
}
\value{
distribution plot (histogram overlayed with density plot) of
osi values for corresponding osi_score column
}
\description{
Title
}
