% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_Pathway_Enrichment.R
\name{olink_pathway_enrichment}
\alias{olink_pathway_enrichment}
\title{Performs pathway enrichment using over-representation analysis (ORA) or gene set enrichment analysis (GSEA) based on statistical test results and full data using clusterProfiler's gsea and enrich functions for various databases.}
\usage{
olink_pathway_enrichment(
  data,
  test_results,
  method = "GSEA",
  ontology = "MSigDb",
  organism = "human"
)
}
\arguments{
\item{data}{NPX data frame in long format with at least protein name (Assay), OlinkID, UniProt,SampleID, QC_Warning, NPX, and LOD}

\item{test_results}{a dataframe of statistical test results including Adjusted_pval and estimate columns. Currently only works for tests with estimates for all assays.}

\item{method}{Either "GSEA" (default) or "ORA"}

\item{ontology}{Supports "MSigDb" (default), "KEGG", "GO", and "Reactome" as arguements. MSigDb contains C2 and C5 genesets. C2 and C5 encompass KEGG, GO, and Reactome.}

\item{organism}{Either "human" (default) or "mouse"}
}
\value{
A data frame of enriched pathway/ontological terms including pvalue, p.adjust, and qvalues, and list of Entrez IDs in that pathway and the data
}
\description{
Performs pathway enrichment using over-representation analysis (ORA) or gene set enrichment analysis (GSEA) based on statistical test results and full data using clusterProfiler's gsea and enrich functions for various databases.
}
\examples{
\donttest{
library(dplyr)
npx_df <- npx_data1 \%>\% filter(!grepl("control", SampleID, ignore.case = TRUE))
ttest_results <- olink_ttest(
  df = npx_df,
  variable = "Treatment",
  alternative = "two.sided"
)
gsea_results <- olink_pathway_enrichment(data = npx_data1, test_results = ttest_results)
ora_results <- olink_pathway_enrichment(
  data = npx_data1,
  test_results = ttest_results, method = "ORA"
)
}
}
\seealso{
\itemize{
\item{\code{\link[OlinkAnalyze:olink_pathway_heatmap]{olink_pathway_heatmap}} for generating a heat map of results}
\item{\code{\link[OlinkAnalyze:olink_pathway_visualization]{olink_pathway_visualization}} for generating a bar graph of results}
}
}
