% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_Pathway_Visualization.R
\name{olink_Pathway_Visualization}
\alias{olink_Pathway_Visualization}
\title{Creates bargraph of top/selected enrichment terms from GSEA or ORA results from olink_Pathway_Enrichment()}
\usage{
olink_Pathway_Visualization(
  enrich_results,
  method = "GSEA",
  keyword = NULL,
  number_of_terms = 20
)
}
\arguments{
\item{enrich_results}{data frame of enrichment results from olink_Pathway_Enrichment()}

\item{method}{method used in olink_Pathway_Enrichment ("GSEA" (default) or "ORA")}

\item{keyword}{(optional) keyword to filter enrichment results on, if not specified, displays top terms}

\item{number_of_terms}{number of terms to display, default is 20}
}
\value{
bargraph as a ggplot object
}
\description{
Creates bargraph of top/selected enrichment terms from GSEA or ORA results from olink_Pathway_Enrichment()
}
\examples{
\donttest{
library(dplyr)
npx_df <- npx_data1 \%>\% filter(!grepl('control',SampleID, ignore.case = TRUE))
ttest_results <- olink_ttest(df=npx_df,
                             variable = 'Treatment',
                             alternative = 'two.sided')
gsea_results <- olink_Pathway_Enrichment(data = npx_data1, test_results = ttest_results)
ora_results <- olink_Pathway_Enrichment(data = npx_data1,
                                      test_results = ttest_results, method = "ORA")
olink_Pathway_Visualization(enrich_results = gsea_results)
olink_Pathway_Visualization(enrich_results = gsea_results, keyword = "immune")
olink_Pathway_Visualization(enrich_results = ora_results, method = "ORA", number_of_terms = 15)

}

}
\seealso{
\itemize{
\item{\code{\link[OlinkAnalyze:olink_Pathway_Enrichment]{olink_Pathway_Enrichment}} for generating enrichment results}
\item{\code{\link[OlinkAnalyze:olink_Pathway_Heatmap]{olink_Pathway_Heatmap}} for generating a heat map of results}
}
}
