% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{olink_pca_plot}
\alias{olink_pca_plot}
\title{Function to plot a PCA of the data}
\usage{
olink_pca_plot(
  df,
  color_g = "QC_Warning",
  x_val = 1,
  y_val = 2,
  label_samples = F,
  drop_assays = F,
  drop_samples = F,
  n_loadings = 0,
  loadings_list = NULL,
  verbose = T
)
}
\arguments{
\item{df}{data frame in long format with Sample Id, NPX and column of choice for colors}

\item{color_g}{Character value indicating which column to use for colors (default QC_Warning)}

\item{x_val}{Integer indicating which principal component to plot along the x-axis (default 1)}

\item{y_val}{Integer indicating which principal component to plot along the y-axis (default 2)}

\item{label_samples}{Logical. If TRUE, points are replaced with SampleID (default FALSE)}

\item{drop_assays}{Logical. All assays with any missing values will be dropped. Takes precedence over sample drop.}

\item{drop_samples}{Logical. All samples with any missing values will be dropped.}

\item{n_loadings}{Integer. Will plot the top n_loadings based on size.}

\item{loadings_list}{Character vector indicating for which OlinkID's to plot as loadings. It is possible to use n_loadings and loadings_list simultaneously.}

\item{verbose}{Logical. Whether warnings about the number of samples and/or assays dropped or imputed should be printed to the console.}
}
\value{
An object of class "ggplot"
}
\description{
Generates a PCA projection of all samples from NPX data along two principal components (default PC2 vs. PC1) including the explained variance and dots colored by QC_Warning using stats::prcomp and ggplot2::ggplot. 
The values are by default scaled and centered in the PCA and proteins with missing NPX values are by default removed from the corresponding assay. 
Unique sample names are required. 
Imputation by the median is done for assays with missingness <10\% for multi-plate projects and <5\% for single plate projects.
}
\examples{
\donttest{olink_pca_plot(df=npx.data, color_g = "QC_Warning")  }
}
\keyword{NPX,}
\keyword{PCA}
