% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{olink_pca_plot}
\alias{olink_pca_plot}
\title{Function to plot a PCA of the data}
\usage{
olink_pca_plot(df, color_g = "QC_Warning", x_val = 1, y_val = 2,
  label_samples = F)
}
\arguments{
\item{df}{data frame in long format with Sample Id, NPX and column of choice for colors}

\item{color_g}{Character value indicating which column to use for colors (default QC_Warning)}

\item{x_val}{Integer indicating which principal component to plot along the x-axis (default 1)}

\item{y_val}{Integer indicating which principal component to plot along the y-axis (default 2)}

\item{label_samples}{Logical. If TRUE, points are replaced with SampleID (default FALSE)}
}
\value{
An object of class "ggplot"
}
\description{
Creates a scatterplot of all samples from an NPX data frame along two principal components. Unique sample names are required.
}
\examples{
\donttest{olink_pca_plot(df=npx.data, color_g = "QC_Warning")  }
}
\keyword{NPX,}
\keyword{PCA}
