% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_ttest.R
\name{olink_ttest}
\alias{olink_ttest}
\title{Function which performs a t-test per protein}
\usage{
olink_ttest(df, variable, ...)
}
\arguments{
\item{df}{NPX data frame in long format with at least protein name (Assay), OlinkID, UniProt and a factor with 2 levels.}

\item{variable}{Character value indicating which column should be used as the grouping variable. Needs to have exactly 2 levels}

\item{...}{Options to be passed to t.test. See ?t.test for more.}
}
\value{
A data frame containing the t-test results for every protein.
}
\description{
Performs a 2-sample t-test on for every protein (by OlinkID) for a given grouping variable.
}
\examples{
\donttest{ttest_results <- olink_ttest(df=npx.data, variable = 'Group', alternative = 'two.sided')}
}
