% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_plot.R
\name{olink_umap_plot}
\alias{olink_umap_plot}
\title{Function to make a UMAP plot from the data}
\usage{
olink_umap_plot(
  df,
  color_g = "QC_Warning",
  x_val = 1,
  y_val = 2,
  label_samples = FALSE,
  config = umap.defaults,
  drop_assays = FALSE,
  drop_samples = FALSE,
  returnEmbedding = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{df}{data frame in long format with Sample Id, NPX and column of choice for colors}

\item{color_g}{Character value indicating which column to use for colors (default QC_Warning)}

\item{x_val}{Integer indicating which UMAP component to plot along the x-axis (default 1)}

\item{y_val}{Integer indicating which UMAP component to plot along the y-axis (default 2)}

\item{label_samples}{Logical. If TRUE, points are replaced with SampleID (default FALSE)}

\item{config}{object of class umap.config, specifying the parameters for the UMAP algorithm.}

\item{drop_assays}{Logical. All assays with any missing values will be dropped. Takes precedence over sample drop.}

\item{drop_samples}{Logical. All samples with any missing values will be dropped.}

\item{returnEmbedding}{Logical. If true, the UMAP results will be returned.}

\item{verbose}{Logical. Whether warnings about the number of samples and/or assays dropped or imputed should be printed to the console.}

\item{...}{coloroption passed to specify color order.}
}
\value{
If returnEmbedding == FALSE (default): An object of class "ggplot". If returnEmbedding == TRUE: a list with 1. an object of class "ggplot", 2. an object of class umap
}
\description{
NOTE: This is a beta version.
Computes a manifold approximation and projection using umap::umap and plots the two specified components. 
Unique sample names are required and imputation by the median is done for assays with missingness <10\% for multi-plate projects and <5\% for single plate projects.
}
\examples{
\donttest{
npx_data <- npx_data1 \%>\%
    mutate(SampleID = paste(SampleID, "_", Index, sep = ""))
olink_pca_plot(df=npx_data, color_g = "QC_Warning")}
}
\keyword{NPX,}
\keyword{UMAP}
