% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{olink_volcano_plot}
\alias{olink_volcano_plot}
\title{Generate linear models and make volcano plot}
\usage{
olink_volcano_plot(long, variable, random, covariates = NULL)
}
\description{
Generate linear models and make volcano plot
}
