% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_wilcox.R
\name{olink_wilcox}
\alias{olink_wilcox}
\title{Function which performs a Mann-Whitney U Test per protein}
\usage{
olink_wilcox(df, variable, pair_id, check_log = NULL, ...)
}
\arguments{
\item{df}{NPX or Quantified_value data frame in long format with at least
protein name (Assay), OlinkID, UniProt and a factor with 2 levels.}

\item{variable}{Character value indicating which column should be used as the
grouping variable. Needs to have exactly 2 levels.}

\item{pair_id}{Character value indicating which column indicates the paired
sample identifier.}

\item{check_log}{Optional named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL}, an
attached \code{check_log} is used when present; otherwise \code{\link[=check_npx]{check_npx()}} will
be run internally using \code{df}.}

\item{...}{Options to be passed to wilcox.test. See \code{?wilcox_test} for
more information.}
}
\value{
A data frame containing the Mann-Whitney U Test results for every
protein. Columns include:
\itemize{
\item{Assay:} "character" Protein symbol
\item{OlinkID:} "character" Olink specific ID
\item{UniProt:} "character" UniProt ID
\item{Panel:} "character" Name of Olink Panel
\item{estimate:} "numeric" median of NPX differences between groups
\item{statistic:} "named numeric" the value of the test statistic with a
name describing it
\item{p.value:} "numeric" p-value for the test
\item{conf.low:} "numeric" confidence interval for the median of differences
(lower end)
\item{conf.high:} "numeric" confidence interval for the median of
differences (upper end)
\item{method:} "character" which wilcoxon method was used
\item{alternative:} "character" describes the alternative hypothesis
\item{Adjusted_pval:} "numeric" adjusted p-value for the test
(Benjamini&Hochberg)
\item{Threshold:} "character" if adjusted p-value is significant or not
(< 0.05)
}
}
\description{
Performs a Welch 2-sample Mann-Whitney U Test at confidence level 0.95 for
every protein (by OlinkID) for a given grouping variable using
stats::wilcox.test and corrects for multiple testing by the
Benjamini-Hochberg method (“fdr”) using stats::p.adjust. Adjusted p-values
are logically evaluated towards adjusted p-value<0.05. The resulting
Mann-Whitney U Test table is arranged by ascending p-values.
}
\examples{
\donttest{
if (rlang::is_installed(pkg = c("broom"))) {
  npx_df <- npx_data1 |>
    dplyr::filter(
    !grepl(
      pattern = "control",
      x = .data[["SampleID"]],
      ignore.case = TRUE
    )
  )

  # Mann-Whitney U Test
  wilcox_results <- OlinkAnalyze::olink_wilcox(
    df = npx_df,
    variable = "Treatment",
    alternative = "two.sided"
  )

  # Paired Mann-Whitney U Test
  wilcox_paired_results <- npx_df |>
    dplyr::filter(
      .data[["Time"]] \%in\% c("Baseline", "Week.6")
    ) |>
    OlinkAnalyze::olink_wilcox(
      variable = "Time",
      pair_id = "Subject"
    )
}
}

}
