% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised.R
\name{plot_rf}
\alias{plot_rf}
\title{plot random forests}
\usage{
plot_rf(
  rf,
  no_of_trees_label = 40,
  accuracy_decrease_label = 0.004,
  return_imp_frame = FALSE
)
}
\description{
plot random forests
}
