% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx.R
\name{read_npx}
\alias{read_npx}
\alias{read_NPX}
\title{Read Olink data in R.}
\usage{
read_npx(
  filename,
  out_df = "tibble",
  sep = NULL,
  long_format = NULL,
  olink_platform = NULL,
  data_type = NULL,
  .ignore_files = c("README.txt"),
  quiet = TRUE,
  legacy = FALSE
)

read_NPX(
  filename,
  out_df = "tibble",
  sep = NULL,
  long_format = NULL,
  olink_platform = NULL,
  data_type = NULL,
  .ignore_files = c("README.txt"),
  quiet = TRUE,
  legacy = FALSE
)
}
\arguments{
\item{filename}{Path to Olink software output file in wide or long format.
Expecting extensions "xls" or "xlsx" for excel files, "csv" or "txt" for delim files, "parquet" for parquet files, and "zip" for compressed files.}

\item{out_df}{The class of the output dataset. One of
"tibble" and "arrow". (default = "tibble")}

\item{sep}{Character separator of delimited input file. One of \code{NULL}
(default) for auto-detection, or ";" and ",".
Used only for delimited output files from Olink software.}

\item{long_format}{Boolean marking format of input file. One of \code{NULL}
(default) for auto-detection, or \code{TRUE} for long format and \code{FALSE} for
wide format files.}

\item{olink_platform}{Olink platform used to generate the input file.
One of \code{NULL} (default) for auto-detection,
or "Target 48", "Flex", "Target 96", "Explore 3072", "Explore HT", and "Focus".}

\item{data_type}{Quantification method of the input data. One of \code{NULL}
(default) for auto-detection, "Ct", "NPX", and "Quantified".}

\item{.ignore_files}{Character vector of files included in the zip-compressed
Olink software output files that should be ignored. Used only for
zip-compressed input files (default = \emph{c("README.txt")}).}

\item{quiet}{Boolean to print a confirmation message when reading the input
file. Applies to excel or delimited input only. \code{TRUE} (default) to not print
and \code{FALSE} to print.}

\item{legacy}{Boolean to run the legacy version of the read_npx function.
\strong{Important: should be used only to wide format files from Target 96 or
Target 48 with NPX Software version earlier than 1.8!} (default \code{FALSE}).}
}
\value{
Dataset, tibble, ArrowObject, with Olink data in long format.
}
\description{
Imports a file exported from Olink software that quantifies protein levels in
NPX, Ct or absolute quantification.

\strong{Note:} Do not modify the Olink software output file prior to
importing it with \code{\link{read_npx}} as it might fail.
}
\examples{
\donttest{
file <- system.file("extdata",
                    "npx_data_ext.parquet",
                    package = "OlinkAnalyze")
read_npx(filename = file,
         olink_platform = "Target 96")
}

}
\author{
Klev Diamanti;
Kathleen Nevola;
Pascal Pucholt;
Christoffer Cambronero;
Boxi Zhang;
Olof Mansson;
Marianne Sandin
}
\keyword{1536;}
\keyword{3072;}
\keyword{48;}
\keyword{96;}
\keyword{Explore}
\keyword{Explore;}
\keyword{Flex;}
\keyword{Focus}
\keyword{HT;}
\keyword{NPX;}
\keyword{Olink}
\keyword{Olink;}
\keyword{Target}
\keyword{csv;}
\keyword{parquet;}
\keyword{xlsx;}
\keyword{zip;}
