% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx.R
\name{read_npx}
\alias{read_npx}
\alias{read_NPX}
\title{Function to read NPX data into long format}
\usage{
read_npx(filename)

read_NPX(filename)
}
\arguments{
\item{filename}{Path to Olink Software output file.}
}
\value{
A "tibble" in long format.
}
\description{
Imports an NPX or QUANT file exported from Olink Software.
No alterations to the output format is allowed.
}
\examples{
\donttest{
file <- system.file("extdata", "Example_NPX_Data.csv", package = "OlinkAnalyze")
read_NPX(file)
}

}
\author{
Klev Diamanti;
  Kathleen Nevola;
  Pascal Pucholt;
  Christoffer Cambronero;
  Boxi Zhang;
  Olof Mansson;
  Marianne Sandin
}
\keyword{1536}
\keyword{3072}
\keyword{3k}
\keyword{Explore}
\keyword{Flex}
\keyword{HT}
\keyword{NPX}
\keyword{Olink}
\keyword{Target48}
\keyword{Target96}
\keyword{csv}
\keyword{parquet}
\keyword{xlsx}
\keyword{zip}
