% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_NPX_data.R
\name{read_NPX}
\alias{read_NPX}
\title{Function to read NPX data into long format}
\usage{
read_NPX(filename)
}
\arguments{
\item{filename}{Path to file NPX Manager output file.}
}
\value{
A tibble in long format.
}
\description{
Imports an NPX file exported from NPX Manager.
No alterations to the output NPX Manager format is allowed.
}
\examples{
\donttest{
file <- system.file("extdata", "Example_NPX_Data.csv", package = "OlinkAnalyze")
read_NPX(file)
}
}
\keyword{NPX}
