% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_delim.R
\name{read_npx_delim}
\alias{read_npx_delim}
\alias{read_npx_csv}
\title{Help function to read "Ct", "NPX", and "Quantified" data
from delimited Olink software output files in R.}
\usage{
read_npx_delim(file, out_df = "arrow", sep = NULL)

read_npx_csv(file, out_df = "arrow", sep = NULL)
}
\arguments{
\item{file}{Path to Olink software output delimited file in wide or long
format. Expecting file extensions
"csv" and "txt".}

\item{out_df}{The class of output data frame. One of
"tibble" and "arrow".}

\item{sep}{Character separator of delimited input file. One of \code{NULL}
(default) for auto-detection, or ";" and ",".
Used only for delimited output files from Olink software.}
}
\value{
Tibble or ArrowObject with Olink data in long format.
}
\description{
The function can handle delimited files in long and wide format.
}
\seealso{
\code{\link{read_npx}}
\code{\link{read_npx_parquet}}
\code{\link{read_npx_zip}}
\code{\link{read_npx_excel}}
\code{\link{read_npx_format}}
\code{\link{read_npx_delim_wide}}
\code{\link{read_npx_delim_long}}
}
\author{
Klev Diamanti;
Christoffer Cambronero;
Kathleen Nevola
}
