% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_delim.R
\name{read_npx_delim}
\alias{read_npx_delim}
\alias{read_npx_csv}
\title{Help function to read "Ct", "NPX", and "Quantified" data
from delimited Olink software output files in R.}
\usage{
read_npx_delim(file, out_df = "arrow")

read_npx_csv(file, out_df = "arrow")
}
\arguments{
\item{file}{Path to Olink software output delimited file in wide or long
format. Expecting file extensions
"csv" or "txt".}

\item{out_df}{The class of the output dataset. One of
"tibble" or "arrow". (default =
"tibble")}
}
\value{
Dataset,
"tibble" or "ArrowObject", with Olink
data in long format.
}
\description{
The function can handle delimited files in long and wide format.
}
\seealso{
\code{\link{read_npx}}
\code{\link{read_npx_parquet}}
\code{\link{read_npx_zip}}
\code{\link{read_npx_excel}}
\code{\link{read_npx_format}}
}
\author{
Klev Diamanti;
Christoffer Cambronero;
Kathleen Nevola
}
