% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_delim.R
\name{read_npx_delim_wide}
\alias{read_npx_delim_wide}
\title{Help function to read wide format
"csv" and "txt" data from
delimited file exported from Olink software.}
\usage{
read_npx_delim_wide(file, sep)
}
\arguments{
\item{file}{Path to Olink software output delimited file in wide format.
Expecting file extensions
"csv" and "txt".}

\item{sep}{Character separator of delimited input file. One of \code{NULL}
(default) for auto-detection, or ";", ",", "    ", "|", and ":".
Used only for delimited output files from Olink software.}
}
\value{
"tibble" or "ArrowObject" with
Olink data in wide format.
}
\description{
The function can handle delimited files in wide format.
}
\seealso{
\code{\link{read_npx_delim_long}}
\code{\link{read_npx_delim}}
}
\author{
Klev Diamanti
}
