% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_excel.R
\name{read_npx_excel}
\alias{read_npx_excel}
\title{Help function to read long or wide format
"Ct", "NPX", or "Quantified" data from
Microsoft
"xls" or "xlsx"
files exported from Olink software in R.}
\usage{
read_npx_excel(file, out_df = "arrow")
}
\arguments{
\item{file}{Path to Olink software output excel file in wide or long format.
Expected file extensions
"xls" or "xlsx".}

\item{out_df}{The class of the output dataset. One of
"tibble" or "arrow". (default =
"tibble")}
}
\value{
Dataset,
"tibble" or "ArrowObject", with Olink
data in long or wide format.
}
\description{
Help function to read long or wide format
"Ct", "NPX", or "Quantified" data from
Microsoft
"xls" or "xlsx"
files exported from Olink software in R.
}
\seealso{
\code{\link{read_npx}}
\code{\link{read_npx_parquet}}
\code{\link{read_npx_zip}}
\code{\link{read_npx_format}}
\code{\link{read_npx_delim}}
}
\author{
Klev Diamanti;
Christoffer Cambronero;
Kathleen Nevola
}
