% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_excel.R
\name{read_npx_excel}
\alias{read_npx_excel}
\title{Help function to read in Olink data in long format from comma or semicolon
delimited files.}
\usage{
read_npx_excel(file, out_df = "arrow")
}
\arguments{
\item{file}{Path to Olink software output in txt or csv.}

\item{out_df}{The class of output data frame to be returned. Accepted values
are "tibble" and "arrow" (default).}
}
\value{
An R6 class ArrowObject or tibble.
}
\description{
Help function to read in Olink data in long format from comma or semicolon
delimited files.
}
\seealso{
[read_npx()]
  [read_npx_parquet()]
  [read_npx_zip()]
  [read_npx_delim()]
  [read_npx_delim_wide()]
}
\author{
Klev Diamanti;
  Christoffer Cambronero;
  Kathleen Nevola
}
\keyword{NPX}
\keyword{csv}
\keyword{delim}
\keyword{sep}
\keyword{txt}
