% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_excel.R
\name{read_npx_excel}
\alias{read_npx_excel}
\title{Help function to read "Ct", "NPX", and "Quantified" data
from excel Olink software output files in R.}
\usage{
read_npx_excel(file, out_df = "arrow")
}
\arguments{
\item{file}{Path to Olink software output excel file in wide or long format.
Expected file extensions
"xls" and "xlsx". # nolint line_length_linter}

\item{out_df}{The class of output data frame. One of
"tibble" and "ArrowObject".}
}
\value{
Tibble or ArrowObject with Olink data in wide or long format.
}
\description{
Help function to read "Ct", "NPX", and "Quantified" data
from excel Olink software output files in R.
}
\seealso{
\code{\link{read_npx}}
\code{\link{read_npx_parquet}}
\code{\link{read_npx_zip}}
\code{\link{read_npx_format}}
\code{\link{read_npx_delim}}
}
\author{
Klev Diamanti;
Christoffer Cambronero;
Kathleen Nevola
}
