% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_format.R
\name{read_npx_format}
\alias{read_npx_format}
\title{Help function to read excel and delimited Olink data files in R and determine
their format, data type and platform.}
\usage{
read_npx_format(
  file,
  out_df = "arrow",
  sep = NULL,
  long_format = NULL,
  olink_platform = NULL,
  data_type = NULL,
  quiet = FALSE,
  legacy = FALSE
)
}
\arguments{
\item{file}{Path to Olink software output file in wide or long format.
Expecting file extensions
"xls", "xlsx", "csv", and "txt".}

\item{out_df}{The class of output data frame. One of
"tibble" and "arrow".}

\item{sep}{Character separator of delimited input file. One of \code{NULL}
(default) for auto-detection, or ";" and ",".
Used only for delimited output files from Olink software.}

\item{long_format}{Boolean marking format of input file. One of \code{NULL}
(default) for auto-detection, \code{TRUE} for long format files or \code{FALSE} for
wide format files.}

\item{olink_platform}{Olink platform used to generate the input file.
One of \code{NULL} (default) for auto-detection,
"Flex", "Focus", "Target 48", and "Target 96".}

\item{data_type}{Quantification method of the input data. One of \code{NULL}
(default) for auto-detection, "Ct", "NPX", and "Quantified".}

\item{quiet}{Boolean to print a confirmation message when reading the input
file. \code{TRUE} (default) to skip printing, and \code{FALSE} to print.}

\item{legacy}{Boolean to enforce returning a list containing olink_platform,
data_type and long_format information together with the dataset. Used only
when \code{\link{read_npx_format}} is called from
\code{\link{read_npx_legacy}}.}
}
\value{
"tibble" or "ArrowObject" with
Olink data in long or wide format.
}
\description{
This function processes Olink software excel or delimited files regardless of
data type, platform or format.

\strong{Olink software excel files} with the extension
"xls" or "xlsx"
are imported in R by the function \code{\link{read_npx_excel}}.

\strong{Olink software delimited files} with suffix
"csv" or "txt"
are imported in R by the functions \code{\link{read_npx_delim}} or
\code{\link{read_npx_csv}}.

Files in wide format are subsequently handled by the function
\code{\link{read_npx_wide}}.

\strong{Olink software files in wide format} always originate from the Olink
qPCR platforms, and are further processed by
\code{\link{read_npx_format_get_format}} and
\code{\link{read_npx_format_get_quant}} to determine the data type and Olink
platform, respectively.
}
\seealso{
\code{\link{read_npx}}
\code{\link{read_npx_format_read}}
\code{\link{read_npx_format_get_format}}
\code{\link{read_npx_format_get_platform}}
\code{\link{read_npx_format_get_quant}}
\code{\link{read_npx_legacy}}
}
\author{
Klev Diamanti;
Kathleen Nevola;
Pascal Pucholt;
Christoffer Cambronero;
Boxi Zhang;
Olof Mansson;
Marianne Sandin
}
