% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_format.R
\name{read_npx_format}
\alias{read_npx_format}
\title{Help function to identify the format, data type and platform from the input
Olink file.}
\usage{
read_npx_format(
  file,
  out_df = "arrow",
  sep = NULL,
  long_format = NULL,
  olink_platform = NULL,
  data_type = NULL,
  quiet = FALSE
)
}
\arguments{
\item{file}{The input Olink file in excel or delimited format.}

\item{out_df}{The class of output data frame to be returned. Accepted values
are "tibble" (default) and "arrow".}

\item{sep}{The separator of the file: NULL (autodetect), comma (,) or
semicolon (;). Ignored if file is in excel format.}

\item{long_format}{Boolean marking if input file is in long (TRUE) or wide
(FALSE) format.}

\item{olink_platform}{The Olink platform used to generate the input file.
Expecting "Explore 3072", "Explore HT", "Target 96", "Target 48", "Flex" or
"Focus".}

\item{data_type}{The quantification in which the data comes in. Expecting one
of NPX, Quantified or Ct.}

\item{quiet}{Print a confirmation message after reading in the input file.}
}
\value{
A tibble in long format or an ArrowObject.
}
\description{
This function should be processing all Olink excel or delimited files
regardless of data type, platform or wide/long format. Files in wide format
should be handled by read_npx_wide; files in long format delimited by comma
or semicolon (csv) should be handled by read_npx_delim; and files in long
format in excel (xlsx) should be handled by read_npx_excel. Only files in
wide format that always originate from qPCR platforma should be further
investigated for data type and olink platform.
}
\author{
Klev Diamanti;
  Kathleen Nevola;
  Pascal Pucholt;
  Christoffer Cambronero;
  Boxi Zhang;
  Olof Mansson;
  Marianne Sandin
}
