% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_utils.R
\name{read_npx_format_colnames}
\alias{read_npx_format_colnames}
\title{Help function checking whether a data set contains NA or empty strings on
its column names}
\usage{
read_npx_format_colnames(df, file)
}
\arguments{
\item{df}{Tibble or ArrowObject with Olink data in wide or long format.}

\item{file}{Path to Olink software output file in wide or long format.
Expecting file extensions \code{csv}, \code{txt}, \code{xls}, \code{xlsx}, \code{parquet} or \code{zip}.}
}
\value{
Error is file contains problematic column names. \code{NULL} otherwise.
}
\description{
Help function checking whether a data set contains NA or empty strings on
its column names
}
\author{
Klev Diamanti
}
