% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_format.R
\name{read_npx_format_get_quant}
\alias{read_npx_format_get_quant}
\title{Help function to determine the type of quantification from the input file in
wide format.}
\usage{
read_npx_format_get_quant(file, data_type = NULL, data_cells)
}
\arguments{
\item{file}{Path to Olink software output file in wide format. Expecting file
extensions
"xls", "xlsx", "csv", and "txt".}

\item{data_type}{Quantification method of the input data. One of \code{NULL}
(default) for auto-detection, "Ct", "NPX", and "Quantified".}

\item{data_cells}{A character vector with the contents of the cell \emph{A2}
from the Olink software file in wide format indicating the quantification
method.}
}
\value{
The name of the data type. One of
"Ct", "NPX", and "Quantified".
}
\description{
This function uses information from the cell \emph{A2} from Olink software
files in wide format to determine the quantification method of the data that
the dataset contains.
}
\seealso{
\code{\link{read_npx_format}}
\code{\link{read_npx_format_read}}
\code{\link{read_npx_format_get_format}}
\code{\link{read_npx_format_get_platform}}
}
\author{
Klev Diamanti
}
