% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_format.R
\name{read_npx_format_read}
\alias{read_npx_format_read}
\title{Help function to read excel and delimited Olink data files in R.}
\usage{
read_npx_format_read(file, sep = NULL, read_n = 3L)
}
\arguments{
\item{file}{Path to Olink software output file in wide or long format.
Expecting file extensions
"xls", "xlsx", "csv", and "txt".}

\item{sep}{Character separator of delimited input file. One of \code{NULL}
(default) for auto-detection, or ";" and ",".
Used only for delimited output files from Olink software.}

\item{read_n}{Number of top rows to read.}
}
\value{
A list with two elements:
\itemize{
\item An ArrowObject (\var{df}) containing the full dataset.
\item A tibble (\var{df_top_n}) containing the \var{read_n} rows of the full
dataset. This subset of data is used to determine \var{long_format},
\var{olink_platform} and \var{data_type}.
}
}
\description{
This function reads Olink software excel or delimited files regardless of
data type, platform or format.

\strong{Olink software excel files} with the extension
"xls" or "xlsx"
are imported in R by the function \code{\link{read_npx_excel}}.

\strong{Olink software delimited files} with suffix
"csv" or "txt"
are imported in R by the functions \code{\link{read_npx_delim}} or
\code{\link{read_npx_csv}}.

\itemize{
\item Files in long format are read with the column row as column names.
\item Files in wide format are read with column names as V1, V2, etc.
}

This function also extracts the first \var{read_n} rows of the dataset read
earlier to determine the Olink platform that generated the file, the data
type and and file format.
}
\seealso{
\code{\link{read_npx_format}}
\code{\link{read_npx_format_get_format}}
\code{\link{read_npx_format_get_platform}}
\code{\link{read_npx_format_get_quant}}
}
\author{
Klev Diamanti
}
