% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_legacy.R
\name{read_npx_legacy}
\alias{read_npx_legacy}
\title{Olink legacy function for reading NPX or absolute quantification data in wide
format into R from qPCR Olink products.}
\usage{
read_npx_legacy(
  file,
  out_df = "tibble",
  olink_platform = NULL,
  data_type = NULL,
  quiet = TRUE
)
}
\arguments{
\item{file}{Path to Olink software output file in wide format. Expected one
of file extensions
"xls", "xlsx", "csv", and "txt".}

\item{out_df}{The class of the output dataset. One of
"tibble" and "arrow". (default = "tibble")}

\item{olink_platform}{Olink platform used to generate the input file.
One of \code{NULL} (default) for auto-detection,
"Flex", "Focus", "Target 48", and "Target 96".}

\item{data_type}{Quantification method of the input data. One of \code{NULL}
(default) for auto-detection, "Ct", "NPX", and "Quantified".}

\item{quiet}{Boolean to print a confirmation message when reading the input
file. Applies to excel or delimited input only. \code{TRUE} (default) to not print
and \code{FALSE} to print.}
}
\value{
Tibble or ArrowObject with Olink data in long format.
}
\description{
This implementation of read_NPX does not cover the latest versions of Olink
files in wide format. Specifically, it supports:
\itemize{
\item \strong{Target 96} output files in wide format \strong{(T96 reports
only NPX)} with the bottom matrix containing one of the following
combinations of rows:
\itemize{
\item Missing Data freq. and LOD. # nolint line_length_linter
\item Missing Data freq., LOD, and Normalization. # nolint line_length_linter
}
\item \strong{Target 48} output files in wide format \strong{NPX} with the
bottom matrix containing the following rows:
Missing Data freq., Normalization, and LOD. # nolint line_length_linter
\item \strong{Target 48} output files in wide format \strong{absolute
Quantification} with the bottom matrix containing the following rows:
Assay warning, LLOQ, Lowest quantifiable level, Missing Data freq., Normalization, Plate_LOD, and ULOQ. # nolint line_length_linter
}

This function would accept data exported in wide format from Olink NPX
Signature 1.7.1 or earlier, or NPX Manager.
}
\seealso{
\code{\link{read_npx_format_read}}
\code{\link{read_npx_format_get_format}}
\code{\link{read_npx_format_get_platform}}
\code{\link{read_npx_format_get_quant}}
}
\author{
Klev Diamanti;
Kathleen Nevola;
Pascal Pucholt;
Christoffer Cambronero;
Boxi Zhang;
Olof Mansson;
Marianne Sandin
}
