% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_legacy.R
\name{read_npx_legacy_help}
\alias{read_npx_legacy_help}
\title{Help function utilizing functions from \code{\link{read_npx_format}} and
\code{\link{read_npx_wide}} to streamline \code{\link{read_npx_legacy}}}
\usage{
read_npx_legacy_help(
  file,
  out_df,
  olink_platform = NULL,
  data_type = NULL,
  data_type_no_accept = c("Ct")
)
}
\arguments{
\item{file}{Path to Olink software output file in wide format. Expected one
of file extensions
"xls", "xlsx", "csv", or "txt".}

\item{out_df}{The class of the output dataset. One of
"tibble" or "arrow". Defaults to
"tibble".}

\item{olink_platform}{Olink platform used to generate the input file.
One of \code{NULL} (default) for auto-detection,
"Flex", "Focus", "Target 48", or "Target 96".}

\item{data_type}{Quantification method of the input data. One of
"Ct", "NPX", or "Quantified". Defaults to
\code{NULL} for auto-detection.}

\item{data_type_no_accept}{Character vector of data types that should be
rejected (default = "Ct").}
}
\value{
A list of objects containing the following:
\itemize{
\item \strong{olink_platform}: auto-detected Olink platform. One of
"Flex", "Focus", "Target 48", and "Target 96".
\item \strong{long_format}: auto-detected Olink format. Should always be
"FALSE".
\item \strong{data_type}: auto-detected Olink data type. One of
"NPX" and "Quantified". # nolint line_length_linter
\item \strong{df_split}: list of 2 tibbles. Top matrix from the Olink wide
file, and middle combined with bottom matrix.
\item \strong{npxs_v}: Olink NPX software version.
\item \strong{bottom_mat_v}: bottom matrix version based on
\var{olink_wide_bottom_matrix}.
\item \strong{format_spec}: specifications of the wide format based on
\var{olink_wide_spec}.
}
}
\description{
Help function utilizing functions from \code{\link{read_npx_format}} and
\code{\link{read_npx_wide}} to streamline \code{\link{read_npx_legacy}}
}
\author{
Klev Diamanti
}
