% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_parquet.R
\name{read_npx_parquet}
\alias{read_npx_parquet}
\title{Helper function to read in Olink Explore parquet files}
\usage{
read_npx_parquet(file)
}
\arguments{
\item{file}{Path to Olink Software parquet output file.}
}
\value{
An R6 class ArrowObject.
}
\description{
Helper function to read in Olink Explore parquet files
}
\examples{
\donttest{
file <- system.file("extdata", "npx_data_ext.parquet",
package = "OlinkAnalyze")
read_npx_parquet(file)
}
}
\keyword{NPX}
