% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_parquet.R
\name{read_npx_parquet}
\alias{read_npx_parquet}
\title{Help function to read Olink parquet files.}
\usage{
read_npx_parquet(file, out_df = "arrow")
}
\arguments{
\item{file}{Path to Olink Software parquet output file.}

\item{out_df}{The class of output data frame to be returned. Accepted values
are "tibble" and "arrow" (default).}
}
\value{
An R6 class ArrowObject.
}
\description{
Help function to read Olink parquet files.
}
\seealso{
[read_npx()]
  [read_npx_delim()]
  [read_npx_delim_wide()]
  [read_npx_zip()]
  [read_npx_excel()]
}
\author{
Klev Diamanti;
  Kathleen Nevola;
  Pascal Pucholt
}
\keyword{NPX}
\keyword{parquet}
