% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide}
\alias{read_npx_wide}
\title{Convert Olink data in wide format with
"Ct", "NPX", and "Quantified" data to
long format.}
\usage{
read_npx_wide(df, file, data_type, olink_platform)
}
\arguments{
\item{df}{A tibble containing the full Olink dataset in wide format.}

\item{file}{Path to Olink software output file in wide format. Expected one
of file extensions
"xls", "xlsx", "csv", or "txt".}

\item{data_type}{Quantification method of the input data. One of
"Ct", "NPX", or "Quantified". Defaults to \code{NULL} for
auto-detection.}

\item{olink_platform}{Olink platform used to generate the input file.
One of \code{NULL} (default) for auto-detection,
"Flex", "Focus", "Target 48", or "Target 96".}
}
\value{
Dataset,
"tibble" or "ArrowObject", with Olink
data in long format.
}
\description{
Convert Olink data in wide format with
"Ct", "NPX", and "Quantified" data to
long format.
}
\seealso{
\code{\link{read_npx_format}}
\code{\link{read_npx_wide_split_row}}
\code{\link{read_npx_wide_npxs_version}}
\code{\link{read_npx_wide_top}}
\code{\link{read_npx_wide_middle}}
\code{\link{read_npx_wide_bottom}}
}
\author{
Klev Diamanti
}
