% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide}
\alias{read_npx_wide}
\title{Convert Olink data in wide format with NPX, Ct or Quantified data to long
format.}
\usage{
read_npx_wide(df, file, data_type, olink_platform)
}
\arguments{
\item{df}{A tibble containing the full Olink dataset in wide format.}

\item{file}{Path to Olink software output file in wide or long format.
Expecting file extensions
xls, xlsx, csv or txt. # nolint}

\item{data_type}{Quantification method of the input data. One of
Ct, NPX or Quantified. # nolint}

\item{olink_platform}{Olink platform used to generate the input file.
One of
Target 48, Flex, Target 96 or Focus. # nolint}
}
\value{
Tibble with Olink data in long format.
}
\description{
Convert Olink data in wide format with NPX, Ct or Quantified data to long
format.
}
\seealso{
\code{\link{read_npx_format}}
\code{\link{read_npx_wide_split_row}}
\code{\link{read_npx_wide_npxs_version}}
\code{\link{read_npx_wide_top}}
\code{\link{read_npx_wide_middle}}
\code{\link{read_npx_wide_bottom}}
}
\author{
Klev Diamanti
}
