% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide_utils.R
\name{read_npx_wide_assay_nrow}
\alias{read_npx_wide_assay_nrow}
\title{Help function to determine the number of rows with assay information in Olink
excel long or wide format files.}
\usage{
read_npx_wide_assay_nrow(file)
}
\arguments{
\item{file}{The input excel file.}
}
\value{
Integer number of rows containing info about assays in Olink excel
wide files.
}
\description{
Help function to determine the number of rows with assay information in Olink
excel long or wide format files.
}
\author{
Klev Diamanti
}
