% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_bottom}
\alias{read_npx_wide_bottom}
\title{Help function that converts df_bottom into long format.}
\usage{
read_npx_wide_bottom(df, file, col_split, assay_cols, format_spec)
}
\arguments{
\item{df}{The bottom data frame from a split Olink wide excel file.}

\item{file}{The input excel file.}

\item{col_split}{The name of the column that splits the Olink wide excel file
into left (assay info) and right  hand side (PlateID and QC_Warning info).}

\item{assay_cols}{Character vector with the names of the columns containing
Olink assays.}

\item{format_spec}{A one-row data frame filtered from olink_wide_excel_spec
with the Olink wide excel file specifications.}
}
\value{
The bottom matrix in long format.
}
\description{
Help function that converts df_bottom into long format.
}
\author{
Klev Diamanti
}
