% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_bottom}
\alias{read_npx_wide_bottom}
\title{Help function that converts the bottom matrix of an Olink wide file to long
format.}
\usage{
read_npx_wide_bottom(
  df,
  file,
  olink_platform,
  data_type,
  col_names,
  format_spec,
  df_plate_panel
)
}
\arguments{
\item{df}{The bottom data frame from a split Olink wide file.}

\item{file}{The input Olink wide file.}

\item{olink_platform}{The Olink platform used to generate the input file.
Expecting one of "Target 96", "Target 48", "Flex" or "Focus".}

\item{data_type}{The quantification in which the data comes in. Expecting one
of "NPX", "Quantified" or "Ct".}

\item{col_names}{Named list of character vectors with the names of the
columns containing Olink data on Assays and Internal Control assays.}

\item{format_spec}{A one-row data frame filtered from olink_wide_spec with
the Olink wide file specifications.}

\item{df_plate_panel}{Data frame with unique combinations of panels and
plates from the combination of top and middle data frames.}
}
\value{
A tibble with the bottom matrix of an Olink wide file in long format.
}
\description{
Help function that converts the bottom matrix of an Olink wide file to long
format.
}
\author{
Klev Diamanti
}
