% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_bottom}
\alias{read_npx_wide_bottom}
\title{Convert the bottom matrix from Olink dataset in wide format to long.}
\usage{
read_npx_wide_bottom(
  df,
  file,
  olink_platform,
  data_type,
  col_names,
  format_spec,
  df_plate_panel
)
}
\arguments{
\item{df}{Bottom matrix of Olink dataset in wide format \var{df_bottom}.}

\item{file}{Path to Olink software output file in wide or long format.
Expecting file extensions
xls, xlsx, csv or txt. # nolint}

\item{olink_platform}{Olink platform used to generate the input file.
One of
Target 48, Flex, Target 96 or Focus. # nolint}

\item{data_type}{Quantification method of the input data. One of
Ct, NPX or Quantified. # nolint}

\item{col_names}{Names list of character vectors containing column names from
each chunk of columns \var{df_top} was split on in function.}

\item{format_spec}{A tibble derived from \var{olink_wide_spec} in the local
environment containing the expected format of the Olink wide file based on
the \var{olink_platform} and \var{data_type}.}

\item{df_plate_panel}{Tibble with unique combinations of panels and plates
from the combination of top and middle data frames.}
}
\value{
A tibble with the bottom matrix of an Olink wide file in long format.
}
\description{
Use chunks of columns from \code{\link{read_npx_wide_top}} to covert the
bottom matrix \var{df_bottom} into a long format tibble.
}
\seealso{
\code{\link{read_npx_wide}}
\code{\link{read_npx_wide_split_row}}
\code{\link{read_npx_wide_npxs_version}}
\code{\link{read_npx_wide_top}}
\code{\link{read_npx_wide_middle}}
}
\author{
Klev Diamanti
}
