% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_bottom_version}
\alias{read_npx_wide_bottom_version}
\title{Additional checks of the bottom matrix of Olink dataset in wide format.}
\usage{
read_npx_wide_bottom_version(df, file, data_type, olink_platform)
}
\arguments{
\item{df}{Bottom matrix of Olink dataset in wide format \var{df_bottom}.}

\item{file}{Path to Olink software output file in wide or long format.
Expecting file extensions \code{csv}, \code{txt}, \code{xls}, or \code{xlsx}.}

\item{data_type}{The quantification in which the data comes in. Expecting one
of "NPX", "Quantified" or "Ct".}

\item{olink_platform}{Olink platform used to generate the input file.
One of \verb{Target 96}, \verb{Target 48}, \code{Flex} or \code{Focus}.}
}
\value{
Tibble with the bottom matrix specifications for the Olink wide file.
}
\description{
The rows included in the bottom matrix have evolved through the years. For us
to be able to support as many such versions as possible we have used the
local environment variable \var{olink_wide_bottom_matrix} to mark these
different versions. This function extract these version and allows us to
check the validity of the data.
}
\seealso{
\code{\link{read_npx_wide_bottom}}
}
\author{
Klev Diamanti
}
