% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_check_top}
\alias{read_npx_wide_check_top}
\title{Help function to determine the number of rows with assay information in Olink
wide format file.}
\usage{
read_npx_wide_check_top(df, file, format_spec)
}
\arguments{
\item{df}{The top data frame from a split Olink wide  file.}

\item{file}{The input Olink wide file.}

\item{format_spec}{A one-row data frame filtered from olink_wide_spec with
the Olink wide file specifications.}
}
\value{
NULL unless an inconsistency is spotted.
}
\description{
Help function to determine the number of rows with assay information in Olink
wide format file.
}
\author{
Klev Diamanti
}
