% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_npxs_version}
\alias{read_npx_wide_npxs_version}
\title{Extract version of NPX Signature from the head matrix of Olink datasets in
wide format.}
\usage{
read_npx_wide_npxs_version(df)
}
\arguments{
\item{df}{Head matrix of Olink datasets in wide format \var{df_head}.}
}
\value{
The version of the NPX Signature software.
}
\description{
Extract version of NPX Signature from the head matrix of Olink datasets in
wide format.
}
\seealso{
\code{\link{read_npx_wide}}
\code{\link{read_npx_wide_split_row}}
\code{\link{read_npx_wide_npxs_version}}
\code{\link{read_npx_wide_top}}
\code{\link{read_npx_wide_middle}}
\code{\link{read_npx_wide_bottom}}
}
\author{
Klev Diamanti
}
