% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{read_npx_wide_top}
\alias{read_npx_wide_top}
\title{Split the top matrix from Olink dataset in wide format.}
\usage{
read_npx_wide_top(df, file, olink_platform, format_spec)
}
\arguments{
\item{df}{Top matrix of Olink dataset in wide format \var{df_top}.}

\item{file}{Path to Olink software output file in wide format. Expected one
of file extensions
"xls", "xlsx", "csv", or "txt".}

\item{olink_platform}{Olink platform used to generate the input file.
One of \code{NULL} (default) for auto-detection,
"Flex", "Focus", "Target 48", or "Target 96".}

\item{format_spec}{A tibble derived from \var{olink_wide_spec} in the local
environment containing the expected format of the Olink wide file based on
the \var{olink_platform} and \var{data_type}.}
}
\value{
A list of data frames from top matrix in long format:
\itemize{
\item Data frame containing Olink assays \var{df_top_oid}
\item Data frame containing plate identifiers \var{df_top_pid}
\item Data frame containing QC warnings \var{df_top_qc_warn}
\item Data frame containing internal control assays \var{df_top_int_ctrl}
\item Data frame containing deviation from internal control assays
\var{df_top_dev_int_ctrl}
}
}
\description{
The function splits the top matrix \var{df_top} into chunks of columns, each
of which contains separate information that will be combined with matching
chunks from \var{df_mid} to convert the wide dataset into a long one.
}
\seealso{
\code{\link{read_npx_wide}}
\code{\link{read_npx_wide_split_row}}
\code{\link{read_npx_wide_npxs_version}}
\code{\link{read_npx_wide_middle}}
\code{\link{read_npx_wide_bottom}}
}
\author{
Klev Diamanti
}
