% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_zip.R
\name{read_npx_zip}
\alias{read_npx_zip}
\title{Helper function to read NPX zip-compressed files with checksum.}
\usage{
read_npx_zip(
  file,
  out_df = "arrow",
  sep = NULL,
  long_format = NULL,
  olink_platform = NULL,
  data_type = NULL,
  .ignore_files = c("README.txt"),
  quiet = FALSE
)
}
\arguments{
\item{file}{Path to Olink Software output zip file.}

\item{out_df}{The class of output data frame to be returned. Accepted values
are "tibble" and "arrow" (default).}

\item{sep}{The separator of the delimited file: NULL (autodetect), comma (,)
or semicolon (;). Used only for delimited Olink software output files.}

\item{long_format}{Boolean marking if input file is in long (TRUE) or wide
(FALSE) format. Ignored for non-excel input files.}

\item{olink_platform}{The Olink platform used to generate the input file.
Expecting "Target 96", "Target 48" or "Flex". Ignored for non-excel input
files.}

\item{data_type}{The quantification in which the data comes in, Expecting on
of NPX, Quantified or Ct. Ignored for non-excel input files.}

\item{.ignore_files}{Vector of files to ignore.}

\item{quiet}{Print a confirmation message after reading in the input file.}
}
\value{
An R6 class ArrowObject.
}
\description{
Helper function to read NPX zip-compressed files with checksum.
}
\seealso{
[read_npx()]
  [read_npx_delim()]
  [read_npx_parquet()]
  [read_npx_excel()]
}
\author{
Klev Diamanti;
  Kathleen Nevola;
  Pascal Pucholt
}
\keyword{NPX}
\keyword{csv}
\keyword{delim}
\keyword{parquet}
\keyword{sep}
\keyword{txt}
