% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_zip.R
\name{read_npx_zip}
\alias{read_npx_zip}
\title{Help function to read "Ct", "NPX", and "Quantified" data
from zip-compressed Olink software output files in R.}
\usage{
read_npx_zip(
  file,
  out_df = "arrow",
  long_format = NULL,
  olink_platform = NULL,
  data_type = NULL,
  .ignore_files = c("README.txt"),
  quiet = FALSE
)
}
\arguments{
\item{file}{Path to Olink software output zip-compressed file in wide or long
format. Expected file extensions
"zip".}

\item{out_df}{The class of the output dataset. One of
"tibble" or "arrow". (default =
"tibble")}

\item{long_format}{Boolean marking format of input file. One of \code{NULL}
(default) for auto-detection, or \code{TRUE} for long format and \code{FALSE} for
wide format files.}

\item{olink_platform}{Olink platform used to generate the input file.
One of \code{NULL} (default) for auto-detection,
or "Target 48", "Flex", "Target 96", "Explore 3072", "Explore HT", "Focus", and "Reveal".}

\item{data_type}{Quantification method of the input data. One of
"Ct", "NPX", or "Quantified". Defaults to \code{NULL} for
auto-detection.}

\item{.ignore_files}{Character vector of files included in the zip-compressed
Olink software output files that should be ignored. Used only for
zip-compressed input files (default = \emph{c("README.txt")}).}

\item{quiet}{Boolean to print a confirmation message when reading the input
file. Applies to excel or delimited input only. \code{TRUE} (default) to not print
and \code{FALSE} to print.}
}
\value{
Dataset,
"tibble" or "ArrowObject", with Olink
data in long format.
}
\description{
A zip-compressed input file might contain a file from the Olink software
containing "Ct", "NPX", and "Quantified" data, a checksum
file and one or more files to be ignored.

\strong{Note:} The zip-compressed file should strictly contain \strong{one}
Olink data file, \strong{none or one} checksum file and
\strong{none or one, or more} files that might be ignored.

\itemize{
\item \strong{Olink file} exported by Olink software in wide or long format.
Expecting file extensions
"xls", "xlsx", "csv", "txt", and "parquet".
This file is subsequently provided as input to \code{\link{read_npx}}.
\item \strong{checksum file} One of
"MD5_checksum.txt" or "checksum_sha256.txt".
depending on the checksum algorithm. The file contains only one line with the
checksum string of characters.
\item \strong{File(s) to be ignored} from the zip file. These files can be
named as a character vector in the argument \var{.ignore_files}.
}
}
\seealso{
\code{\link{read_npx}}
\code{\link{read_npx_parquet}}
\code{\link{read_npx_excel}}
\code{\link{read_npx_format}}
\code{\link{read_npx_delim}}
}
\author{
Klev Diamanti;
Kathleen Nevola;
Pascal Pucholt
}
