% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx_wide.R
\name{red_npx_wide_top_mid_long}
\alias{red_npx_wide_top_mid_long}
\title{Combine top and middle matrices in long format.}
\usage{
red_npx_wide_top_mid_long(df_top_list, df_middle_list, data_type, format_spec)
}
\arguments{
\item{df_top_list}{List of data frames from the top matrix. Output of
function \code{\link{read_npx_wide_top}}.}

\item{df_middle_list}{List of data frames from the middle matrix. Output of
function \code{\link{read_npx_wide_middle}}.}

\item{data_type}{Quantification method of the input data. One of \code{NULL}
(default) for auto-detection, "Ct", "NPX", and "Quantified".}

\item{format_spec}{A tibble derived from \var{olink_wide_spec} in the local
environment containing the expected format of the Olink wide file based on
the \var{olink_platform} and \var{data_type}.}
}
\value{
Tibble in long format combining the top and middle matrices.
}
\description{
Combined corresponding chunks of columns from the top and middle matrix
that were computed from \code{\link{read_npx_wide_top}} and
\code{\link{read_npx_wide_middle}}, respectively.
}
\seealso{
\code{\link{read_npx_wide_top}}
\code{\link{read_npx_wide_middle}}
}
\author{
Klev Diamanti
}
