% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_npx.R
\name{run_check_npx}
\alias{run_check_npx}
\title{Check and run \code{\link[=check_npx]{check_npx()}} if not provided.}
\usage{
run_check_npx(df, check_log = NULL)
}
\arguments{
\item{df}{A "tibble" or "ArrowObject"
from \code{\link{read_npx}}.}

\item{check_log}{A named list returned by \code{\link[=check_npx]{check_npx()}}. If \code{NULL},
\code{\link[=check_npx]{check_npx()}} will be run internally using \code{df}.}
}
\value{
A list containing the following elements:
\itemize{
\item \strong{col_names} List of column names from the input data frame
marking the columns to be used in downstream analyses.
\item \strong{oid_invalid} Character vector of invalid \var{OlinkID}.
\item \strong{assay_na} Character vector of assays with all samples having
\emph{NA} values.
\item \strong{sample_id_dups} Character vector of duplicate \var{SampleID}.
\item \strong{sample_id_na} Character vector containing \var{SampleID} of
samples with quantified values \emph{NA} for all assays.
\item \strong{col_class} Data frame with columns of incorrect type
including column key \var{col_key}, column name \var{col_name}, detected
column type \var{col_class} and expected column type
\var{expected_col_class}.
\item \strong{assay_qc} Character vector containing \var{OlinkID} of assays
with at least one assay warning.
\item\strong{non_unique_uniprot} Character vector of \var{OlinkID} mapped
to more than one \var{UniProt} ID.
}
}
\description{
Check and run \code{\link[=check_npx]{check_npx()}} if not provided.
}
\details{
This function acts as a wrapper for \code{\link[=check_npx]{check_npx()}}. It will check if the
input \code{check_log} provided by the user is valid. If not, it will throw
relevant errors or warnings. Alternatively, if \code{check_log} was not provided
by the user, it will run \code{\link[=check_npx]{check_npx()}} to provide \code{check_log} to enable
downstream functions to run.#'
}
\author{
Klev Diamanti
}
\keyword{internal}
